### 2021-11-15 #485 @oggioniale
Submission: ReLTER
**Reviewers:**
Submitting Author Name: Alessandro Oggioni
Submitting Author Github Handle: @oggioniale
Other Package Authors Github handles: (comma separated, delete if none) @micha-silver, @ptagliolato
Repository:  https://github.com/oggioniale/ReLTER
Version submitted: 1.0.0
Submission type: Standard
Version accepted: TBD

---

```
Package: ReLTER
Title: An interface to the eLTER for the R statistical programming environment
Version: 1.0.0
        person("Alessandro", "Oggioni", email = "oggioni.a@irea.cnr.it",
          role = c("aut", "cre", "fnd"),
          comment = c(ORCID = "0000-0002-7997-219X")
        ),
        person("Micha", "Silver", email = "silverm@post.bgu.ac.il",
          role = c("aut", "ctb"),
          comment = c(ORCID = "0000-0002-1128-1325",
          " Micha Silver is a coauthor of functions get_ilter_generalinfo and
          get_site_info for category Boundaries, he develop get_site_ODS
          function and he also is a reviewer of English wording for all
          functions.")
        ),
        person("Paolo", "Tagliolato", email = "tagliolato.a@irea.cnr.it",
          role = c("aut", "ctb"),
          comment = c(ORCID = "0000-0002-0261-313X",
          " Paolo Tagliolato is a coauthor of package, he wrote the function
          do_Q(), on which some ReLTER internal functions are based")
        )
      )
Maintainer: The package maintainer Alessandro Oggioni, phD (2020)
    <oggioniale@gmail.com>
Description: ReLTER is an R package that: provides access to DEIMS-SDR
    (https://deims.org/), allows interact with software implemented by eLTER
    Research Infrastructure (RI) and improves the data/information shared by
    them.
    ReLTER is a R package devoted to access, interact and improve the
    information and the data shared by Long Term Ecological Research (LTER)
    network. This package is born within eLTER H2020 major project that will
    help advance the development of European Long-Term Ecosystem Research
    Infrastructures (eLTER RI - https://elter-ri.eu).
    The ReLTER package functions in particular allow to:
    - retrive the information about entities (e.g. sites, datasets, and
    activities) shared by DEIMS-SDR (see e.g. get_site_info function);
    - interact with the [ODSEurope](maps.opendatascience.eu) managed by members
    of the [Geo-harmonizer](https://opendatascience.eu/geoharmonizer-project/)
    project starting with the dataset shared by [DEIMS-SDR](https://deims.org/)
    (see e.g. [get_site_ODS](https://oggioniale.github.io/ReLTER/reference/get_site_ODS.html)
    function);
    - use the site informations for download data from other platforms (see
    e.g. get_site_ODS function);
    - improve the quality of the dataset (see e.g. get_id_worms).
    Functions currently implemented are derived from the discussion of the
    needs declared by eLTER users community.
    The ReLTER package shall definitely follow the progress of eLTER-RI
    infrastructure and evolving with the improvements and develop of
    new tools.
    jsonlite (>= 1.7.2),
    sf (>= 0.9-5),
    tibble (>= 3.0.5),
    taxize (>= 0.9.97),
    sp (>= 1.4-2),
    rgeos (>= 0.5-3),
    rosm (>= 0.2.5),
    raster (>= 3.3-13),
    terra (>= 1.4-11),
    tmap (>= 3.1),
    grid (>= 3.5.1),
    leaflet (>= 2.0.3),
    qrencoder (>= 0.1.0),
    waffle (>= 0.7.0),
    dtplyr,
    ISOcodes (>= 2021.02.24),
    RColorBrewer (>= 1.1-2),
    ggforce (>= 0.3.2),
    ggplot2 (>= 3.3.5),
    jqr (>= 1.2.0),
    mapview (>= 2.9.0),
    qrcode (>= 0.1.3),
    scales (>= 1.1.1),
    xml2 (>= 1.3.2),
    magrittr (>= 2.0.1),
    stats (>= 4.0.3)
Language: en-GB
```

ReLTER package devoted to access, interact and improve the information and the data shared on World Wide Web by Long Term Ecological Research (LTER) network. Most of this data has geographic characteristics and the package provides some functions
to accessing and manipulating that.

The target audience are primarily the site managers, researchers and operators of International Long Term Ecological Research ([ILTER](https://www.ilter.network)) network of networks. Currently, ILTER encompasses 39 national networks which together operate more than 600 sites and more than 1300 involved persons. ILTER's focus is on long-term data series, site-based research and monitoring. ReLTER access, interact and improve the information and the data shared by European Long-Term Ecosystem Research Infrastructures ([eLTER RI](https://elter-ri.eu)) trough web services and APIs. The target audience in perspective, since ReLTER relates to the data and information released by the largest network dedicated to ecological research, will be all users who need to work on ecology and environmental scientific area.

To the our knowledge there aren't currently R packages that accomplish the same thing of ReLTER. [MetaEgress](https://github.com/BLE-LTER/MetaEgress) and rOpenSci [popler](https://github.com/ropensci/popler) packages are developed for same community but none of them have aims comparable with ReLTER.

N/A

N/A
