### 2021-11-15 6eb1b85
Added CONTRIBUTING.md
