### 2021-11-16 #485 @micha-silver
I put a CONTRIBUTING.md file into the dev branch. It's just a template that was suggested on the rOpenSci page, and I edited it so the URL to your github was correct.
