### 2021-11-18 7b09fb9
fixed typo in test-get_site_research_topics and test-get_site_related_resources
