+++ tests/testthat/test-get_site_research_topics.R
@@ -5,7 +5,7 @@ library(testthat)
 test_that("Expect error if internet connection is down", {
   testthat::expect_error(
     httptest::without_internet(
-      result <- ReLTER::getSiteResearchTopics(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')
+      result <- ReLTER::get_site_research_topics(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')
     ),
     "GET"
   )
@@ -14,7 +14,7 @@ test_that("Expect error if internet connection is down", {
 skip_if_offline(host = "deims.org")
 
 test_that("Output of site reseearch topics function constructs ‘tibble’ as expected", {
-  result <- ReLTER::getSiteResearchTopics(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')
+  result <- ReLTER::get_site_research_topics(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 9)
   expect_true(all(names(result) == c(
@@ -44,12 +44,12 @@ test_that("Output of site reseearch topics function constructs ‘tibble’ as e
 })
 
 test_that("Wrong input (but URL) constructs a tibble with empty data", {
-  result <- ReLTER::getSiteResearchTopics(deimsid = 'https://deims.org/ljhnhbkihubib')
+  result <- ReLTER::get_site_research_topics(deimsid = 'https://deims.org/ljhnhbkihubib')
   expect_type(result, "NULL")
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
-  result <- ReLTER::getSiteResearchTopics(deimsid = 'ljhnhbkihubib')
+  result <- ReLTER::get_site_research_topics(deimsid = 'ljhnhbkihubib')
   expect_type(result, "NULL")
 })
 
