+++ tests/testthat/test-get_site_related_resources.R
@@ -5,7 +5,7 @@ library(testthat)
 test_that("Expect error if internet connection is down", {
   testthat::expect_error(
     httptest::without_internet(
-      result <- ReLTER::getSiteRelatedResources(
+      result <- ReLTER::get_site_related_resources(
         deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe'
       )
     ),
@@ -16,7 +16,7 @@ test_that("Expect error if internet connection is down", {
 skip_if_offline(host = "deims.org")
 
 test_that("Output of site related resources function constructs ‘tibble’ as expected", {
-  result <- ReLTER::getSiteRelatedResources(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')
+  result <- ReLTER::get_site_related_resources(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 9)
   expect_true(all(names(result) == c(
@@ -46,12 +46,12 @@ test_that("Output of site related resources function constructs ‘tibble’ as 
 })
 
 test_that("Wrong input (but URL) constructs a tibble with empty data", {
-  result <- ReLTER::getSiteRelatedResources(deimsid = 'https://deims.org/ljhnhbkihubib')
+  result <- ReLTER::get_site_related_resources(deimsid = 'https://deims.org/ljhnhbkihubib')
   expect_type(result, "NULL")
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
-  result <- ReLTER::getSiteRelatedResources(deimsid = 'ljhnhbkihubib')
+  result <- ReLTER::get_site_related_resources(deimsid = 'ljhnhbkihubib')
   expect_type(result, "NULL")
 })
 
