### 2021-11-18 ec62581
Manage pipe internally
This is required to correctly run examples containing "%>%"
