+++ R/get_ilter_research_topics.R
@@ -10,7 +10,6 @@
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom jsonlite fromJSON
 #' @importFrom dplyr bind_rows distinct as_tibble
-#' @importFrom magrittr %>%
 #' @export
 #' @examples
 #' \dontrun{
