+++ R/get_site_boundaries.R
@@ -19,7 +19,6 @@
 #' @importFrom sf st_as_sf write_sf st_write
 #' @importFrom leaflet leaflet addTiles addPolygons
 #' @importFrom mapview mapshot
-#' @importFrom magrittr %>%
 #' @export
 #' @keywords internal
 #' @examples
