+++ R/produce_site_parameters_pie.R
@@ -13,7 +13,6 @@
 #' @importFrom utils data
 #' @importFrom ggplot2 theme_minimal theme element_blank element_text ggplot geom_text aes coord_fixed scale_x_continuous scale_y_continuous scale_color_manual
 #' @importFrom ggforce geom_arc_bar
-#' @importFrom magrittr %>%
 #' @export
 #' @examples
 #' \dontrun{
