+++ R/get_ilter_parameters.R
@@ -10,7 +10,6 @@
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom jsonlite fromJSON
 #' @importFrom dplyr as_tibble bind_rows distinct
-#' @importFrom magrittr %>%
 #' @export
 #' @examples
 #' \dontrun{
