### 2021-11-18 285de35
Import xslt
A method of the library is used in the package, but no functions are explicitly called; so a function (xml_xslt) is called so that the package is put in the NAMESPACE
