+++ DESCRIPTION
@@ -81,6 +81,7 @@ Imports:
     qrcode (>= 0.1.3),
     scales (>= 1.1.1),
     xml2 (>= 1.3.2),
+    xslt,
     magrittr (>= 2.0.1),
     stats (>= 4.0.3)
 Suggests: 
