+++ NAMESPACE
@@ -113,3 +113,4 @@ importFrom(worrms,wm_records_names)
 importFrom(xml2,read_xml)
 importFrom(xml2,xml_find_all)
 importFrom(xml2,xml_ns)
+importFrom(xslt,xml_xslt)
