### 2021-11-18 9182dfd
Fix to correctly run example in covr
