+++ man/do_Q.Rd
@@ -32,7 +32,7 @@ q <- '{title: .title,
     envCharacteristics: .attributes.environmentalCharacteristics
   }'
 
-do_Q(q, jj)
+ReLTER:::do_Q(q, jj)
 
 }
 \author{
