### 2021-11-18 04e28b7
Added 'show_map' parameter, with default FALSE
