+++ R/get_site_ODS.R
@@ -2,7 +2,8 @@
 #'
 #' @description This function acquires various raster layers from
 #' ODS Europe:  https://maps.opendatascience.eu/
-#' and crops to an eLTER site boundary, which is obtained from the DEIMS-SDR sites API.
+#' and crops to an eLTER site boundary, which is obtained 
+#' from the DEIMS-SDR sites API.
 #'
 #' @param deimsid  a `character`. The DEIMS ID of the site from
 #' DEIMS-SDR website. More information about DEIMS ID from:
@@ -25,14 +26,19 @@
 #'    aggregated and rasterized first to 10m spatial resolution
 #'    and after downsampled to 30m by spatial average.
 #' 
-#' Natura2000: Protected areas rasterized from NATURA 2000 (A, B and C site categories)
-#'    and OSM (IUCN Ia, IUCN Ib, IUCN 2, IUCN 3, IUCN 4, IUCN 5, IUCN 6 and others categories),
-#'    first to 10m spatial resolution and after downsampled to 30m by spatial average.
+#' Natura2000: Protected areas rasterized from NATURA 2000 
+#'    (A, B and C site categories)
+#'    and OSM (IUCN Ia, IUCN Ib, IUCN 2, IUCN 3, IUCN 4, IUCN 5, IUCN 6
+#'    and others categories),
+#'    first to 10m spatial resolution and after downsampled 
+#'    to 30m by spatial average.
 #'    The overlap areas are indicated in a new category. 
 #' 
-#' NDVI:  NDVI time-series, derived from the Landsat quarterly temporal composites
+#' NDVI:  NDVI time-series, 
+#'    derived from the Landsat quarterly temporal composites
 #'
-#' All datasets are georeferenced to the EPSG:3035 coordinate reference system.
+#' All datasets are georeferenced to the 
+#' EPSG:3035 coordinate reference system.
 #' and all except clc2018 have 30 meters resolution
 #'
 #' @return The function returns a SpatRaster object (from the `terra` package)
@@ -72,35 +78,44 @@ get_site_ODS <- function(deimsid, dataset = "landcover") {
   # Define viscurl virtual dataset
   ods_url <- paste0("/vsicurl/", ods_url)
   # Links to individual datasets
-  landcover <- "lcv/lcv_landcover.hcl_lucas.corine.rf_p_30m_0..0cm_2019_eumap_epsg3035_v0.1.tif"
-  clc2018 <- "lcv/lcv_landcover.clc_corine_c_100m_0..0cm_2018_eumap_epsg3035_v2020.tif"
-  osm_buildings <- "lcv/lcv_building_copernicus.osm_p_30m_0..0cm_2018..2021_eumap_epsg3035_v0.1.tif"
-  natura2000 <- "adm/adm_protected.area_natura2000.osm_p_30m_0..0cm_2019..2021_eumap_epsg3035_v0.1.tif"
-  ndvi_spring <- "lcv/lcv_ndvi_landsat.glad.ard_p50_30m_0..0cm_201903_eumap_epsg3035_v1.0.tif"
-  ndvi_summer <- "lcv/lcv_ndvi_landsat.glad.ard_p50_30m_0..0cm_201906_eumap_epsg3035_v1.0.tif"
-  ndvi_autumn <- "lcv/lcv_ndvi_landsat.glad.ard_p50_30m_0..0cm_201909_eumap_epsg3035_v1.0.tif"
-  ndvi_winter <- "lcv/lcv_ndvi_landsat.glad.ard_p50_30m_0..0cm_201912_eumap_epsg3035_v1.0.tif"
-  full_url <- dplyr::case_when(dataset == "landcover" ~ paste0(ods_url, landcover),
[…]
