### 2021-11-19 d279e8e
some minor changes for goodpractice and other checks: added stringr (>= 1.4.0), worrms (>= 0.4.2), spelling (>= 2.2) and xslt (>= 1.4.3) into DESCRIPTION file and Imports section LICENSE is now mentioned in the DESCRIPTION file deleted the .gitignoreAle hidden file from package directory changed in following file non-ASCII characters: produce_site_map.R added dependencies for xslt into get_sos_procedurelist.R file changed dependencies for files: get_ilter_generalinfo.R, get_network_related_resources.R, get_site_related_resources.R, get_network_research_topics.R, taxon_id_worms_refine.R, get_il
[…]
