+++ R/get_site_info.R
@@ -29,7 +29,7 @@
 #'   category = c("EnvCharacts", "Affiliations", "Boundaries")
 #' )
 #' siteInfo
-#' 
+#'
 #' site <- get_site_info(
 #'   deimsid = "https://deims.org/79d6c1df-570f-455f-a929-6cfe5c4ca1e9",
 #'   category = "Boundaries"
@@ -50,17 +50,17 @@ get_site_info <- function(deimsid, category = NA) {
   )
   export <- httr::GET(url = url)
   jj <- suppressMessages(httr::content(export, "text"))
-  status <- jj %>% 
-    jqr::jq(as.character('{status: .errors.status}')) %>% 
+  status <- jj %>%
+    jqr::jq(as.character("{status: .errors.status}")) %>%
     textConnection() %>%
     jsonlite::stream_in(simplifyDataFrame = TRUE) %>%
-    dtplyr::lazy_dt() %>% 
+    dtplyr::lazy_dt() %>%
     dplyr::as_tibble()
   if (is.na(status)) {
     invisible(
       utils::capture.output(
         siteInfo <- dplyr::as_tibble(
-          ReLTER:::do_Q(q, jj)
+          do_Q(q, jj)
         )
       )
     )
@@ -69,7 +69,7 @@ get_site_info <- function(deimsid, category = NA) {
     } else {
       # add 'Affiliations' info
       if (any(stringr::str_detect(category, "Affiliations"))) {
-        siteAffil <- ReLTER:::get_site_affiliations(deimsid = deimsid)
+        siteAffil <- get_site_affiliations(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
           siteAffil,
@@ -87,10 +87,9 @@ get_site_info <- function(deimsid, category = NA) {
       } else {
         siteInfo <- siteInfo
       }
-      
       # add 'Contacts' info
       if (any(stringr::str_detect(category, "Contacts"))) {
-        siteConta <- ReLTER:::get_site_contact(deimsid = deimsid)
+        siteConta <- get_site_contact(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
           siteConta,
@@ -108,10 +107,9 @@ get_site_info <- function(deimsid, category = NA) {
       } else {
         siteInfo <- siteInfo
       }
-      
       # add 'EnvCharacts' info
       if (any(stringr::str_detect(category, "EnvCharacts"))) {
-        siteEnvCh <- ReLTER:::get_site_envcharacts(deimsid = deimsid)
+        siteEnvCh <- get_site_envcharacts(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
           siteEnvCh,
@@ -129,10 +127,9 @@ get_site_info <- function(deimsid, category = NA) {
       } else {
         siteInfo <- siteInfo
       }
-      
       # add 'General' info
       if (any(stringr::str_detect(category, "General"))) {
-        siteGener <- ReLTER:::get_site_general(deimsid = deimsid)
+        siteGener <- get_site_general(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
           siteGener,
@@ -150,10 +147,9 @@ get_site_info <- function(deimsid, category = NA) {
       } else {
         siteInfo <- siteInfo
       }
-      
       # add 'Infrastructure' info
       if (any(stringr::str_detect(category, "Infrastructure"))) {
-        siteInfra <- ReLTER:::get_site_infrastructure(deimsid = deimsid)
+        siteInfra <- get_site_infrastructure(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
           siteInfra,
@@ -171,10 +167,9 @@ get_site_info <- function(deimsid, category = NA) {
[…]
