+++ R/get_network_related_resources.R
@@ -11,13 +11,14 @@
 #' resources shared by the network's sites.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom jsonlite fromJSON
-#' @importFrom dplyr bind_rows distinct as_tibble
+#' @importFrom dplyr bind_rows distinct as_tibble select
 #' @importFrom magrittr %>%
 #' @export
 #' @examples
 #' \dontrun{
 #' listRelatedResources <- get_network_related_resources(
-#'   networkDEIMSID = "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
+#'   networkDEIMSID =
+#'   "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
 #' )
 #' listRelatedResources[1:10, ]
 #' }
@@ -47,8 +48,7 @@ get_network_related_resources <- function(networkDEIMSID) {
     uniteSiteRelatedResources <- dplyr::bind_rows(allSiteRelatedResources)
     relatedResourcesNetworkList <- uniteSiteRelatedResources$relatedResources
     relatedResourcesNetworkDF <- dplyr::bind_rows(relatedResourcesNetworkList)
-    # relatedResourcesNetworkDF$uri <- paste0(relatedResourcesNetworkDF$relatedResourcesId$prefix, relatedResourcesNetworkDF$relatedResourcesId$suffix)
-    relatedResourcesNetworkDF <- relatedResourcesNetworkDF %>% 
+    relatedResourcesNetworkDF <- relatedResourcesNetworkDF %>%
       dplyr::select(relatedResourcesTitle, uri, relatedResourcesChanged)
     uniqueSiteRelatedResources <- dplyr::as_tibble(
       dplyr::distinct(
@@ -57,7 +57,8 @@ get_network_related_resources <- function(networkDEIMSID) {
     )
     uniqueSiteRelatedResources
   } else {
-    message("\n---- The requested page could not be found. Please check again the Network.iD ----\n")
-    uniqueSiteParameters <- NULL
+    message("\n----\nThe requested page could not be found.
+Please check again the Network.iD\n----\n")
+    uniqueSiteRelatedResources <- NULL
   }
 }

[…]
