+++ R/produce_site_qrcode.R
@@ -15,19 +15,8 @@
 #'
 ### function produce_site_qrcode
 produce_site_qrcode <- function(deimsid) {
-  suppressWarnings({ 
-    # FIXME replace qrcode::qrcode_gen() with qrcode::qr_code() properly
-    qrcode::qrcode_gen(
-      deimsid,
-      ErrorCorrectionLevel = "L",
-      dataOutput = FALSE,
-      plotQRcode = TRUE,
-      wColor = "White",
-      bColor = "#1479BC",
-      mask = 1,
-      softLimitFlag = TRUE
-    )
+  suppressWarnings({
+    qrcode::qr_code(deimsid, ecl = "L") %>%
+      plot(col = c("White", "#1479BC"))
   })
-  # qrcode::qr_code(deimsid, ecl = "L") %>%
-  #   plot(col=c("White", "#1479BC"))
 }
