### 2021-11-20 7b5ce87
Changes to get_site_info and minor edits of line length
