+++ man/get_ilter_generalinfo.Rd
@@ -4,20 +4,23 @@
 \alias{get_ilter_generalinfo}
 \title{eLTER get_ilter_generalinfo function}
 \usage{
-get_ilter_generalinfo(country_name = NA, site_name = NA)
+get_ilter_generalinfo(country_name = NA, site_name = NA, show_map = FALSE)
 }
 \arguments{
-\item{country_name}{A `character`. This character string filters the full set
-of DEIMS sites by country name. Partial matching is supported.}
+\item{country_name}{A `character`. This character string filters 
+the full set of DEIMS sites by country name. Partial matching is supported.}
 
 \item{site_name}{A `character`. This character string filters by site name
 where, again, partial matching is supported}
+
+\item{show_map}{A `boolean`. If TRUE a Leaflet map of site locations 
+is shown. Default FALSE}
 }
 \value{
 An `sf` object of the bounding boxes of sites in the filtered list,
 containing the name, DEIMS ID, longitude, latitude, average altitude,
-and affiliation of the filtered ILTER sites. If no bounding box is available,
-the centroid is returned.
+and affiliation of the filtered ILTER sites. 
+If no bounding box is available,the centroid is returned.
 }
 \description{
 This function downloads generic information
@@ -29,22 +32,16 @@ strings is specified, then a filtered subset of the sites is acquired.
 \examples{
 \dontrun{
 listOfAllSites <- get_ilter_generalinfo()
-nrow(listOfAllSites)
-
+length(listOfAllSites[,1])
 sitesAustria <- get_ilter_generalinfo(country_name = "Austri")
 # (matches Austria, but not Australia)
 length(sitesAustria$title)
-
-eisenwurzen <- get_ilter_generalinfo(
-  country_name = "Austri",
-  site_name =" Eisen"
-)
+eisenwurzen <- get_ilter_generalinfo(country_name = "Austri",
+                                     site_name =" Eisen")
 eisenwurzen[,1:2]
 eisenwurzen_deimsid <- eisenwurzen$uri
 eisenwurzen_deimsid
-
 }
-
 }
 \author{
 Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}

[…]
