### 2021-11-20 b755b76
Changes to get_site_info and minor edits of line length
