+++ man/get_site_ODS.Rd
@@ -25,8 +25,8 @@ i.e. writeRaster(ds_site, "site_dataset.tif")
 \description{
 This function acquires various raster layers from
 ODS Europe:  https://maps.opendatascience.eu/
-and crops to an eLTER site boundary, which is obtained from the DEIMS-SDR
-sites API.
+and crops to an eLTER site boundary, which is obtained 
+from the DEIMS-SDR sites API.
 }
 \details{
 Supported datasets from the ODS repository include:
@@ -38,14 +38,19 @@ OSM buildings: Buildings according to OSM polygons
    and the Copernicus impervious build-up layer (2018),
    aggregated and rasterized first to 10m spatial resolution
    and after downsampled to 30m by spatial average.
-Natura2000: Protected areas rasterized from NATURA 2000 (A, B and C site
-   categories) and OSM (IUCN Ia, IUCN Ib, IUCN 2, IUCN 3, IUCN 4, IUCN 5,
-   IUCN 6 and others categories), first to 10m spatial resolution and
-   after downsampled to 30m by spatial average.
-   The overlap areas are indicated in a new category.
-NDVI:  NDVI time-series, derived from the Landsat quarterly temporal
-composites
-All datasets are georeferenced to the EPSG:3035 coordinate reference system.
+Natura2000: Protected areas rasterized from NATURA 2000 
+   (A, B and C site categories)
+   and OSM (IUCN Ia, IUCN Ib, IUCN 2, IUCN 3, IUCN 4, IUCN 5, IUCN 6
+   and others categories),
+   first to 10m spatial resolution and after downsampled 
+   to 30m by spatial average.
+   The overlap areas are indicated in a new category. 
+
+NDVI:  NDVI time-series, 
+   derived from the Landsat quarterly temporal composites
+
+All datasets are georeferenced to the 
+EPSG:3035 coordinate reference system.
 and all except clc2018 have 30 meters resolution
 }
 \examples{
