### 2021-11-22 #485 @ldecicco-USGS
Hi @oggioniale ! Thanks for the submission. Once you've made appropriate changes, can you run `pkgcheck::pkgcheck()` yourself  and confirm the summary at the top (or result of summary() applied to check results) gives all ✔️ and no ✖️? It should also confirm, "This package may be submitted." Thanks!

Similarly, if you want to look at the `goodpractice` output, you can run that locally with `goodpractice::gp()`

Once you've done that, let me know and I'll re-check the package here.

From a rOpenSci-fit perspective, I think this package looks good to proceed to a review. I'll begin looking around for an editor.
