### 2021-11-28 722d5a1
solved all the issues indicated by good practice
