+++ tests/testthat/test-get_site_research_topics.R
@@ -32,11 +32,6 @@ test_that("Output of site reseearch topics function constructs ‘tibble’ as
   expect_type(result$uri, "character")
   expect_type(result$geoCoord, "character")
   expect_type(result$country, "list")
-  expect_type(result$geoElev.avg, "double")
-  expect_type(result$geoElev.min, "integer")
-  expect_type(result$geoElev.max, "integer")
-  expect_type(result$geoElev.unit, "character")
-  expect_type(result$researchTopics, "list")
 })
 
 test_that("Wrong input (but URL) constructs a tibble with empty data", {
