+++ R/get_site_ODS.R
@@ -1,7 +1,7 @@
 #' @title eLTER get_site_ODS function
 #' @description This function acquires various raster layers from
 #' ODS Europe:  https://maps.opendatascience.eu/
-#' and crops to an eLTER site boundary, which is obtained 
+#' and crops to an eLTER site boundary, which is obtained
 #' from the DEIMS-SDR sites API.
 #'
 #' @param deimsid  a `character`. The DEIMS ID of the site from
@@ -20,18 +20,18 @@
 #'    and the Copernicus impervious build-up layer (2018),
 #'    aggregated and rasterized first to 10m spatial resolution
 #'    and after downsampled to 30m by spatial average.
-#' Natura2000: Protected areas rasterized from NATURA 2000 
+#' Natura2000: Protected areas rasterized from NATURA 2000
 #'    (A, B and C site categories)
 #'    and OSM (IUCN Ia, IUCN Ib, IUCN 2, IUCN 3, IUCN 4, IUCN 5, IUCN 6
 #'    and others categories),
-#'    first to 10m spatial resolution and after downsampled 
+#'    first to 10m spatial resolution and after downsampled
 #'    to 30m by spatial average.
-#'    The overlap areas are indicated in a new category. 
-#' 
-#' NDVI:  NDVI time-series, 
+#'    The overlap areas are indicated in a new category.
+#'
+#' NDVI:  NDVI time-series,
 #'    derived from the Landsat quarterly temporal composites
 #'
-#' All datasets are georeferenced to the 
+#' All datasets are georeferenced to the
 #' EPSG:3035 coordinate reference system.
 #' and all except clc2018 have 30 meters resolution
 #' @return The function returns a SpatRaster object (from the `terra` package)
@@ -60,7 +60,7 @@
 #'   dataset = "ndvi_summer"
 #' )
 #' siteNDVI
-#'   terra::plot(siteNDVI)
+#' terra::plot(siteNDVI)
 #' }
 #'
 ### function get_site_ODS
@@ -124,7 +124,7 @@ get_site_ODS <- function(deimsid, dataset = "landcover") {
   }
   # Crop and mask the raster dataset to the boundary polygon
   # The boundary must be transformed first
-  # to the European CRS (EPSG:3034) used by ODS 
+  # to the European CRS (EPSG:3034) used by ODS
   boundary <- sf::st_transform(boundary, terra::crs(ds))
   ds_site <- terra::mask(terra::crop(ds, boundary), terra::vect(boundary))
   return(ds_site)
