+++ R/produce_site_qrcode.R
@@ -1,22 +1,31 @@
 #' @title eLTER produce_site_qrcode function
 #' @description This function obtain the QRCode of any DEIMS-SDR
 #' entities of the eLTER.
-#' @param deimsid A `character`. The DEIMS ID of site from
+#' @param deimsid a `character`. The DEIMS ID of entities from
 #' DEIMS-SDR website. More information about DEIMS iD in this
 #' \href{https://deims.org/docs/deimsid.html}{page}.
-#' @return The output of the function is a QRcode `image`.
+#' @param do_plot a `boolean`. Plot the computed QRCode. Default FALSE.
+#' @return The QR code as a logical matrix with "qr_code" class.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom qrcode qrcode_gen
 #' @export
 #' @examples
-#' produce_site_qrcode(
+#' qrcode <- produce_site_qrcode(
 #'   deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
+#' )
+#' plot(qrcode)
+#' 
+#' produce_site_qrcode(
+#'   deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe",
+#'   do_plot = TRUE
 #' )
 #'
 ### function produce_site_qrcode
-produce_site_qrcode <- function(deimsid) {
-  suppressWarnings({
-    qrcode::qr_code(deimsid, ecl = "L") %>%
+produce_site_qrcode <- function(deimsid, do_plot = FALSE) {
+  res <- qrcode::qr_code(deimsid, ecl = "L")
+  if (do_plot) {
+    res %>% 
       plot(col = c("White", "#1479BC"))
-  })
+  }
+  return(res)
 }
