### 2021-11-30 #485 @oggioniale
Dear @maurolepore,
we finish the check of the package following the suggestions and using the goodpractice::gp().
After this, now the response is "Wee! Fabulous package! Keep up the incredible work!"
So now the package may be submitted.

Regards
Ale
