### 2021-11-30 #485 @mpadge
sorry @maurolepore just working on that - it'll be back up soon, and checks will appear here. @oggioniale note you can and should [run these checks locally with `pkgcheck::pkgcheck()`](https://docs.ropensci.org/pkgcheck/) to confirm "this package may be submitted" prior to proceeding further.
