### 2021-11-30 #485 @maurolepore
Thanks @mpadge! I believe @oggioniale diligently did run `pkgcheck::pkcheck()` locally. My call to the bot was just to confirm and leave a formal record here.
