### 2021-11-30 #485 @mpadge
@oggioniale @maurolepore Our system has been updated a bit since the initial checks. You'll notice it now reports:

- :heavy_multiplication_x: Package has no continuous integration checks.

This is true, and the previous reports which gave a :heavy_check_mark: for that were inaccurate. Details of the build failure can be seen in the *goodpractice and other checks* section.
