### 2021-12-01 #485 @maurolepore
Dear @oggioniale,

Thanks again for this submission and for you latest improvements. Soon I'll
search for two reviewers that will help you make this package even better, but
first I would like to note some issues I see.

All issues are numbered in case you need to refer to them in your response. Some
issues have a check-box -- please address them as soon as possible; I'll search
for reviewers once they have been resolved. The remaining issues with no
check-box are important but not urgent.

### Technical checks

Please ensure you can check all boxes under "Technical checks". I see two
missing checks:

- [ ] (01) LICENCE: In DESCRIPTION I see "License: GPL-3". Please run
`usethis::use_gpl3_license()`, commit any change, then check this box.

- [ ] (02) Continuous Integration (CI): You can add a standard CI workflow with
`usethis::use_github_action_check_standard()`. If the standard workflow doesn't
cover the version of R you declare in the field "Depends:" of DESCRIPTION, then
you need a fuller CI workflow. You may run instead
`usethis::use_github_action("check-full.yaml")` and remove the checks for
versions of R that are older than the one this package supports.

### Installation and R CMD check

- [ ] (03) Please ensure the package installs correctly on a fresh R environment
(e.g. on a new project on https://rstudio.cloud) and passes R CMD check cleanly.

* I fail to install the package on my computer:

```r
> devtools::install_github("oggioniale/ReLTER")
Downloading GitHub repo oggioniale/ReLTER@HEAD
✓  checking for file ‘/tmp/Rtmpwupekz/remotes260a10984a29/oggioniale-ReLTER-05a9522/DESCRIPTION’ ...
─  preparing ‘ReLTER’:
✓  checking DESCRIPTION meta-information ...
   Omitted ‘LazyData’ from DESCRIPTION
─  building ‘ReLTER_1.0.0.tar.gz’

Installing package into ‘/home/mauro/R/x86_64-pc-linux-gnu-library/4.1’
(as ‘lib’ is unspecified)
* installing *source* package ‘ReLTER’ ...
** using staged installation
** R
** inst
** byte-compile and prepare package for lazy loading
Error: object ‘ISO_3166_1’ is not exported by 'namespace:ISOcodes'
Execution halted
ERROR: lazy loading failed for package ‘ReLTER’
* removing ‘/home/mauro/R/x86_64-pc-linux-gnu-library/4.1/ReLTER’
...
```

* The [bot reports](https://github.com/ropensci/software-review/issues/485#issuecomment-982765579) _R CMD check process failed with message: 'Build process failed'_. It also
reports _Package coverage failed_ but this may be because the package can't be
installed so please first address the issues with installation and R CMD check.

### DESCRIPTION

- [ ] (04) Please review the grammar of the documentation, specially DESCRIPTION
because CRAN inspects it carefully. Having a grammatically correct and clear
documentation will help the reviewers better understand what this package does,
and eventually will also help the users. I'm not a native English speaker, and
the reviewers might not either. If you can, please find a native speaker in your
network. Even if you are a native English speaker, it usually helps to get
feedback from someone with fresh eyes.

- [ ] (05) Please correct the letter case of the title. CRAN enforces the use of
title case. You may use `tools::toTitleCase()`.

- [ ] (06) Please shorten the title to 65 characters or less. CRAN has rejected
packages based on this. Consider removing the phrase "for the R statistical
programming environment"; it uses 45 characters and CRAN may consider it is
redundant. For example use "An Interface to the European Long-Term Ecosystem
Research".

See: https://github.com/DavisVaughan/extrachecks

* (07) Under "Authors@R:", the second element of the `comment` (after the ORCID)
might be hard to maintain. If you change the name of a function you may forget
to update the name here. Instead you may use the tag `@author` in the help file
of the relevant functions, but just adding the name of an author in DESCRIPTION
may be enough.

* (08) Under "Imports:", I see most dependencies require a specific version.
This may make the package harder to maintain and install. For example note that
ggplot2 requires specific versions of only a few of all its dependencies:
https://github.com/tidyverse/ggplot2/blob/main/DESCRIPTION.

### Non-standard files and directories

* (09) Do you need README.html? Usually README.Rmd and README.md are enough.

* (10) Do you need these directories?: R_newFunctions/, figure/, miscellaneous/,
xslt/. They play no role in the installed R package and may confuse the
reviewers.
