+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ReLTER
 Type: Package
-Title: An interface to the eLTER for the R statistical programming environment
+Title: An R interface for the eLTER community
 Version: 1.0.0
 Authors@R: c(
         person("Alessandro", "Oggioni", email = "oggioni.a@irea.cnr.it", 
@@ -11,45 +11,44 @@ Authors@R: c(
           role = c("aut", "ctb"),
           comment = c(ORCID = "0000-0002-1128-1325",
           "Micha Silver is a coauthor of functions get_ilter_generalinfo and
-          get_site_info for category Boundaries, he develop get_site_ODS
+          get_site_info for category Boundaries, he has developed get_site_ODS
           function and he also is a reviewer of English wording for all
           functions")
         ),
         person("Paolo", "Tagliolato", email = "tagliolato.a@irea.cnr.it",
           role = c("aut", "ctb"),
           comment = c(ORCID = "0000-0002-0261-313X",
-          "Paolo Tagliolato is a coauthor of package, he wrote the function
+          "Paolo Tagliolato is a coauthor of the package, he wrote the function
           do_Q() and developed the related pattern to interact with external
-          API on which several ReLTER internal functions are based")
+          API on which several ReLTER functions are based")
         )
       )
 Maintainer: The package maintainer Alessandro Oggioni, phD (2020)
     <oggioniale@gmail.com>
-Description: ReLTER is an R package that: provides access to DEIMS-SDR
-    (https://deims.org/), allows interact with software implemented by eLTER
-    Research Infrastructure (RI) and improves the data/information shared by
-    them.
-    ReLTER is a R package devoted to access, interact and improve the
-    information and the data shared by Long Term Ecological Research (LTER)
-    network. This package is born within eLTER H2020 major project that will
+Description: ReLTER provides access to DEIMS-SDR (https://deims.org/),
+    and allows interaction with data and software implemented by eLTER
+    Research Infrastructure (RI) thus improving data sharing among European 
+    LTER projects.
+    ReLTER uses the R language to access and interact with the DEIMS-SDR
+    archive of information shared by the Long Term Ecological Research (LTER)
+    network. This package grew within eLTER H2020 as a major project that will
     help advance the development of European Long-Term Ecosystem Research
     Infrastructures (eLTER RI - https://elter-ri.eu).
     The ReLTER package functions in particular allow to:
-    - retrive the information about entities (e.g. sites, datasets, and
+    - retrieve the information about entities (e.g. sites, datasets, and
     activities) shared by DEIMS-SDR (see e.g. get_site_info function);
     - interact with the [ODSEurope](maps.opendatascience.eu) managed by members
     of the [Geo-harmonizer](https://opendatascience.eu/geoharmonizer-project/)
     project starting with the dataset shared by [DEIMS-SDR](https://deims.org/)
     (see e.g. [get_site_ODS](https://oggioniale.github.io/ReLTER/reference/get_site_ODS.html)
     function);
-    - use the site informations for download data from other platforms (see
-    e.g. get_site_ODS function);
+    - use the eLTER site informations to download and crop geospatial data
+    from other platforms (see e.g. get_site_ODS function);
     - improve the quality of the dataset (see e.g. get_id_worms).
-    Functions currently implemented are derived from the discussion of the
-    needs declared by eLTER users community.
-    The ReLTER package shall definitely follow the progress of eLTER-RI
-    infrastructure and evolving with the improvements and develop of
-    new tools.
+    Functions currently implemented are derived from discussions of the
+    needs among the eLTER users community.
+    The ReLTER package will continue to follow the progress of eLTER-RI
+    and evolve, adding new tools and improvements as required.
 License: GPL-3 + file LICENSE
 Encoding: UTF-8
 LazyData: true
