+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: ReLTER
 Type: Package
-Title: An R interface for the eLTER community
+Title: An R Interface for the eLTER Community
 Version: 1.0.0
 Authors@R: c(
         person("Alessandro", "Oggioni", email = "oggioni.a@irea.cnr.it", 
@@ -49,7 +49,7 @@ Description: ReLTER provides access to DEIMS-SDR (https://deims.org/),
     needs among the eLTER users community.
     The ReLTER package will continue to follow the progress of eLTER-RI
     and evolve, adding new tools and improvements as required.
-License: GPL-3 + file LICENSE
+License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
 Depends: R (>= 3.5.0)
