+++ NAMESPACE
@@ -32,7 +32,7 @@ export(produce_site_qrcode)
 export(taxon_id_pesi)
 export(taxon_id_worms)
 export(taxon_id_worms_refine)
-importFrom(ISOcodes,ISO_3166_1)
+import(ISOcodes)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(dplyr,as_tibble)
 importFrom(dplyr,bind_rows)
