### 2021-12-02 fe3cb18
Adds LICENCE.md
