### 2021-12-02 e4da87c
for solving the issues rOpenSci number 03, 06 and 09 about https://github.com/ropensci/software-review/issues/485
