+++ tests/testthat/test-produce_network_points_map.R
@@ -96,8 +96,6 @@ test_that("Output of site affiliation information function constructs ‘sf' wit
       "https://deims.org/networks/e904354a-f3a0-40ce-a9b5-61741f66c824",
     countryCode = "DEU"
   )
-  result_sp <- sf::as_Spatial(result$coordinates)
-  result_valid <- rgeos::gIsValid(result_sp, byid = FALSE, reason = TRUE)
-  expect_type(result_valid, "character")
-  expect_match(result_valid, "Valid Geometry")
+  result_valid <- sf::st_is_valid(result$coordinates)
+  expect_true(any(result_valid))
 })
