### 2021-12-02 228318b
Delete R_newFunctions directory
