### 2021-12-02 7cae772
Delete figure directory
