### 2021-12-02 3ee51f9
Delete miscellaneus directory
