### 2021-12-02 d409e79
deleted the comments about the authors as suggested by issue https://github.com/ropensci/software-review/issues/485
