+++ DESCRIPTION
@@ -9,18 +9,11 @@ Authors@R: c(
         ),
         person("Micha", "Silver", email = "silverm@post.bgu.ac.il",
           role = c("aut", "ctb"),
-          comment = c(ORCID = "0000-0002-1128-1325",
-          "Micha Silver is a coauthor of functions get_ilter_generalinfo and
-          get_site_info for category Boundaries, he has developed get_site_ODS
-          function and he also is a reviewer of English wording for all
-          functions")
+          comment = c(ORCID = "0000-0002-1128-1325")
         ),
         person("Paolo", "Tagliolato", email = "tagliolato.a@irea.cnr.it",
           role = c("aut", "ctb"),
-          comment = c(ORCID = "0000-0002-0261-313X",
-          "Paolo Tagliolato is a coauthor of the package, he wrote the function
-          do_Q() and developed the related pattern to interact with external
-          API on which several ReLTER functions are based")
+          comment = c(ORCID = "0000-0002-0261-313X")
         )
       )
 Maintainer: The package maintainer Alessandro Oggioni, phD (2020)
