+++ README.md
@@ -24,6 +24,8 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 <!-- Once you have submitted a package and it has passed editor checks, add a peer-review badge via -->
 <!-- [![](https://badges.ropensci.org/<issue_id>_status.svg)](https://github.com/ropensci/software-review/issues/<issue_id>) -->
 <!-- where issue_id is the number of the issue in the software-review repository. -->
+
+[![R-CMD-check](https://github.com/oggioniale/ReLTER/workflows/R-CMD-check/badge.svg)](https://github.com/oggioniale/ReLTER/actions)
 <!-- badges: end -->
 
 `{ReLTER}` is an R package that: provides access to
@@ -109,6 +111,9 @@ You can install the development version of `{ReLTER}` from
 install.packages("devtools")
 devtools::install_github("oggioniale/ReLTER")
 ```
+
+If you wish to help develop this package, please follow the
+[contributing guidelines](CONTRIBUTING.md).
 
 ## :memo: Examples
 
@@ -203,12 +208,16 @@ tDataset
 #> Bounding box:  xmin: 11.88721 ymin: 43.20518 xmax: 15.86426 ymax: 45.91294
 #> Geodetic CRS:  WGS 84
 #> # A tibble: 1 × 34
-#>   title   abstract      keywords  uri    type  dateRange.from dateRange.to relatedSite contacts.corres… contacts.creator contacts.metada… observationPara… observationSpec… dataPolicy doi   onlineLocation
-#> * <chr>   <chr>         <list>    <chr>  <chr> <chr>          <lgl>        <list>      <list>           <list>           <lgl>            <list>           <list>           <list>     <chr> <list>        
-#> 1 LTER N… The present … <df [5 ×… https… data… 1965-01-01     NA           <df [1 × 4… <df [1 × 4]>     <df [14 × 4]>    NA               <df [1 × 2]>     <df [3 × 2]>     <df [1 × … 10.3… <df [1 × 3]>  
-#> # … with 18 more variables: legal.accessUse <list>, legal.rights <lgl>, legal.legalAct <lgl>, legal.citation <lgl>, method.instrumentation <lgl>, method.qualityAssurance <lgl>, method.methodUrl <list>,
-#> #   method.methodDescription <list>, method.samplingTimeUnit.label <chr>, method.samplingTimeUnit.uri <lgl>, method.spatialDesign.label <chr>, method.spatialDesign.uri <lgl>,
-#> #   method.spatialScale.label <chr>, method.spatialScale.uri <lgl>, method.temporalResolution.label <chr>, method.temporalResolution.uri <lgl>, boundaries <POLYGON [°]>, boundariesDescription <chr>
+#>   title  abstract   keywords uri   type  dateRange.from dateRange.to relatedSite
+#> * <chr>  <chr>      <list>   <chr> <chr> <chr>          <lgl>        <list>     
+#> 1 LTER … The prese… <df [5 … http… data… 1965-01-01     NA           <df [1 × 4…
+#> # … with 26 more variables: contacts.corresponding <list>,
+#> #   contacts.creator <list>, contacts.metadataProvider <lgl>,
+#> #   observationParameters <list>, observationSpecies <list>, dataPolicy <list>,
+#> #   doi <chr>, onlineLocation <list>, legal.accessUse <list>,
+#> #   legal.rights <lgl>, legal.legalAct <lgl>, legal.citation <lgl>,
+#> #   method.instrumentation <lgl>, method.qualityAssurance <lgl>,
+#> #   method.methodUrl <list>, method.methodDescription <list>, …
 ```
 
[…]
