### 2021-12-03 #485 @oggioniale
I checked step by step the issues.

- [x] (01) issue solved by [commit fe3cb18](https://github.com/oggioniale/ReLTER/commit/fe3cb1852741fe06048df0cecd2a7db28a400d46)

- [x] (02) by the use of `usethis::use_github_action("check-standard")` following the suggestion of [GitHub Actions for the R language page](https://github.com/r-lib/actions/blob/v1/examples/README.md)
issue solved by [committ 2ef2a55](https://github.com/oggioniale/ReLTER/commit/2ef2a5505e756c5024fa7c7db1ae88cd06eadc32)

- [x] (03) after `rcmdcheck::rcmdcheck(".")` the result is:
```
# ── R CMD check results ──────── ReLTER 1.0.0 ────
# Duration: 5m 44.3s
#
# 0 errors ✓ | 0 warnings ✓ | 0 notes ✓
```
 issue solved by [commit e4da87c](https://github.com/oggioniale/ReLTER/commit/e4da87c32613788301b9bfeb0a7fae41a31f3449) and [commit 51dc729](https://github.com/oggioniale/ReLTER/commit/51dc729764ff462077db81b65a6d66dc9cd9487e)

- [x] (04) English revision of the Description has done by @micha-silver
issue solved by [commit a92127e](https://github.com/oggioniale/ReLTER/commit/a92127ed50cca39c91d7dbbfcb1f837d2859c85b)

- [x] (05) issue solved by [commit a92127e](https://github.com/oggioniale/ReLTER/commit/a92127ed50cca39c91d7dbbfcb1f837d2859c85b)

- [x] (06) changed the title as suggestion "Title: An R interface for the eLTER community"
issue solved by [commit e4da87c](https://github.com/oggioniale/ReLTER/commit/e4da87c32613788301b9bfeb0a7fae41a31f3449)

- [x] (07) issue solved by [commit d409e7](https://github.com/oggioniale/ReLTER/commit/d409e79eb85b26e0db76e0706e1815ea7fa67063)

- [ ] (08) we work on that in the next few hours

- [x] (09) issue solved by [commit cac8273](https://github.com/oggioniale/ReLTER/commit/cac8273cb20889ada732a5fe2fb9431ee0636668) and [commit e4da87c](https://github.com/oggioniale/ReLTER/commit/e4da87c32613788301b9bfeb0a7fae41a31f3449)

- [x] (10) I deleted R_newFunctions/, figure/ and miscellaneous/ folders, I maintained the xslt/ folder because is used in the package.
issue solved by [commit 228318b](https://github.com/oggioniale/ReLTER/commit/228318b33911d0129deb27c2729f8cd1b5ff0d37), [commit   7cae772](https://github.com/oggioniale/ReLTER/commit/7cae772eb91725f2c080e6d64d700c30286b26de) and [commit 3ee51f9](https://github.com/oggioniale/ReLTER/commit/3ee51f9ce68e7d4cf6cdba6bf111afd84259096a)
