### 2021-12-03 #485 @maurolepore
Thanks @oggioniale!

* 01 and 02: Thanks! I checked the boxes for license and CI in the section _Technical checks_ at the top. I now see the new file LICENSE -- R CMD check should throw a NOTE because it's a non-standard file. Did you run `usethis::use_gpl3_license()`? This should add not LICENSE but LICENSE.md and an entry in .Rbuildignore. For example:

```r
> use_gpl3_license()
✔ Setting active project to '/home/mauro/tmp/ReLTER'
✔ Adding '^LICENSE\\.md$' to '.Rbuildignore'
```

* 03: Please see the [bot checks above](https://github.com/ropensci/software-review/issues/485#issuecomment-985639398) under the section _2. goodpractice and other checks > Details of goodpractice and other checks (click to open)_. Surprisingly, the R CMD check results are different from what you get. See if you can fix the problems. I confirmed the issues in a fresh R environment: https://rstudio.cloud/project/1568984

* 05: I can't find the source now but I'm pretty sure CRAN will want single quotes around 'ReLTER' in your title -- or be safer and spell it out.

* 06: I'm not sure but it CRAN might also consider redundant the "R" in the title -- so maybe it's safer to use: _An Interface for 'eLTER' community_

Thanks for considering all other comments too.
