### 2021-12-03 e665ac8
Adds httptest to DESCRIPTION and testthat
