+++ DESCRIPTION
@@ -46,7 +46,7 @@ License: GPL-3
 Encoding: UTF-8
 LazyData: true
 Depends: R (>= 3.5.0)
-Imports:
+Imports: 
     ISOcodes (>= 2021.02.24),
     RColorBrewer (>= 1.1-2),
     dplyr (>= 1.0.0),
@@ -78,9 +78,10 @@ Imports:
     xml2 (>= 1.3.2),
     xslt (>= 1.4.3)
 Suggests: 
-    testthat (>= 3.0.0),
+    httptest,
     knitr,
-    rmarkdown
+    rmarkdown,
+    testthat (>= 3.0.0)
 VignetteBuilder: knitr
 URL: https://github.com/oggioniale/ReLTER
 BugReports: https://github.com/oggioniale/ReLTER/issues
