### 2021-12-05 6af320c
Fixes check errors in get_ilter_info.R
