### 2021-12-05 fd32f74
Fixes test-get_network_envcharacts.R
