### 2021-12-05 06776a6
Fixes test error in get_network_sites
