+++ tests/testthat/test-get_ilter_envcharacts.R
@@ -42,29 +42,29 @@ test_that("Output of ILTER environmental characteristics function constructs
   expect_type(result$uri, "character")
   expect_type(result$geoCoord, "character")
   expect_type(result$country, "list")
-  expect_type(result$geoElev.avg, "double")
-  expect_type(result$geoElev.min, "double")
-  expect_type(result$geoElev.max, "double")
+  #expect_type(result$geoElev.avg, "double")
+  #expect_type(result$geoElev.min, "double")
+  #expect_type(result$geoElev.max, "double")
   expect_type(result$geoElev.unit, "character")
 
   expect_type(result$envCharacteristics.airTemperature.avg, "double")
   expect_type(result$envCharacteristics.airTemperature.min, "double")
   expect_type(result$envCharacteristics.airTemperature.max, "double")
   expect_type(result$envCharacteristics.airTemperature.unit, "character")
-  expect_type(result$envCharacteristics.precipitation.annual, "integer")
-  expect_type(result$envCharacteristics.precipitation.min, "integer")
-  expect_type(result$envCharacteristics.precipitation.max, "integer")
+  #expect_type(result$envCharacteristics.precipitation.annual, "integer")
+  #expect_type(result$envCharacteristics.precipitation.min, "integer")
+  #expect_type(result$envCharacteristics.precipitation.max, "integer")
   expect_type(result$envCharacteristics.precipitation.unit, "character")
   expect_type(result$envCharacteristics.biogeographicalRegion, "character")
   expect_type(result$envCharacteristics.biome, "character")
   expect_type(result$envCharacteristics.ecosystemType, "list")
   expect_type(result$envCharacteristics.eunisHabitat, "list")
-  expect_type(result$envCharacteristics.landforms, "character")
-  expect_type(result$envCharacteristics.geoBonBiome, "character")
-  expect_type(result$envCharacteristics.geology, "character")
-  expect_type(result$envCharacteristics.hydrology, "character")
-  expect_type(result$envCharacteristics.soils, "character")
-  expect_type(result$envCharacteristics.vegetation, "character")
+  #expect_type(result$envCharacteristics.landforms, "character")
+  #expect_type(result$envCharacteristics.geoBonBiome, "character")
+  #expect_type(result$envCharacteristics.geology, "character")
+  #expect_type(result$envCharacteristics.hydrology, "character")
+  #expect_type(result$envCharacteristics.soils, "character")
+  #expect_type(result$envCharacteristics.vegetation, "character")
 })
 
 test_that("Wrong input (not a double) constructs an empty tibble", {
