### 2021-12-06 1570dcd
Fixes no internet error in get_network_sites.R
