### 2021-12-06 e10635d
executed usethis::use_gpl3_license()
