### 2021-12-06 9fd5c04
for solve https://github.com/ropensci/software-review/issues/485#issuecomment-985825397
