### 2021-12-06 #485 @oggioniale
Hi @maurolepore,
please check the package after this [commit](https://github.com/oggioniale/ReLTER/commit/1ede2f54340e7b67b4121a2790749296d0816584)

- [x] 01 - 02  I executed now `usethis::use_gpl3_license()` (@micha-silver had made the required changes regarding the licence but I corrected them!! Sorry for that!).
- [x] 03. We (thank you @micha-silver) changed something and I run again the R CMD check. The result is:
```
> rcmdcheck::rcmdcheck(".")
── R CMD build ───────────────────────────────────────────────────────────────────────────────────────────────────
✓  checking for file ‘.../DESCRIPTION’ (344ms)
─  preparing ‘ReLTER’: (4.9s)
✓  checking DESCRIPTION meta-information ...
✓  creating vignettes (58s)
─  checking for LF line-endings in source and make files and shell scripts (830ms)
   Removed empty directory ‘ReLTER/tests/testthat/_snaps’
─  building ‘ReLTER_1.0.0.tar.gz’

── R CMD check ───────────────────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/private/var/folders/p1/110rx8q101z0wn0bwh4njrcw0000gn/T/RtmpGxjm5n/file897a4bd54de8/ReLTER.Rcheck’ (443ms)
─  using R version 4.0.3 (2020-10-10)
─  using platform: x86_64-apple-darwin17.0 (64-bit)
─  using session charset: UTF-8
✓  checking for file ‘ReLTER/DESCRIPTION’ (1.6s)
─  checking extension type ... Package
─  this is package ‘ReLTER’ version ‘1.0.0’
─  package encoding: UTF-8
✓  checking package namespace information
✓  checking package dependencies (2.7s)
✓  checking if this is a source package ...
✓  checking if there is a namespace
✓  checking for executable files (2.3s)
✓  checking for hidden files and directories ...
✓  checking for portable file names ...
✓  checking for sufficient/correct file permissions ...
✓  checking whether package ‘ReLTER’ can be installed (34.8s)
✓  checking installed package size ...
✓  checking package directory
✓  checking ‘build’ directory
✓  checking DESCRIPTION meta-information (649ms)
✓  checking top-level files
✓  checking for left-over files ...
✓  checking index information (472ms)
✓  checking package subdirectories (505ms)
✓  checking R files for non-ASCII characters ...
✓  checking R files for syntax errors ...
✓  checking whether the package can be loaded (25.2s)
✓  checking whether the package can be loaded with stated dependencies (24.9s)
✓  checking whether the package can be unloaded cleanly (49.8s)
✓  checking whether the namespace can be loaded with stated dependencies (17.1s)
✓  checking whether the namespace can be unloaded cleanly (9.5s)
✓  checking dependencies in R code (10.5s)
✓  checking S3 generic/method consistency (12.7s)
✓  checking replacement functions (9.4s)
✓  checking foreign function calls (9.9s)
✓  checking R code for possible problems (29.4s)
✓  checking Rd files (1.1s)
✓  checking Rd metadata ...
✓  checking Rd cross-references (496ms)
✓  checking for missing documentation entries (25.1s)
✓  checking for code/documentation mismatches (35.1s)
✓  checking Rd \usage sections (12s)
✓  checking Rd contents ...
✓  checking for unstated dependencies in examples ...
✓  checking R/sysdata.rda (399ms)
✓  checking installed files from ‘inst/doc’ ...
✓  checking files in ‘vignettes’
✓  checking examples (27s)
✓  checking for unstated dependencies in ‘tests’ ...
─  checking tests (499ms)
Registered S3 methods overwritten by 'stars':
  method             from
  st_bbox.SpatRaster sf
  st_crs.SpatRaster  sf
✓  Running ‘testthat.R’ (9m 5.7s)
✓  checking for unstated dependencies in vignettes (600ms)
✓  checking package vignettes in ‘inst/doc’ ...
─  checking running R code from vignettes
     ‘ReLTER.Rmd’ using ‘UTF-8’... OK
    NONE
✓  checking re-building of vignette outputs (5.5s)
✓  checking PDF version of manual (6.9s)

── R CMD check results ───────────────────────────────────────────────────────────────────────── ReLTER 1.0.0 ────
Duration: 15m 7.7s

0 errors ✓ | 0 warnings ✓ | 0 notes ✓
```

- [x] 05. Do you mean the single quote around the word eLTER in the title? I do that.
- [x] 06. The title is now: "An Interface for the 'eLTER' Community"

-------

About Continuous Integration (CI) I am reading the errors to see how to solve them.
