### 2021-12-06 #485 @maurolepore
Thanks @oggioniale,

I still see differences between the R CMD check results you share and the ones I get. My priority now is to find a way you can reproduce what I see. This is crucial for you and the reviewers to maintain a productive conversation about your code.

To ensure the differences are not due to different computing environments I suggest we use the exact same computing environment. Here is the rstudio cloud project I used: https://rstudio.cloud/project/3335565. Please  create your own copy of this project and keep pulling the latest commits. You can use the terminal tab of that rstudio project -- the remote `origin` points to your repo (`git remote -vv`) so you can run `git pull`. Alternatively you may update R and all packages to the latest version -- which is pretty much what I use all the time.

Right now the HEAD of this repo points to this commit:

```bash
6d153e0 (HEAD -> main, origin/main, origin/HEAD) Merge pull request #28 from oggioniale/dev
```

R CMD check reports one error and three notes. The notes include one about non-standard files/directories which you can fix with `usethis::use_build_ignore()`.

```
N  checking top-level files
   Non-standard files/directories found at top level:
     ‘CONTRIBUTING.md’ ‘LICENSE.md’ ‘README.Rmd’ ‘_pkgdown.yaml’
     ‘codemeta.json’ ‘docs’ ‘pkgdown’ ‘xslt’
```

Please ensure R CMD check passes cleanly. Let me know if you have questions.
