### 2021-12-07 #485 @oggioniale
Thanks @maurolepore.

As you suggested I'm working on the (RStudio Cloud)[https://rstudio.cloud/project/3335565]. Actually I'm using Mac OSx and the error is actually on the function `get_site_ODS`.
This function was created by @micha-silver and I don't have a long experience on the package `terra`. @micha-silver please can you verify this error on the (RStudio Cloud)[https://rstudio.cloud/project/3335565] enviroment.

@micha-silver I checked the `get_site_ODS`and the error happenes for the site with DEIMS.Id "https://deims.org/79d6c1df-570f-455f-a929-6cfe5c4ca1e9".
All the steps in the function are right, but the line:
`ds_site <- terra::mask(terra::crop(ds, boundary), terra::vect(boundary))`
causes the system to crash. Only for that site and not for the others sites in the test.
