### 2021-12-07 #485 @micha-silver
body p { margin-bottom: 0cm; margin-top: 0pt; }

    Here is what I see:

    > library(testthat)
        > test_file("tests/testthat/test-get_site_ODS.R")

        ══ Testing test-get_site_ODS.R
═════════════════════════════════════════════════════════════════════════════
        [ FAIL 0 | WARN 0 | SKIP 0 | PASS 0 ]
        ---- Test get_site_ODS() ----
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying... ]
         Imported 1 records. Simplifying...
        [ FAIL 0 | WARN 1 | SKIP 0 | PASS 9 ][1] "No boundary for
        requested DEIMS site."
         Imported 1 records. Simplifying...0 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...8 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...9 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...0 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...1 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...2 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...3 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...4 ]
         Imported 1 records. Simplifying...
         Imported 1 records. Simplifying...5 ]
         Imported 1 records. Simplifying...
        [ FAIL 0 | WARN 1 | SKIP 0 | PASS 26 ]

        ── Warning (test-get_site_ODS.R:35:11): Function correctly
        returns NULL for site with no boundary ─────────────────────

        ----
        This site does not have boundaries uploaded to DEIMS-SDR.Please
        verify in the site page
        (https://deims.org/8a313716-ceed-4f41-8b0b-a8197bfc304a)
        ----

        Backtrace:
         1. ReLTER::get_site_ODS(deimsid = deimsids[id])
        test-get_site_ODS.R:35:10
         2. ReLTER::get_site_info(deimsid, category = "Boundaries")
         3. ReLTER::get_site_boundaries(deimsid = deimsid)

        [ FAIL 0 | WARN 1 | SKIP 0 | PASS 26 ]

    Looks good to me. The warning is when user enters
        a deimsid for a site that does not have a boundary shapefile.
        Maybe I should change that to just a message??

    On 07/12/2021 17:18, Alessandro Oggioni
      wrote:

      Thanks @maurolepore.
      As you suggested I'm working on the (RStudio
        Cloud)[https://rstudio.cloud/project/3335565]. Actually I'm
        using Mac OSx and the error is actually on the function get_site_ODS.
        This function was created by @micha-silver
        and I don't have a long experience on the package terra.
        @micha-silver
        please can you verify this error on the (RStudio
        Cloud)[https://rstudio.cloud/project/3335565] enviroment.
      @micha-silver
        I checked the get_site_ODSand the error happenes
        for the site with DEIMS.Id "https://deims.org/79d6c1df-570f-455f-a929-6cfe5c4ca1e9".
        All the steps in the function are right, but the line:
        ds_site <- terra::mask(terra::crop(ds, boundary),
          terra::vect(boundary))
        causes the system to crash. Only for that site and not for the
        others sites in the test.
