### 2021-12-07 #485 @oggioniale
@micha-silver did you execute this in the RStudio Cloud or whit the R v.4.1.2 in your machine?
Because if I execute the R CMD check in local (Mac OSx v.12.0.1 (21A559), RStudio v.2021.09.0 Build 351 and R v.4.0.3), it don't return any error.
