### 2021-12-07 #485 @oggioniale
@micha-silver I tested again the error within RStudio Cloud provided by @maurolepore.
By executing step by step the rows within the `get_site_ODS` function all the rows are fine except for the command `ds_site <- terra::mask(terra::crop(ds, boundary), terra::vect(boundary))`.
This command is executed fine in my local machine. In particular the problem is about the `terra::crop(ds, boundary)`.

Can it be a version problem of the `terra` package? The version on my computer is 1.4-11 while the one on RStudio Cloud is 1.4-22.

Could it be a problem about the size of the shape to be used for cutting? Because the extension of other sites is more little.
