### 2021-12-08 #485 @maurolepore
Thanks @oggioniale and @micha-silver for your prompt response.

I'm glad to see you're working to meet the submission requirements. To keep this thread focused on the peer-review process, I suggest you move this "internal" conversation to an issue/PR to the oggioniale/ReLTER repo.
