### 2021-12-08 #485 @oggioniale
@maurolepore in my local machine with this:
```
R version 4.1.2 (2021-11-01)
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS Monterey 12.0.1

Matrix products: default
LAPACK: /Library/Frameworks/R.framework/Versions/4.1/Resources/lib/libRlapack.dylib

locale:
[1] en_US.UTF-8/en_US.UTF-8/en_US.UTF-8/C/en_US.UTF-8/en_US.UTF-8

attached base packages:

 [1] xmlparsedata_1.0.5 compiler_4.1.2     prettyunits_1.1.1  remotes_2.4.2      tools_4.1.2        digest_0.6.29      pkgbuild_1.2.1     evaluate_0.14
 [9] jsonlite_1.7.2     praise_1.0.0       lifecycle_1.0.1    rlang_0.4.12       rex_1.2.1          whoami_1.3.0       cli_3.1.0          rstudioapi_0.13
[17] yaml_2.2.1         curl_4.3.2         xfun_0.28          xopen_1.0.0        fastmap_1.1.0      cyclocomp_1.1.0    knitr_1.36         withr_2.4.3
[25] httr_1.4.2         xml2_1.3.3         desc_1.4.0         fs_1.5.1           rprojroot_2.0.2    glue_1.5.1         R6_2.5.1           processx_3.5.2
[33] rcmdcheck_1.4.0    rmarkdown_2.11     callr_3.7.0        purrr_0.3.4        lintr_2.0.1        covr_3.5.1         magrittr_2.0.1     htmltools_0.5.2
[41] clisymbols_1.2.0   ps_1.6.0           usethis_2.1.3      goodpractice_1.0.2 lazyeval_0.2.2     crayon_1.4.2
```
The result of R CMD check is:
```
> rcmdcheck::rcmdcheck(".")
── R CMD build ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✓  checking for file ‘.../DESCRIPTION’ (475ms)
─  preparing ‘ReLTER’: (4.7s)
✓  checking DESCRIPTION meta-information ...
✓  creating vignettes (37.4s)
─  checking for LF line-endings in source and make files and shell scripts (1.2s)
   Removed empty directory ‘ReLTER/tests/testthat/_snaps’
   Omitted ‘LazyData’ from DESCRIPTION
─  building ‘ReLTER_1.0.0.tar.gz’

── R CMD check
─  using log directory ‘/private/var/folders/p1/110rx8q101z0wn0bwh4njrcw0000gn/T/Rtmp8dCnq3/filee2505040bc03/ReLTER.Rcheck’ (465ms)
─  using R version 4.1.2 (2021-11-01)
─  using platform: x86_64-apple-darwin17.0 (64-bit)
─  using session charset: UTF-8
✓  checking for file ‘ReLTER/DESCRIPTION’ (1.3s)
─  checking extension type ... Package
─  this is package ‘ReLTER’ version ‘1.0.0’
─  package encoding: UTF-8
✓  checking package namespace information ...
✓  checking package dependencies (2.5s)
✓  checking if this is a source package ...
✓  checking if there is a namespace
✓  checking for executable files (2.9s)
✓  checking for hidden files and directories ...
✓  checking for portable file names ...
✓  checking for sufficient/correct file permissions ...
✓  checking whether package ‘ReLTER’ can be installed (29.9s)
✓  checking installed package size
✓  checking package directory
✓  checking ‘build’ directory
✓  checking DESCRIPTION meta-information ...
✓  checking top-level files
✓  checking for left-over files ...
✓  checking index information ...
✓  checking package subdirectories (542ms)
✓  checking R files for non-ASCII characters ...
✓  checking R files for syntax errors ...
✓  checking whether the package can be loaded (6.6s)
✓  checking whether the package can be loaded with stated dependencies (7.5s)
✓  checking whether the package can be unloaded cleanly (7.8s)
✓  checking whether the namespace can be loaded with stated dependencies (8.3s)
✓  checking whether the namespace can be unloaded cleanly (10.6s)
✓  checking dependencies in R code (11.8s)
✓  checking S3 generic/method consistency (9.8s)
✓  checking replacement functions (18.3s)
✓  checking foreign function calls (13.7s)
✓  checking R code for possible problems (19.4s)
✓  checking Rd files (362ms)
✓  checking Rd metadata ...
✓  checking Rd cross-references ...
✓  checking for missing documentation entries (6.6s)
✓  checking for code/documentation mismatches (18.5s)
✓  checking Rd \usage sections (9.5s)
✓  checking Rd contents ...
✓  checking for unstated dependencies in examples ...
✓  checking R/sysdata.rda ...
✓  checking installed files from ‘inst/doc’ ...
✓  checking files in ‘vignettes’ ...
✓  checking examples (13.2s)
✓  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
✓  Running ‘testthat.R’ (2m 46.7s)
✓  checking for unstated dependencies in vignettes ...
✓  checking package vignettes in ‘inst/doc’ ...
─  checking running R code from vignettes
     ‘ReLTER.Rmd’ using ‘UTF-8’... OK
    NONE
✓  checking re-building of vignette outputs (1.6s)
✓  checking PDF version of manual (4.5s)

── R CMD check results ────────────── ReLTER 1.0.0 ────
Duration: 6m 19s

0 errors ✓ | 0 warnings ✓ | 0 notes ✓
```
