### 2021-12-08 #485 @maurolepore
Thanks @oggioniale you are on a good track, but R CMD checkmust pass cleanly on everyone's computers. Fortunately you seem to have identified a specific problem and a person that can help fix it. Also you now have two tools that will help you:

1. A fresh RStudio cloud project that allows you to work on your package in an R environment that reflects a "general" developer/user/reviewer than your own machine: https://rstudio.cloud/project/3335565

2. CI: You now have workflows that run R CMD check on multiple platforms and with multiple versions of R, every time you push a commit to GitHub. Once all checks successfully pass green you will have solved the problem. This is not the case with the latest commit I see on GitHub:

--

BTW, which is your latest local commit? Are you in sync with the repo on GitHub? I'm surprised you don't get these R CMD check NOTE:

```
N  checking top-level files
   Non-standard files/directories found at top level:
     ‘CONTRIBUTING.md’ ‘LICENSE.md’ ‘README.Rmd’ ‘_pkgdown.yaml’
     ‘codemeta.json’ ‘docs’ ‘pkgdown’ ‘xslt’
```

Did you forget to commit and/or push any changes you did locally?
