### 2021-12-08 0ce1d19
Fixes get_network_sites for test without_internet
