### 2021-12-09 b54f7b6
Fixes test-get_site_ODS, removed Zone Atelier Alps
