### 2021-12-09 #485 @micha-silver
Hi @oggioniale  and @maurolepore:

I think I have found and fixed the two remaining issues. I tested on Rstudio cloud, and  `test_local()` now finish cleanly (except for 1 warning). Changes are pushed to the `dev` branch.

1. Since both @oggioniale and I did not find a problem with the test for `get_site_ODS` on our home machines, I checked on the Rstudio.cloud instance, and I noticed that the session is allocated only 1 GB of RAM. That probably was causing the crashes that we saw when testing with larger eLTER sites. To overcome this I replaced the tested sites with smaller ones.
2. I also changed the function `get_network_sites` so that it correctly returns NULL for the `without_internet` test.
3. @maurolepore  Can you give us some hints why the files ‘CONTRIBUTING.md’ ‘LICENSE.md’ ‘README.Rmd’ etc are giving that note about Non-standard files ??
4.  The remaining warning is, I think, due to a duplicate `approxNA` method in both `raster` and `terra` packages. I think we will be able to solve this rewriting some bits that require `raster` and then removing it.

@oggioniale  If these changes are acceptable, please merge into `main` and restart the CI and tests.
