### 2021-12-09 #485 @maurolepore
Thanks @micha-silver! Great progress!!!

> @maurolepore Can you give us some hints why the files ‘CONTRIBUTING.md’ ‘LICENSE.md’ ‘README.Rmd’ etc are giving that note about Non-standard files ??

You can fix that note with the help of [`usethis::use_build_ignore()`](https://usethis.r-lib.org/reference/use_build_ignore.html) (hinted [above](https://github.com/ropensci/software-review/issues/485#issuecomment-987197954)).

Formally, R packages know about some so-called "standard" files and directories, like DESCRIPTION and R/. If your package includes non-standard files and directories you must add their names to a file called .Rbuildignore.

The usethis package can help you add some of those files and update .Rbuildignore as needed -- all at once. For example, [this comment above](https://github.com/ropensci/software-review/issues/485#issuecomment-985825397) shows that `usethis::use_gpl3_license()` adds LICENSE.md to .Rbuildignore:

```
> use_gpl3_license()
✔ Setting active project to '/home/mauro/tmp/ReLTER'
✔ Adding '^LICENSE\\.md$' to '.Rbuildignore'
```
