### 2021-12-10 fd97a83
Update do_Q.R
