+++ .github/workflows/check-standard.yaml
@@ -2,10 +2,8 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main, dev]
-  pull_request:
-    branches: [main]
-
+    branches: [dev]
+  
 name: R-CMD-check
 
 jobs:
