### 2021-12-11 74692c2
added command about installation dependancies
