### 2021-12-12 02a0b24
Update check-standard.yaml
