+++ .github/workflows/check-standard.yaml
@@ -25,7 +25,7 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_KEEP_PKG_SOURCE: yes
-
+      
     steps:
       - uses: actions/checkout@v2
 
@@ -35,6 +35,7 @@ jobs:
         with:
           r-version: ${{ matrix.config.r }}
           http-user-agent: ${{ matrix.config.http-user-agent }}
+          use-public-rspm: true
 
       - uses: r-lib/actions/setup-r-dependencies@v1
         with:
@@ -68,5 +69,4 @@ jobs:
             libv8-dev \
             libssl-dev \
             libcairo2-dev \
-            libgit2-dev \
-            r-cran-phangorn
+            libgit2-dev
