### 2021-12-13 #485 @oggioniale
Dear @maurolepore,
I worked for solve some issues about CI.
Currently I had the errors on the Windows and Ubuntu (oldrel).
In particular the error in Ubuntu is about `Error: <callr_remote_error: Cannot install packages:` (see image). I don't have experience about this. Do you have some suggestions? I followed some discussion online in order to find this issue but I didn't find the solution.

I'm working to install a Windows server machine and an Ubuntu machine to try to reproduce the error. If for the Windows one I am confident, for the Ubuntu one I don't understand why there is an error only for the old release.
