### 2021-12-13 #485 @micha-silver
I wonder if this [issue](https://github.com/r-lib/pak/issues/339) from the `pak` github page is relevant?
