### 2021-12-13 d68b5e3
deleted the pan installation from source
