### 2021-12-13 6001c78
Initializes get_site_MODIS()
