+++ man/get_site_MODIS.Rd
@@ -1,0 +1,119 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/get_site_MODIS.R
+\name{get_site_MODIS}
+\alias{get_site_MODIS}
+\title{eLTER get_site_MODIS function}
+\usage{
+get_site_MODIS(
+  deimsid,
+  earthdata_user,
+  earthdata_passwd,
+  product = "Vegetation_Indexes_Monthly_1Km (M*D13A3)",
+  from_date = NA,
+  to_date = NA,
+  bands = c("NDVI", "EVI"),
+  scale = TRUE,
+  save_ts_dir = NA,
+  out_folder = tempdir(),
+  show_products = FALSE,
+  show_bands = NA
+)
+}
+\arguments{
+\item{deimsid}{a `character`. The DEIMS ID of the site from
+DEIMS-SDR website. More information about DEIMS ID from:
+\href{https://deims.org/docs/deimsid.html}{page}.}
+
+\item{earthdata_user}{`character`. For authentication. 
+Please create an account on: https://urs.earthdata.nasa.gov/users/new
+and enter authentication details here.}
+
+\item{earthdata_passwd}{`character` For authentication on Earthdata.}
+
+\item{product}{a `character`. The requested MODIS product.
+Must be the full product name.
+Use `show_products = TRUE` to get a list of products.
+Some examples:
+      "Vegetation_Indexes_Monthly_1Km (M*D13A3)"
+        "LST_3band_emissivity_8day_1km (M*D21A2)"
+    "LAI_8Days_500m (MCD15A2H)"
+Default is "Vegetation_Indexes_Monthly_1Km (M*D13A3)"}
+
+\item{from_date, }{`character`, the start date for acquiring MODIS
+formatted as YYYY.MM.DD
+Must be later than 2000.02.01. (The earliest MODIS imagery was Feb 2000)}
+
+\item{to_date, }{`character`, the last date for acquiring MODIS
+formatted as YYYY.MM.DD.}
+
+\item{bands}{`character`, string array of bands in the requested product.
+Use the "show_bands" parameter first to get a list of available bands.
+Some examples: "LST_Day_1km", or multiple bands as c("NDVI", "EVI")
+Default is c("NDVI", "EVI")}
+
+\item{scale}{`boolean`. The MODIS raster values will be scaled (multiplied)
+by a factor. Some MODIS products are scaled up and saved as 16 bit integers.
+For example, Vegetation indices are scaled by 10000, and Land surface temperatures
+are scaled by 50. Scaling factors are derived from product metadata.
+For additional details refer to the  MODIS Product Tables: 
+https://modis.gsfc.nasa.gov/data/dataprod/
+Default TRUE (scaling applied)}
+
+\item{save_ts_dir, }{`character`. Full path to directory to save a CSV file
+of the time series of site averaged MODIS bands values.
+Default is NA (saving is disabled). 
+This directory is created if it does not exist.}
+
+\item{out_folder, }{`character`, Where to save the downloaded MODIS rasters
+Default is tempdir(). 
+(Note that if this parameter is left at default the downloaded MODIS rasters
+will be deleted on exit, along with other temporary R files.)}
+
+\item{show_products, }{a `boolean`, if TRUE, then show a list of all
+available MODIS products, and exit. Default is FALSE.}
+
+\item{show_bands, }{a `character`, the name of a MODIS product.
+The parameter could be one of: 
+    "M*D13Q1", (Vegetation Indexes_16Days_250m),
+    "M*D21A2", (LST_3band_emissivity_8day_1km),
+    "MCD12Q1". (LandCover_Type_Yearly_500m)
+When this parameter is supplied, show a list of the bands
+for that product, and exit. Default is NA.}
+}
+\value{
[…]
