+++ NAMESPACE
@@ -12,6 +12,7 @@ export(get_network_parameters)
 export(get_network_related_resources)
 export(get_network_research_topics)
 export(get_network_sites)
+export(get_site_MODIS)
 export(get_site_ODS)
 export(get_site_affiliations)
 export(get_site_boundaries)
@@ -33,6 +34,7 @@ export(taxon_id_pesi)
 export(taxon_id_worms)
 export(taxon_id_worms_refine)
 import(ISOcodes)
+import(MODIStsp)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(dplyr,as_tibble)
 importFrom(dplyr,bind_rows)
