### 2021-12-13 1b76dbb
Coerce UTF-8 to avoid error "invalid multibyte string"
