+++ R/get_site_general.R
@@ -9,7 +9,7 @@
 #' status, yearEstablished, yearClosed, hierarchy, siteName, short name, site
 #' type, protection level, images.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
-#' @importFrom httr GET content
+#' @importFrom httr RETRY content
 #' @importFrom utils capture.output
 #' @importFrom dplyr as_tibble
 #' @export
@@ -34,7 +34,7 @@ get_site_general <- function(deimsid) {
     "api/sites/",
     sub("^.+/", "", deimsid)
   )
-  export <- httr::GET(url = url)
+  export <- httr::RETRY("GET", url = url, times = 5)
   jj <- suppressMessages(httr::content(export, as="text", encoding="UTF-8"))
   status <- jj %>%
     jqr::jq(as.character("{status: .errors.status}")) %>%
