+++ R/get_dataset_info.R
@@ -7,7 +7,7 @@
 #' @return The output of the function is a `tibble` with main features
 #' of the site and the related resources collected by site.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
-#' @importFrom httr GET content
+#' @importFrom httr RETRY content
 #' @importFrom dplyr as_tibble
 #' @importFrom dtplyr lazy_dt
 #' @importFrom utils capture.output
@@ -49,7 +49,7 @@ get_dataset_info <- function(datasetid) {
     "api/datasets/",
     sub("^.+/", "", datasetid)
   )
-  export <- httr::GET(url = url)
+  export <- httr::RETRY("GET", url = url, times = 5)
   jj <- suppressMessages(httr::content(export, as = "text",
                                        encoding = "UTF-8"))
   status <- jj %>%
