### 2021-12-14 82d347a
New function get_id to easily run test in offline mode (to avoid HTTP error 500)
