+++ tests/testthat/test-get_ilter_parameters.R
@@ -5,12 +5,14 @@ library(testthat)
 skip_on_cran()
 
 test_that("Expect error if internet connection is down", {
+  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_ilter_parameters(sitesNum = 10)
     ),
     "GET"
   )
+  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
