+++ tests/testthat/setup.R
@@ -1,1 +1,2 @@
 library(httptest)
+Sys.setenv('LOCAL_DEIMS' = TRUE) # run get_id(..., test = TRUE)

[…]
