+++ tests/testthat/test-get_site_contact.R
@@ -3,6 +3,7 @@ message("\n---- Test get_site_contact() ----")
 library(testthat)
 
 test_that("Expect error if internet connection is down", {
+  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_site_contact(
@@ -11,6 +12,7 @@ test_that("Expect error if internet connection is down", {
     ),
     "GET"
   )
+  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
@@ -37,13 +39,17 @@ test_that("Output of site contact function constructs 'tibble' as expected", {
 })
 
 test_that("Wrong input (but URL) constructs a tibble with empty data", {
+  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
   result <- ReLTER::get_site_contact(
     deimsid = "https://deims.org/ljhnhbkihubib"
   )
   expect_type(result, "NULL")
+  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
+  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
   result <- ReLTER::get_site_contact(deimsid = "ljhnhbkihubib")
   expect_type(result, "NULL")
+  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
 })
