+++ tests/testthat/test-produce_site_parameters_waffle.R
@@ -3,6 +3,7 @@ message("\n---- Test produce_site_parameters_waffle() ----")
 library(testthat)
 
 test_that("Expect error if internet connection is down", {
+  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::produce_site_parameters_waffle(
@@ -11,6 +12,7 @@ test_that("Expect error if internet connection is down", {
     ),
     "GET"
   )
+  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
@@ -32,15 +34,19 @@ test_that("Output of chart waffle of parameters function constructs 'sf' and
 })
 
 test_that("Wrong input (but URL) constructs a NULL object", {
+  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
   result <- ReLTER::produce_site_parameters_waffle(
     deimsid = "https://deims.org/ljhnhbkihubib"
   )
   expect_type(result, "NULL")
+  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
+  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
   result <- ReLTER::produce_site_parameters_waffle(
     deimsid = "ljhnhbkihubib"
   )
   expect_type(result, "NULL")
+  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
 })
