+++ tests/testthat/test-get_ilter_envcharacts.R
@@ -19,7 +19,7 @@ skip_if_offline(host = "deims.org")
 
 test_that("Output of ILTER environmental characteristics function constructs
           ‘tibble’ as expected", {
-  result <- ReLTER::get_ilter_envcharacts(sitesNum = 50)
+  result <- ReLTER::get_ilter_envcharacts(sitesNum = 10)
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 26)
   expect_true(all(names(result) == c(
