### 2021-12-15 1b2cac7
Exclude some tests on CI
