+++ tests/testthat/test-get_ilter_generalinfo.R
@@ -16,6 +16,7 @@ test_that("Expect error if internet connection is down", {
 })
 
 skip_if_offline(host = "deims.org")
+skip_on_ci()
 
 test_that("Output of ILTER general info function constructs 'tibble' as
           expected", {
