+++ tests/testthat/test-get_network_research_topics.R
@@ -18,6 +18,7 @@ test_that("Expect error if internet connection is down", {
 })
 
 skip_if_offline(host = "deims.org")
+skip_on_ci()
 
 test_that("Output of network research topics function constructs ‘tibble’ as
           expected", {
