### 2021-12-15 c2c0b28
Run this test in normal mode
