### 2021-12-15 a5cf639
Avoid NOTEs during checks
