+++ R/get_site_contact.R
@@ -28,7 +28,7 @@ get_site_contact <- function(deimsid) {
        geoElev: .attributes.geographic.elevation,
        generalInfo: .attributes.contact
       }'
-  jj <- ReLTER:::get_id(deimsid, "sites")
+  jj <- get_id(deimsid, "sites")
   if (is.na(attr(jj, "status"))) {
     invisible(
       utils::capture.output(
