+++ R/get_site_related_resources.R
@@ -27,7 +27,7 @@ get_site_related_resources <- function(deimsid) {
        geoElev: .attributes.geographic.elevation,
        relatedResources: .attributes.relatedResources
       }'
-  jj <- ReLTER:::get_id(deimsid, "sites")
+  jj <- get_id(deimsid, "sites")
   if (is.na(attr(jj, "status"))) {
     invisible(
       utils::capture.output(
