### 2021-12-15 bfece11
define test variable for networkId
