+++ tests/testthat/test-get_network_sites.R
@@ -8,8 +8,8 @@ test_that("Expect error if internet connection is down", {
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_network_sites(
-        networkDEIMSID =
-          "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3")
+        networkDEIMSID = TESTURLNetwork
+      )
     ),
     "GET"
   )
@@ -20,8 +20,7 @@ skip_if_offline(host = "deims.org")
 test_that("Output of network sites information function constructs 'sf' and
           'data.frame' as expected", {
   result <- ReLTER::get_network_sites(
-    networkDEIMSID =
-      "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
+    networkDEIMSID = TESTURLNetwork
   )
   expect_s3_class(result, "sf")
   expect_s3_class(result, "tbl_df")
