+++ R/get_site_contact.R
@@ -11,7 +11,6 @@
 #' @importFrom httr GET content
 #' @importFrom utils capture.output
 #' @importFrom dplyr as_tibble
-#' @export
 #' @keywords internal
 #' @examples
 #' tSiteContact <- get_site_contact(
