+++ R/get_site_boundaries.R
@@ -21,7 +21,6 @@
 #' @importFrom sf st_as_sf write_sf st_write
 #' @importFrom leaflet leaflet addTiles addPolygons
 #' @importFrom mapview mapshot
-#' @export
 #' @keywords internal
 #' @examples
 #' tSiteBoundaries <- get_site_boundaries(
