+++ tests/testthat/test-get_network_related_resources.R
@@ -8,8 +8,7 @@ test_that("Expect error if internet connection is down", {
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_network_related_resources(
-        networkDEIMSID =
-          "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
+        networkDEIMSID = TESTURLNetwork
       )
     ),
     "GET"
@@ -21,8 +20,7 @@ skip_if_offline(host = "deims.org")
 test_that("Output of network related resources function constructs 'tibble' as
           expected", {
   result <- ReLTER::get_network_related_resources(
-    networkDEIMSID =
-      "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
+    networkDEIMSID = TESTURLNetwork
   )
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 3)
