+++ R/get_site_research_topics.R
@@ -10,7 +10,6 @@
 #' @importFrom httr GET content
 #' @importFrom utils capture.output
 #' @importFrom dplyr as_tibble
-#' @export
 #' @keywords internal
 #' @examples
 #' tSiteResearchTopics <- get_site_research_topics(
