+++ R/get_site_envcharacts.R
@@ -13,7 +13,6 @@
 #' @importFrom httr GET content
 #' @importFrom utils capture.output
 #' @importFrom dplyr as_tibble
-#' @export
 #' @keywords internal
 #' @examples
 #' tSiteEnvCharacts <- get_site_envcharacts(
