+++ tests/testthat/setup.R
@@ -1,1 +1,2 @@
 library(httptest)
+TESTURLNetwork = "https://deims.org/networks/e0f680c2-22b1-4424-bf54-58aa9b7476a0"
