+++ tests/testthat/test-get_network_research_topics.R
@@ -8,8 +8,8 @@ test_that("Expect error if internet connection is down", {
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_network_research_topics(
-        networkDEIMSID =
-          "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3")
+        networkDEIMSID = TESTURLNetwork
+      )
     ),
     "GET"
   )
@@ -20,8 +20,8 @@ skip_if_offline(host = "deims.org")
 test_that("Output of network research topics function constructs ‘tibble’ as
           expected", {
   result <- ReLTER::get_network_research_topics(
-    networkDEIMSID =
-      "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3")
+    networkDEIMSID = TESTURLNetwork
+  )
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 2)
   expect_true(all(names(result) == c(
