### 2021-12-15 09a0846
Add test variable for site deimsId
