+++ tests/testthat/test-get_site_ODS.R
@@ -6,7 +6,7 @@ test_that("Expect error if internet connection is down", {
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_site_info(
-        deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe",
+        deimsid = TESTURLSite,
         category = "Boundaries"
       )
     ),
