+++ tests/testthat/setup.R
@@ -1,3 +1,4 @@
 library(httptest)
-TESTURLNetwork = "https://deims.org/networks/e0f680c2-22b1-4424-bf54-58aa9b7476a0"
+TESTURLNetwork <- "https://deims.org/networks/e0f680c2-22b1-4424-bf54-58aa9b7476a0"
+TESTURLSite <- "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
 Sys.setenv('LOCAL_DEIMS' = TRUE) # run get_id(..., test = TRUE)
