+++ R/get_site_infrastructure.R
@@ -27,7 +27,7 @@ get_site_infrastructure <- function(deimsid) {
        geoElev: .attributes.geographic.elevation,
        generalInfo: .attributes.infrastructure
       }'
-  jj <- ReLTER:::get_id(deimsid, "sites")
+  jj <- get_id(deimsid, "sites")
   if (is.na(attr(jj, "status"))) {
     invisible(
       utils::capture.output(
