+++ tests/testthat/test-get_network_envcharacts.R
@@ -5,7 +5,7 @@ library(testthat)
 skip_on_cran()
 
 test_that("Expect error if internet connection is down", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_network_envcharacts(
@@ -14,7 +14,7 @@ test_that("Expect error if internet connection is down", {
     ),
     "GET"
   )
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
@@ -49,48 +49,24 @@ test_that("Output of network environmental characteristics function constructs
   expect_type(result$uri, "character")
   expect_type(result$geoCoord, "character")
   expect_type(result$country, "list")
-  #expect_type(result$geoElev.avg, "double")
-  #expect_type(result$geoElev.min, "double")
-  #expect_type(result$geoElev.max,"double") 
-  # in alcuni casi potrebbe essere "double"
-  expect_type(result$geoElev.unit, "character")
-
-  expect_type(result$envCharacteristics.airTemperature.avg, "double")
-  expect_type(result$envCharacteristics.airTemperature.min, "double")
-  expect_type(result$envCharacteristics.airTemperature.max, "double")
-  expect_type(result$envCharacteristics.airTemperature.unit, "character")
-  #expect_type(result$envCharacteristics.precipitation.annual, "double")
-  #expect_type(result$envCharacteristics.precipitation.min, "double")
-  #expect_type(result$envCharacteristics.precipitation.max, "double")
-  expect_type(result$envCharacteristics.precipitation.unit, "character")
-  expect_type(result$envCharacteristics.biogeographicalRegion, "character")
-  expect_type(result$envCharacteristics.biome, "character")
-  expect_type(result$envCharacteristics.ecosystemType, "list")
-  expect_type(result$envCharacteristics.eunisHabitat, "list")
-  #expect_type(result$envCharacteristics.landforms, "character")
-  #expect_type(result$envCharacteristics.geoBonBiome, "character")
-  #expect_type(result$envCharacteristics.geology, "character")
-  #expect_type(result$envCharacteristics.hydrology, "character")
-  #expect_type(result$envCharacteristics.soils, "character")
-  #expect_type(result$envCharacteristics.vegetation, "character")
 })
 
 test_that("Wrong input (but URL) constructs a NULL object", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   result <- ReLTER::get_network_envcharacts(
     networkDEIMSID = "https://deims.org/network/ljhnhbkihubib"
   )
   expect_true(is.null(result))
   expect_true(is.null(ncol(result)))
   expect_true(length(result) == 0)
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   result <- ReLTER::get_network_envcharacts(networkDEIMSID = "ljhnhbkihubib")
   expect_true(is.null(result))
[…]
