+++ tests/testthat/test-get_dataset_info.R
@@ -3,7 +3,7 @@ message("\n---- Test get_dataset_info() ----")
 library(testthat)
 
 test_that("Expect error if internet connection is down", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_dataset_info(
@@ -13,7 +13,7 @@ test_that("Expect error if internet connection is down", {
     ),
     "GET"
   )
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
@@ -114,19 +114,19 @@ test_that("Output of dataset function constructs ‘tibble’ as expected", {
 })
 
 test_that("Wrong input (but URL) constructs a NULL object", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   result <- ReLTER::get_dataset_info(
     datasetid = "https://deims.org/dataset/ljhnhbkihubib"
   )
   expect_type(result, "NULL")
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   result <- ReLTER::get_dataset_info(datasetid = "ljhnhbkihubib")
   expect_type(result, "NULL")
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
 })
 
 test_that("Output of get dataset information function constructs 'sf' with
