+++ tests/testthat/test-get_network_parameters.R
@@ -5,7 +5,7 @@ library(testthat)
 skip_on_cran()
 
 test_that("Expect error if internet connection is down", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
       result <- ReLTER::get_network_parameters(
@@ -14,7 +14,7 @@ test_that("Expect error if internet connection is down", {
     ),
     "GET"
   )
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
@@ -35,22 +35,21 @@ test_that("Output of network parameters function constructs 'tibble' as
 })
 
 test_that("Wrong input (but URL) constructs a NULL object", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   result <- ReLTER::get_network_parameters(
     networkDEIMSID = "https://deims.org/network/ljhnhbkihubib"
   )
   expect_true(is.null(result))
   expect_true(is.null(ncol(result)))
   expect_true(length(result) == 0)
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
-  Sys.setenv('LOCAL_DEIMS' = FALSE) # set online mode
+  Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   result <- ReLTER::get_network_parameters(networkDEIMSID = "ljhnhbkihubib")
   expect_true(is.null(result))
   expect_true(is.null(ncol(result)))
   expect_true(length(result) == 0)
-  Sys.setenv('LOCAL_DEIMS' = TRUE) # restore test mode
-})   
-
+  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+})
