+++ README.Rmd
@@ -218,3 +218,7 @@ For a exhaustive list of contributors please visit [authors page](https://oggion
 
 ![BGU](man/figures/bgu_logo.png){height="80"}
 <!--a href="https://in.bgu.ac.il/en/"><img src="man/figures/bgu_logo.png" height="40" align="left" /></a-->
+
+## :thumbsup: Acknowledgements 
+
+This work has been partially funded from the European Union’s Horizon 2020 research and innovation programme under the [H2020 eLTER-Plus Project](https://elter-ri.eu/elter-plus) grant agreement No 871128.
