### 2021-12-21 95adc4e
Delete R_newFunctions directory
