+++ R_newFunctions/todo_getDataset.R
@@ -1,170 +1,0 @@
-# 1. eLTER site: https://deims.org/api/sites/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6
-
-# 2. in the relatedResources the list of resources with prefix and suffix (e.g. prefix": "https://deims.org/dataset/","suffix": "cd1fb6f8-5e57-11e3-aa73-005056ab003f")
-
-# 3. eLTER relatedResources: 
-
-# 3.1 1st case: https://deims.org/api/datasets/cd1fb6f8-5e57-11e3-aa73-005056ab003f where the title is "B2Share Landing Page"
-# ... onlineDistribution": {
-#     "dataPolicyUrl": null,
-#     "doi": null,
-#     "onlineLocation": [
-#         {
-#             "function": "url",
-#             "url": {
-#                 "title": "B2Share Landing Page",
-#                 "value": "https://b2share.eudat.eu/records/2ec9cd568cfd40aa899beadcc8fe1b86"
-#             },
-#             "email": null
-#         }
-#     ]
-# }, ...
-
-# ISSUE:
-# a. for this dataset https://deims.org/dataset/de97b80c-8285-43d4-8ffa-1bea3671ed05 the landing page is the DOI of resources 
-#    (http://doi.org/10.23728/b2share.5db216134fd348e98c67e76fbc12630a). By this link is not usefull for reach or retrive (see point 3.1.1) the JSON version of the B2Share page of dataset 
-# b. if the repository of the data is different to b2share the "url$title" or the "landingPageURI" values must be changed in according with. 
-#    The reference to the file must be defined in DEIMS-SDR withou any other steps!
-# c. as a consequence of issue b - the "downloadLink", in the future, must be the unique valid link to reach the data!
-# d. in the case of https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef LTER Northern Adriatic Sea (Italy) marine data from 1965 to 2015 the ePIC_PID has 2 URI: i. the first point to the dataset
-#    ii. the second to the txt file for metadata of the dataset
-# SOLVED e. in the case of https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef LTER Northern Adriatic Sea (Italy) marine data from 1965 to 2015 the file is has a CSV extension.
-#    How to work with different extension without know before it? The b2share don't provide to a extension field.
-# f. in the case of CSV I don't know the separation characters
-# g. in the case of 'https://deims.org/dataset/cd1fb6f8-5e57-11e3-aa73-005056ab003f' LTER Zöbelboden, Austria, Air chemistry, 2012 the
-#    xsl is comprensive of more than one tabs and is not possible to reach which is the one with the data.
-# h. in the case of Zöbelboden site (e.g. 'https://deims.org/dataset/79cf34dd-0c01-4d67-809b-67416dfeccb0' LTER Zöbelboden, Austria, Foliage chemistry, 2019) the landing
-#    page link is the DOI provided by B2Share and, following this, is not possible to reach the URI of the landing page of B2share where I can reach the file.
-# i. in the case of all datasets provided by Mar Piccolo di Taranto site (e.g. 'https://deims.org/dataset/b7dfe900-9592-4154-9458-d99510b8e066') the metada cointain B2share DOI 
[…]
