### 2021-12-21 #485 @oggioniale
Dear @maurolepore,
we have been working in these last 2 weeks to close the issues arisen after `rcmdcheck` on the Ubuntu OSs.
Now we pretty sure that all is fine, also the [R-CMD-Check is fine](https://github.com/oggioniale/ReLTER/actions/runs/1607561807).

Please let we know the next steps.
