### 2021-12-22 #485 @maurolepore
That's great news! Thanks a lot for your efforts and sorry for my slow response. I've been very busy. I see the checks passing green on GH-Actions! :rocket:

The checks you're running are enough for now, but note DESCRIPTION declares this package works with [R >= 3.5.0](https://github.com/oggioniale/ReLTER/blob/bf50ec78b3b37752d5b5ac92a37567aec3c9a062/DESCRIPTION#L52) so eventually you'll need to extend your workflow to include old releases back to 3.5.0. For inspiration see this [check-full.yaml](https://github.com/r-lib/actions/blob/245dbd85428574c5ffbcd0e2417f6966e2f4c789/examples/check-full.yaml#L31).

I'm running the checks via our bot, then review this thread and if all is okay will start looking for reviewers. I'll keep you posted.

:crossed_fingers:
