### 2021-12-22 #485 @maurolepore
@oggioniale please note:

The bot fails to build the package (see above "2. goodpractice and other checks"):

```r
Error : Build failed, unknown error, standard output:

checking for file ‘ReLTER/DESCRIPTION’ ... OK
preparing ‘ReLTER’:
checking DESCRIPTION meta-information ... OK
installing the package to build vignettes
-----------------------------------
ERROR: dependencies ‘jqr’, ‘xslt’ are not available for package ‘ReLTER’
removing ‘/tmp/Rtmp1CudbO/Rinst63e435db54b2/ReLTER’
-----------------------------------
ERROR: package installation failed
```

And the [project on rstudio cloud](https://rstudio.cloud/project/3335565) shows many tests fail with an error similar to this one:

```r
Error (test-get_ilter_generalinfo.R:23:3): Output of ILTER general info function constructs 'tibble' as
          expected
Error in `get_id(deimsid, "sites")`: deimsid '5de9ae4c-c705-4cd5-b55a-2195c936f7fb' cannot be used in test mode with resource = 'sites'.
Backtrace:
 1. ReLTER::get_ilter_generalinfo(country_name = NA, site_name = NA) test-get_ilter_generalinfo.R:23:2
 2. base::lapply(...) /cloud/project/R/get_ilter_generalinfo.R:85:2
 3. ReLTER:::FUN(X[[i]], ...)
 4. ReLTER::get_id(deimsid, "sites") /cloud/project/R/get_site_info.R:48:2
 ```

The rstudio.cloud project may be a good platform for you to try explore why the package works fine in your own computers and on GitHub Actions but not on other environments.
