### 2021-12-22 #485 @mpadge
@maurolepore @oggioniale This could be due to the RStudio system deps server being intermittently unreliable, and our system then failing to install the system requirements for `jqr` and `xslt`. I'll re-deploy the check system with them pre-installed and run the checks again. Sorry for any inconvenience!
