### 2021-12-22 #485 @oggioniale
@maurolepore I'm disappointed on the situation.
Before to write the comment I checked all on local OS, otherwise I would never have told you that everything worked.

So
1. about the dependencies ‘jqr’, ‘xslt’ where the bot fails? Because the status of the bot in my GitHub profile is "Success";
2. I ran the `rcmdcheck::rcmdcheck(".")`in the [project on rstudio cloud](https://rstudio.cloud/project/3335565) and the output is
```
── R CMD check results ─────────────────────────────── ReLTER 1.0.0 ────
Duration: 5m 53.5s
0 errors ✓ | 0 warnings ✓ | 0 notes ✓
```

if I ran `goodpractice::gp(path = ".")` in the [project on rstudio cloud](https://rstudio.cloud/project/3335565) the output is

```
── GP ReLTER ───────────────────────────────────────────────
It is good practice to

    R/get_activity_info.R:39:NA
    R/get_activity_info.R:40:NA
    R/get_activity_info.R:41:NA
    R/get_activity_info.R:56:NA
    R/get_activity_info.R:57:NA
    ... and 1015 more lines
─────────────────────────────────────────────────────────
```

Nothing about the problem of the dependencies and of the errors.
Please, Let me know how to reproduce the errors, the issues or the fails.
