### 2021-12-22 #485 @oggioniale
@maurolepore
I added to che [check-standard.yaml](https://github.com/oggioniale/ReLTER/edit/dev/.github/workflows/check-standard.yaml) the installation of library libxslt1-dev.

I also changed the R version in DESCRIPTION.
