+++ .github/workflows/check-standard.yaml
@@ -85,7 +85,6 @@ jobs:
             libv8-dev \
             libssl-dev \
             libcairo2-dev \
-            libgit2-dev \
-            libxslt1-dev
+            libgit2-dev
 
       - uses: r-lib/actions/check-r-package@v1
