+++ DESCRIPTION
@@ -49,7 +49,7 @@ Description: ReLTER provides access to DEIMS-SDR (https://deims.org/),
 License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
-Depends: R (>= 4.1.2)
+Depends: R (>= 3.5.0)
 Imports: 
     RColorBrewer (>= 1.1-2),
     dplyr (>= 1.0.0),
