### 2021-12-22 #485 @maurolepore
@oggioniale,

I'm sorry this is so frustrating. You've come a long way and I'm confident you'll there soon.

I now realize running `rcmdcheck::rcmdcheck(".")` but I (and apparently the bot) is running the more common commands: `devtools::test()` and `devtools::check()`. You can also run them from RStudio's "build" tab:

Consider the project we've been using (https://rstudio.cloud/project/3335565) at this (last) commit:

```
commit 5f9b04796e446655761614b28cfecf693fa51526 (HEAD -> main, origin/main, origin/HEAD)
Author: Alessandro Oggioni <oggioniale@gmail.com>
  Date:   Wed Dec 22 15:14:42 2021 +0100

revert
```

Indeed I see that `rcmdcheck::rcmdcheck(".")` throws 0 errors, warnings, and notes.  But `devtools::test()` shows the same 7 failures that the bot detects. They are all similar to this one:

```
Error (test-get_ilter_generalinfo.R:23:3): Output of ILTER general info function constructs 'tibble' as
          expected
Error in `get_id(deimsid, "sites")`: deimsid '5de9ae4c-c705-4cd5-b55a-2195c936f7fb' cannot be used in test mode with resource = 'sites'.
Backtrace:
 1. ReLTER::get_ilter_generalinfo(country_name = NA, site_name = NA) test-get_ilter_generalinfo.R:23:2
 2. base::lapply(...) /cloud/project/R/get_ilter_generalinfo.R:85:2
 3. ReLTER:::FUN(X[[i]], ...)
 4. ReLTER::get_id(deimsid, "sites") /cloud/project/R/get_site_info.R:48:2
 ```

Please run `devtools::test()` yourself to confirm.
