### 2022-01-06 85ece40
Adds rescaling of NDVI
