### 2022-01-11 f52c700
Run test mode only on CRAN and CI
