+++ tests/testthat/test-produce_site_parameters_pie.R
@@ -12,7 +12,7 @@ test_that("Expect error if internet connection is down", {
     ),
     "GET"
   )
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
@@ -48,7 +48,7 @@ test_that("Wrong input (but URL) constructs a NULL object", {
     deimsid = "https://deims.org/ljhnhbkihubib"
   )
   expect_type(result, "NULL")
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
 
 test_that("Wrong input (not URL) constructs an empty tibble", {
@@ -57,5 +57,5 @@ test_that("Wrong input (not URL) constructs an empty tibble", {
     deimsid = "ljhnhbkihubib"
   )
   expect_type(result, "NULL")
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
