+++ tests/testthat/test-produce_network_points_map.R
@@ -14,7 +14,7 @@ test_that("Expect error if internet connection is down", {
     ),
     "GET"
   )
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
 
 skip_if_offline(host = "deims.org")
@@ -45,7 +45,7 @@ test_that("Wrong networkDEIMSID (but URL) constructs a NULL object", {
     countryCode = "DEU"
   )
   expect_type(result, "NULL")
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
 
 test_that("Wrong networkDEIMSID (not URL) constructs an empty tibble", {
@@ -55,7 +55,7 @@ test_that("Wrong networkDEIMSID (not URL) constructs an empty tibble", {
     countryCode = "DEU"
   )
   expect_type(result, "NULL")
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
 
 test_that("Wrong countryCode constructs a NULL object", {
@@ -85,7 +85,7 @@ test_that("Wrong both networkDEIMSID (but URL) and countryCode constructs
     countryCode = "EEA"
   )
   expect_type(result, "NULL")
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
 
 test_that("Wrong both networkDEIMSID (not URL) and countryCode constructs
@@ -96,7 +96,7 @@ test_that("Wrong both networkDEIMSID (not URL) and countryCode constructs
     countryCode = "EEA"
   )
   expect_type(result, "NULL")
-  Sys.setenv("LOCAL_DEIMS" = TRUE) # restore test mode
+  Sys.setenv("LOCAL_DEIMS" = test_mode) # restore test mode
 })
 
 test_that("Output of site affiliation information function constructs ‘sf' with
