+++ .github/workflows/check-standard.yaml
@@ -2,7 +2,7 @@
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
   push:
-    branches: [main]
+    branches: [main,dev]
   pull_request:
     branches: [main]
 
