+++ tests/testthat/test-get_ilter_envcharacts.R
@@ -8,7 +8,7 @@ test_that("Expect error if internet connection is down", {
   Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   testthat::expect_error(
     httptest::without_internet(
-      result <- ReLTER::get_ilter_envcharacts(sitesNum = 10)
+      result <- ReLTER::get_ilter_envcharacts(sitesNum = 5)
     ),
     "GET"
   )
@@ -19,7 +19,7 @@ skip_if_offline(host = "deims.org")
 
 test_that("Output of ILTER environmental characteristics function constructs
           ‘tibble’ as expected", {
-  result <- ReLTER::get_ilter_envcharacts(sitesNum = 10)
+  result <- ReLTER::get_ilter_envcharacts(sitesNum = 5)
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 26)
   expect_true(all(names(result) == c(
