### 2022-01-14 fe0f5d6
Remove dependency "mapview" (loading mapview causes warnings)
