+++ NAMESPACE
@@ -59,7 +59,6 @@ importFrom(leaflet,addPolygons)
 importFrom(leaflet,addTiles)
 importFrom(leaflet,leaflet)
 importFrom(magrittr,"%>%")
-importFrom(mapview,mapshot)
 importFrom(qrcode,qr_code)
 importFrom(raster,getData)
 importFrom(raster,reclassify)
