+++ DESCRIPTION
@@ -63,7 +63,6 @@ Imports:
     jsonlite (>= 1.7.2),
     leaflet (>= 2.0.3),
     magrittr (>= 2.0.1),
-    mapview (>= 2.9.0),
     qrcode (>= 0.1.3),
     raster (>= 3.3-13),
     rgeos (>= 0.5-3),
