### 2022-01-14 5c33594
Remove stringr and scales dependencies
