+++ R/get_site_info.R
@@ -21,7 +21,6 @@
 #' @importFrom dtplyr lazy_dt
 #' @importFrom dplyr as_tibble left_join
 #' @importFrom utils capture.output
-#' @importFrom stringr str_detect
 #' @importFrom leaflet leaflet addTiles addPolygons
 #' @export
 #' @examples
@@ -58,7 +57,7 @@ get_site_info <- function(deimsid, category = NA) {
       siteInfo <- siteInfo
     } else {
       # add 'Affiliations' info
-      if (any(stringr::str_detect(category, "Affiliations"))) {
+      if (any(grepl("Affiliations", category))) {
         siteAffil <- get_site_affiliations(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
@@ -78,7 +77,7 @@ get_site_info <- function(deimsid, category = NA) {
         siteInfo <- siteInfo
       }
       # add 'Contacts' info
-      if (any(stringr::str_detect(category, "Contacts"))) {
+      if (any(grepl("Contacts", category))) {
         siteConta <- get_site_contact(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
@@ -98,7 +97,7 @@ get_site_info <- function(deimsid, category = NA) {
         siteInfo <- siteInfo
       }
       # add 'EnvCharacts' info
-      if (any(stringr::str_detect(category, "EnvCharacts"))) {
+      if (any(grepl("EnvCharacts", category))) {
         siteEnvCh <- get_site_envcharacts(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
@@ -118,7 +117,7 @@ get_site_info <- function(deimsid, category = NA) {
         siteInfo <- siteInfo
       }
       # add 'General' info
-      if (any(stringr::str_detect(category, "General"))) {
+      if (any(grepl("General", category))) {
         siteGener <- get_site_general(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
@@ -138,7 +137,7 @@ get_site_info <- function(deimsid, category = NA) {
         siteInfo <- siteInfo
       }
       # add 'Infrastructure' info
-      if (any(stringr::str_detect(category, "Infrastructure"))) {
+      if (any(grepl("Infrastructure", category))) {
         siteInfra <- get_site_infrastructure(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
@@ -158,7 +157,7 @@ get_site_info <- function(deimsid, category = NA) {
         siteInfo <- siteInfo
       }
       # add 'Parameters' info
-      if (any(stringr::str_detect(category, "Parameters"))) {
+      if (any(grepl("Parameters", category))) {
         siteParam <- get_site_parameters(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
@@ -178,7 +177,7 @@ get_site_info <- function(deimsid, category = NA) {
         siteInfo <- siteInfo
       }
       # add 'RelateRes' info
-      if (any(stringr::str_detect(category, "RelateRes"))) {
+      if (any(grepl("RelateRes", category))) {
         siteRelat <- get_site_related_resources(deimsid = deimsid)
         siteInfo <- dplyr::left_join(
           siteInfo,
@@ -198,7 +197,7 @@ get_site_info <- function(deimsid, category = NA) {
         siteInfo <- siteInfo
       }
       # add 'ResearchTop' info
-      if (any(stringr::str_detect(category, "ResearchTop"))) {
+      if (any(grepl("ResearchTop", category))) {
         siteResea <- get_site_research_topics(deimsid = deimsid)
[…]
