+++ NAMESPACE
@@ -64,7 +64,6 @@ importFrom(raster,getData)
 importFrom(raster,reclassify)
 importFrom(rgeos,gSimplify)
 importFrom(rosm,osm.raster)
-importFrom(scales,percent)
 importFrom(sf,as_Spatial)
 importFrom(sf,st_as_sf)
 importFrom(sf,st_as_sfc)
@@ -75,7 +74,6 @@ importFrom(sf,st_simplify)
 importFrom(sf,st_transform)
 importFrom(sf,st_write)
 importFrom(sf,write_sf)
-importFrom(stringr,str_detect)
 importFrom(taxize,eubon_search)
 importFrom(terra,crop)
 importFrom(terra,crs)
