### 2022-01-14 c53e184
Clean some roxygen tags
