+++ .github/workflows/check-standard.yaml
@@ -91,7 +91,7 @@ jobs:
         # temporary, see https://stackoverflow.com/questions/70674136/r-package-warning-multiple-methods-tables-found-for-direction
         # remove when no longer needed
         run: |
-          install.packages("raster")
+          remotes::install_cran("raster", type = "source", repos = "https://stat.ethz.ch/CRAN/", force = TRUE)
         shell: Rscript {0}
 
       - uses: r-lib/actions/check-r-package@v1
