### 2022-01-18 #485 @oggioniale
@maurolepore,
the last commits and PR solved the failures of test and check when execute the `devtools::test()` and `devtools::check`.
As previously also `rcmdcheck::rcmdcheck(".")

Let me know please.
