+++ R/produce_zenodo_record.R
@@ -1,0 +1,194 @@
+#' @title eLTER produce_zenodo_record function
+#' @description This function allows to deposit a dataset to
+#' \href{https://zenodo.org}{Zenodo} repository. The function use the facility
+#' of `zen4r`package. Blondel, Emmanuel, & Barde, Julien. (2021). zen4R: R
+#' Interface to Zenodo REST API (0.5-2). Zenodo.
+#' \url{https://doi.org/10.5281/zenodo.5741143}.
+#' @param mytoken a `character`. Scopes assign permissions to your personal
+#' access token. A personal access token works just like a normal OAuth access
+#' token for authentication against the API. This token can be created
+#' \href{https://zenodo.org/account/settings/applications/tokens/new/}{here}.
+#' @param myfile a `character`. A file for a given Zenodo deposited record.
+#' @param dataset_type a `character`. It is the type of the record to be
+#' upload. Currently only "dataset" is possible, because the aim of the
+#' function is create a record of dataset in Zenodo repository. Default
+#' "dataset".
+#' @param dataset_title a `character`. The title of the dataset.
+#' @param dataset_description a `character`. The description of the dataset.
+#' @param dataset_authors a `tibble`. It is a list of creator(s) of the
+#' dataset. The approach is to use the firstname, lastname, affiliation,
+#' orcid of the authors of the dataset.
+#' @param dataset_license a `character`. Only one value is possible. It is the
+#' license with which the dataset is released. The license should be set
+#' with the Zenodo id of the license. Default "CC-BY-SA-4.0".
+#' @param dataset_accessRight a `character`. Only one value is possible.
+#' Default "open".
+#' @param dataset_version a `character`. It is a version of the dataset. Only
+#' one value is possible. Default "1.0".
+#' @param dataset_language a `character`. It is the language of the dataset.
+#' Only one value is possible. Default "eng".
+#' @param dataset_keywords a `character`. A multiple values are possible. The
+#' keyword to the record of dataset.
+#' @param dataset_relatedIdentifier a `tibble`. It is the related entities
+#' (e.g. scientific paper, data paper, etc.) with the detaset. The tibble is
+#' composed by 2 variables: relation and identifier. Relation can be one of
+#' among following values: "isCitedBy", "cites", "isSupplementTo",
+#' "isSupplementedBy", "isNewVersionOf", "isPreviousVersionOf",
+#' "isPartOf", "hasPart", "compiles", "isCompiledBy", "isIdenticalTo",
+#' "isAlternateIdentifier". While identifier is any type of identifier
+#' (e.g. DOI).
+#' @param dataset_community a `character`. It is a name of community as created
+#' in Zenodo. A multiple values are possible. Default "lter-italy".
+#' @param dataset_grants a `character`. A multiple values are possible. Put a
+#' list of project ids as well as showed by the European Commission via
+#' \href{http://www.openaire.eu/}{OpenAIRE}.
+#' @param dataset_publish a `boolean`. State whether the Zenodo record is to
+#' be published. The parameter publish can be set to TRUE (to use CAUTIOUSLY,
+#' only if you want to publish your record). Default FALSE.
+#' @return A link (URL) of the deposited record in Zenodo. The user must then:
[…]
