+++ R/get_site_speciesOcc.R
@@ -1,0 +1,235 @@
+#' @title eLTER get_site_speciesOcc function
+#' @description This function acquires data from
+#' GBIF \url{https://www.gbif.org} (via `rgbif`),
+#' iNaturalist \url{https://www.inaturalist.org/} and
+#' OBIS \url{https://obis.org/} and crops to an eLTER site
+#' boundary, which is obtained from the DEIMS-SDR sites API.
+#' @param deimsid  a `character`. The DEIMS ID of the site from
+#' DEIMS-SDR website. More information about DEIMS ID from:
+#' \href{https://deims.org/docs/deimsid.html}{page}.
+#' @param list_DS a `character`. Data source to get data from, any
+#' combination of gbif, inat and/or obis.
+#' @param show_map a `boolean`. If TRUE a Leaflet map with occurences
+#' is shown. Default FALSE.
+#' @param limit a `numeric`. Number of records to return. This is passed
+#' across all sources. To specify different limits for each source, use
+#' the options for each source (gbifopts, bisonopts, inatopts, and ebirdopts).
+#' See Details for more. Default: 500 for each source. BEWARE: if you have a
+#' lot of species to query for (e.g., n = 10), that's 10 * 500 = 5000, which
+#' can take a while to collect. So, when you first query, set the limit to
+#' something smallish so that you can get a result quickly, then do more as
+#' needed.
+#' @return The output of the function is a tibble with the list of occurrences
+#' retived from species data sources selected into parameter list_DS.
+#' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
+#' @importFrom leaflet layersControlOptions addLayersControl addLegend
+#' @importFrom leaflet addCircleMarkers addTiles addProviderTiles leaflet
+#' @importFrom leaflet colorFactor
+#' @importFrom RColorBrewer brewer.pal
+#' @importFrom tibble as_tibble
+#' @importFrom dplyr select mutate
+#' @importFrom spocc occ2df obis_search occ
+#' @importFrom sf st_as_text
+#' @importFrom rgbif wkt_parse
+#' @export
+#' @examples
+#' \dontrun{
+#' # terrestrial site Saldur River Catchment
+#' occ_SRC <- get_site_speciesOcc(
+#'   deimsid =
+#'   "https://deims.org/97ff6180-e5d1-45f2-a559-8a7872eb26b1",
+#'   list_DS = c("gbif", "inat"),
+#'   show_map = TRUE,
+#'   limit = 100
+#' )
+#' occ_SRC
+#'
+#' # marine site Gulf of Venice only obis
+#' occ_GoV <- get_site_speciesOcc(
+#'   deimsid =
+#'   "https://deims.org/758087d7-231f-4f07-bd7e-6922e0c283fd",
+#'   list_DS = c("obis"),
+#'   show_map = TRUE,
+#'   limit = 100
+#' )
+#' occ_GoV
+#'
+#' # marine site Gulf of Venice only gbif, inat and obis
+#' occ_GoV_all <- get_site_speciesOcc(
+#'   deimsid =
+#'   "https://deims.org/758087d7-231f-4f07-bd7e-6922e0c283fd",
+#'   list_DS = c("gbif", "inat", "obis"),
+#'   show_map = TRUE,
+#'   limit = 300
+#' )
+#' occ_GoV_all
+#' }
+#'
+### function get_site_speciesOcc
+get_site_speciesOcc <- function(
+  deimsid,
+  list_DS,
+  show_map = FALSE,
+  limit = 500
+) {
+  # First check that site has a boundary ----
+  boundary <- ReLTER::get_site_info(
+    deimsid,
[…]
