### 2022-01-18 #485 @maurolepore
Yay! Congratulations @oggioniale! I'll start looking for reviewers.

In the mean time please try increase the test coverage. I see "Package coverage is 61.3% (should be at least 75%)".

Reading this short section may help: https://mastering-shiny.org/scaling-testing.html#code-coverage
