+++ DESCRIPTION
@@ -80,7 +80,7 @@ Imports:
     xml2 (>= 1.3.2),
     xslt (>= 1.4.3),
     zen4R (>= 0.5-2),
-    spocc (>= 1.2.0.91)
+    spocc (>= 1.2.0)
 Suggests: 
     httptest,
     knitr,
