### 2022-01-20 #485 @maurolepore
@oggioniale,

Could you please list 3 people you think might be good reviewers for ReLTER and briefly explain why? I would aim to use only one but your list of three -- along with your explanations -- would help me better understand the ways in which you think an external eye can add value.

I'll complement the list with a search of my own.
