### 2022-01-20 #485 @oggioniale
@maurolepore

I would suggest:

1. [Will Bolton](https://github.com/WillOnGit) for his involvement in the activities of the project ([eLTER-Plus](https://elter-ri.eu/elter-plus)) in which the ReLTER package was born.
2. [Daniel Nüst](https://github.com/nuest) for the spatio-temporal and geoinformatics skills, but also for his focus on the principles of open and reproducible science.
3. [Carl Boettiger](https://github.com/cboettig) for his closeness to the activities of the community of LTER US (the ReLTER package was born in the European LTER community). He is a science adviser to NCEAS - National Center for Ecological Analysis and Synthesis, the ReLTER package is dedicated to the Ecological scientists community. But also for the interests in open science, data science, and ecoinformatics.
