+++ DESCRIPTION
@@ -79,8 +79,7 @@ Imports:
     worrms (>= 0.4.2),
     xml2 (>= 1.3.2),
     xslt (>= 1.4.3),
-    zen4R (>= 0.5-2),
-    spocc (>= 1.2.0.91)
+    zen4R (>= 0.5-2)
 Suggests: 
     httptest,
     knitr,
