### 2022-01-21 8ef4af9
Revert "excluding the use of rgbif package"
This reverts commit 6286e45ecf5ab3f279a613fb7cf53e56cd4f031e.
