+++ R/get_site_speciesOcc.R
@@ -29,7 +29,8 @@
 #' @importFrom tibble as_tibble
 #' @importFrom dplyr select mutate
 #' @importFrom spocc occ2df obis_search occ
-#' @importFrom sf st_as_text st_as_sfc st_bbox
+#' @importFrom sf st_as_text
+#' @importFrom rgbif wkt_parse
 #' @export
 #' @examples
 #' \dontrun{
@@ -80,13 +81,10 @@ get_site_speciesOcc <- function(
     print("No boundary for requested DEIMS site.")
     return(NULL)
   } else {
-    bbox_wkt <- sf::st_as_text(
-      sf::st_as_sfc(
-        sf::st_bbox(
-          boundary
-        )
-      )
-    )
+    site_wkt <- sf::st_as_text(boundary$boundaries,
+                               EWKT = FALSE)
+    bbox_wkt <- rgbif::wkt_parse(site_wkt,
+                                 geom_big = "bbox")
   }
 
   # download occurrence by SPOCC by provide data sources ----
@@ -159,7 +157,7 @@ get_site_speciesOcc <- function(
       leaflet::addTiles(
         "http://{s}.basemaps.cartocdn.com/light_only_labels/{z}/{x}/{y}.png"
       )
-
+  
     groups <- unique(occ_df$prov)
     for (i in groups) {
       data <- occ_df[occ_df$prov == i, ]
@@ -192,7 +190,7 @@ get_site_speciesOcc <- function(
           collapsed = FALSE
         )
       )
-
+  
     print(occ_map)
     # create tibble ----
     occ_list <- vector(
@@ -211,7 +209,7 @@ get_site_speciesOcc <- function(
     if ("obis" %in% list_DS) {
       occ_list$obis <- site_occ_spocc_obis$results
     }
-
+  
     occ_list
   } else {
     occ_map <- NULL
