+++ NAMESPACE
@@ -70,6 +70,7 @@ importFrom(magrittr,"%>%")
 importFrom(qrcode,qr_code)
 importFrom(raster,getData)
 importFrom(raster,reclassify)
+importFrom(rgbif,wkt_parse)
 importFrom(rgeos,gSimplify)
 importFrom(rosm,osm.raster)
 importFrom(sf,as_Spatial)
