### 2022-01-21 34852da
add missing tests to increase coverage
