+++ tests/testthat/test-get_ilter_parameters.R
@@ -33,3 +33,14 @@ test_that("Wrong input (not a double) constructs an empty tibble", {
   result <- ReLTER::get_ilter_parameters(sitesNum = "aa")
   expect_type(result, "NULL")
 })
+
+test_that("If sitesNum is 0", {
+  lterILTERSites <- as.list(
+    jsonlite::fromJSON("https://deims.org/api/sites")
+  )
+  number_sites <- length(lterILTERSites$title)
+  result <- ReLTER::get_ilter_envcharacts(sitesNum = 0)
+  expect_s3_class(result, "tbl_df")
+  expect_true(ncol(result) == 2)
+  expect_true(nrow(result) == number_sites)
+})
