+++ tests/testthat/test-taxon_id_worms_refine.R
@@ -2,6 +2,8 @@ message("\n---- Test taxon_id_worms_refine() ----")
 
 library(testthat)
 
+skip_on_ci()
+
 test_that("Output of taxon worms function constructs 'tibble' as expected
 with choiceNumber NULL", {
   input_table <- tibble::tibble(
