+++ tests/testthat/test-get_sos_procedurelist.R
@@ -2,6 +2,8 @@ message("\n---- Test get_sos_procedurelist() ----")
 
 library(testthat)
 
+skip_on_ci()
+
 test_that("Output of SOS procedure list function constructs 'list'
           as expected", {
             result <- get_sos_procedurelist(
