+++ R/get_sos_procedurelist.R
@@ -12,7 +12,9 @@
 #' @export
 #' @examples
 #' \dontrun{
-#' get_sos_procedurelist(sosHost = "http://getit.lteritalia.it/observations/sos/kvp?")
+#' get_sos_procedurelist(
+#'   sosHost = "http://getit.lteritalia.it/observations/sos/kvp?"
+#' )
 #' }
 #'
 ### function get_sos_procedurelist
