+++ R/get_dataset_info.R
@@ -127,4 +127,3 @@ Please check again the Dataset.iD\n----\n")
   print(map)
   geoDataset
 }
-  
