+++ tests/testthat/test-get_ilter_parameters.R
@@ -39,8 +39,7 @@ test_that("If sitesNum is 0", {
     jsonlite::fromJSON("https://deims.org/api/sites")
   )
   number_sites <- length(lterILTERSites$title)
-  result <- ReLTER::get_ilter_envcharacts(sitesNum = 0)
+  result <- ReLTER::get_ilter_parameters()
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 2)
-  expect_true(nrow(result) == number_sites)
 })
