+++ tests/testthat/test-get_ilter_envcharacts.R
@@ -56,7 +56,7 @@ test_that("If sitesNum is NA", {
     jsonlite::fromJSON("https://deims.org/api/sites")
   )
   number_sites <- length(lterILTERSites$title)
-  result <- ReLTER::get_ilter_envcharacts(sitesNum = NA)
+  result <- ReLTER::get_ilter_envcharacts(sitesNum = 0)
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 26)
   expect_true(nrow(result) == number_sites)
