### 2022-01-25 ba7eef3
Revert "Codecov in the package"
