+++ .travis.yml
@@ -1,15 +1,0 @@
-language: r
-
-sudo: required
-
-env: _R_CHECK_CRAN_INCOMING_=FALSE
-
-r_packages:
-  - covr
-  - devtools
-
-r_github_packages:
-  - oggioniale/ReLTER
-
-after_success:
-  - Rscript -e 'library(devtools);test_coverage()'
