### 2022-01-25 #485 @oggioniale
@maurolepore
The last commit in the repo improve the code coverage (now more than 82%).

Ale
