+++ .github/workflows/check-standard.yaml
@@ -95,3 +95,7 @@ jobs:
         shell: Rscript {0}
 
       - uses: r-lib/actions/check-r-package@v1
+
+      - name: Test coverage
+        run: covr::codecov()
+        shell: Rscript {0}
