+++ .github/workflows/check-standard.yaml
@@ -97,5 +97,5 @@ jobs:
       - uses: r-lib/actions/check-r-package@v1
 
       - name: Test coverage
-        run: covr::codecov()
+        run: devtools::test_coverage()
         shell: Rscript {0}
