+++ DESCRIPTION
@@ -83,7 +83,8 @@ Suggests:
     rmarkdown,
     ISOcodes,
     testthat (>= 3.0.0),
-    devtools
+    devtools,
+    covr
 VignetteBuilder: knitr
 URL: https://github.com/oggioniale/ReLTER
 BugReports: https://github.com/oggioniale/ReLTER/issues
