+++ DESCRIPTION
@@ -84,7 +84,8 @@ Suggests:
     ISOcodes,
     testthat (>= 3.0.0),
     devtools,
-    covr
+    covr,
+    DT
 VignetteBuilder: knitr
 URL: https://github.com/oggioniale/ReLTER
 BugReports: https://github.com/oggioniale/ReLTER/issues
