+++ .github/workflows/check-standard.yaml
@@ -95,7 +95,13 @@ jobs:
         shell: Rscript {0}
 
       - uses: r-lib/actions/check-r-package@v1
-
+      
+      - name: Install dependencies
+        run: |
+          install.packages(c("devtools"))
+          devtools::install_dev_deps(dependencies = TRUE)
+        shell: Rscript {0}
+        
       - name: Test coverage
         run: devtools::test_coverage()
         shell: Rscript {0}
