+++ DESCRIPTION
@@ -83,7 +83,6 @@ Suggests:
     rmarkdown,
     ISOcodes,
     testthat (>= 3.0.0),
-    devtools,
     covr
 VignetteBuilder: knitr
 URL: https://github.com/oggioniale/ReLTER
