### 2022-01-27 d5e84c3
Implements get_site_MODIS function
