+++ tests/testthat/test-get_site_MODIS.R
@@ -2,7 +2,7 @@ message("\n---- Test get_site_MODIS() ----")
 
 library(testthat)
 skip_if_offline(host = "deims.org")
-skip_if_offline(host = "https://e4ftl01.cr.usgs.gov/")
+skip_if_offline(host = "e4ftl01.cr.usgs.gov")
 
 # Check for requested products or bands list (function returns NULL)
 test_that("When show_products is TRUE, or show_bands is character,
@@ -14,84 +14,72 @@ test_that("When show_products is TRUE, or show_bands is character,
 })
 
 # Check for password and username
-test_that("Username and password are character strings", {
-  result1 <- get_site_MODIS(
-          deimsid = "https://deims.org/86e3a1ca-b2ba-4b06-b096-71447df52841",
+test_that("Username/Password are character string", {
+          expect_error(get_site_MODIS(
+          deimsid = "https://deims.org/45722713-80e3-4387-a47b-82c97a6ef62b",
           earthdata_user = "abc",
           earthdata_passwd = NULL,
           from_date = "2017.01.01",
-          to_date = "2017.30.03"
-          )
-  result2 <- get_site_MODIS(
-          deimsid = "https://deims.org/86e3a1ca-b2ba-4b06-b096-71447df52841",
+          to_date = "2017.30.03"), NULL)
+
+          expect_error(get_site_MODIS(
+          deimsid = "https://deims.org/45722713-80e3-4387-a47b-82c97a6ef62b",
           earthdata_user = NULL,
           earthdata_passwd = "abc",
           from_date = "2017.01.01",
-          to_date = "2017.30.03"
-          )
-  expect_error(result1, regexp = "Missing")
-  expect_error(result2, regexp = "Missing")
+          to_date = "2017.30.03"), NULL)
 })
 
 # Check for invalid dates
 test_that("Check for invalid dates", {
-  result1 <- get_site_MODIS(
-          deimsid = "https://deims.org/86e3a1ca-b2ba-4b06-b096-71447df52841",
+  expect_error(get_site_MODIS(
+          deimsid = "https://deims.org/45722713-80e3-4387-a47b-82c97a6ef62b",
           earthdata_user = "abc",
           earthdata_passwd = "abc",
           from_date = "1999.01.01",
-          to_date = "2002.30.03"
-          )
-  result2 <- get_site_MODIS(
-          deimsid = "https://deims.org/86e3a1ca-b2ba-4b06-b096-71447df52841",
+          to_date = "2002.30.03"), NULL)
+  expect_error(get_site_MODIS(
+          deimsid = "https://deims.org/45722713-80e3-4387-a47b-82c97a6ef62b",
           earthdata_user = "abc",
           earthdata_passwd = "abc",
           from_date = "2017.01.01",
-          to_date = "2017.13.03"
-          )
-  result3 <- get_site_MODIS(
-          deimsid = "https://deims.org/86e3a1ca-b2ba-4b06-b096-71447df52841",
+          to_date = "2017.13.03"), NULL)
+  expect_error(get_site_MODIS(
+          deimsid = "https://deims.org/45722713-80e3-4387-a47b-82c97a6ef62b",
           earthdata_user = "abc",
           earthdata_passwd = "abc",
           from_date = "2017.02.01",
-          to_date = "2017.02.30"
-          )
-  result4 <- get_site_MODIS(
-          deimsid = "https://deims.org/86e3a1ca-b2ba-4b06-b096-71447df52841",
+          to_date = "2017.02.30"), NULL)
+  expect_error(get_site_MODIS(
+          deimsid = "https://deims.org/45722713-80e3-4387-a47b-82c97a6ef62b",
           earthdata_user = "abc",
           earthdata_passwd = "abc",
[…]
