+++ NAMESPACE
@@ -23,8 +23,6 @@ export(produce_site_parameters_waffle)
 export(produce_site_qrcode)
 export(taxon_id_pesi)
 export(taxon_id_worms)
-export(taxon_id_worms_refine)
-import(ISOcodes)
 import(MODIStsp)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(dplyr,as_tibble)
