### 2022-01-27 d1f1fdb
Fixes time series plot output
