### 2022-01-27 0c3182d
Fixes check for NDVI in get_site_ODS
