### 2022-01-27 82d3d86
Adds MODIStsp to DESCRIPTION file
