+++ DESCRIPTION
@@ -76,7 +76,8 @@ Imports:
     waffle (>= 0.7.0),
     worrms (>= 0.4.2),
     xml2 (>= 1.3.2),
-    xslt (>= 1.4.3)
+    xslt (>= 1.4.3),
+    MODIStsp
 Suggests: 
     httptest,
     knitr,
