### 2022-01-27 2fd83c4
Fixes NDVI rescaling in get_site_ODS
