### 2022-01-27 5293309
Fixes access to MODIS credentials from sysdata.rda
