### 2022-01-27 a21bf25
Adds MODIStsp package to DEPENDENCIES
