### 2022-01-27 a21d739
Adds MODIStsp package to DEPENDENCIES
