+++ DESCRIPTION
@@ -49,7 +49,8 @@ Description: ReLTER provides access to DEIMS-SDR (https://deims.org/),
 License: GPL (>= 3)
 Encoding: UTF-8
 LazyData: true
-Depends: R (>= 3.5.0)
+Depends: 
+    R (>= 3.5.0),
 Imports: 
     RColorBrewer (>= 1.1-2),
     dplyr (>= 1.0.0),
@@ -77,7 +78,7 @@ Imports:
     worrms (>= 0.4.2),
     xml2 (>= 1.3.2),
     xslt (>= 1.4.3),
-    MODIStsp
+    MODIStsp (>= 2.0.0),
 Suggests: 
     httptest,
     knitr,
