### 2022-01-28 6e56a78
new function get_site_speciesOcc
