+++ NAMESPACE
@@ -15,6 +15,7 @@ export(get_network_sites)
 export(get_site_MODIS)
 export(get_site_ODS)
 export(get_site_info)
+export(get_site_speciesOcc)
 export(get_sos_procedurelist)
 export(produce_network_points_map)
 export(produce_site_map)
@@ -56,9 +57,15 @@ importFrom(httr,content)
 importFrom(jqr,jq)
 importFrom(jsonlite,fromJSON)
 importFrom(jsonlite,stream_in)
+importFrom(leaflet,addCircleMarkers)
+importFrom(leaflet,addLayersControl)
+importFrom(leaflet,addLegend)
 importFrom(leaflet,addMarkers)
 importFrom(leaflet,addPolygons)
+importFrom(leaflet,addProviderTiles)
 importFrom(leaflet,addTiles)
+importFrom(leaflet,colorFactor)
+importFrom(leaflet,layersControlOptions)
 importFrom(leaflet,leaflet)
 importFrom(magrittr,"%>%")
 importFrom(qrcode,qr_code)
@@ -69,6 +76,7 @@ importFrom(rosm,osm.raster)
 importFrom(sf,as_Spatial)
 importFrom(sf,st_as_sf)
 importFrom(sf,st_as_sfc)
+importFrom(sf,st_as_text)
 importFrom(sf,st_bbox)
 importFrom(sf,st_crs)
 importFrom(sf,st_is_valid)
@@ -76,6 +84,9 @@ importFrom(sf,st_simplify)
 importFrom(sf,st_transform)
 importFrom(sf,st_write)
 importFrom(sf,write_sf)
+importFrom(spocc,obis_search)
+importFrom(spocc,occ)
+importFrom(spocc,occ2df)
 importFrom(taxize,eubon_search)
 importFrom(terra,crop)
 importFrom(terra,crs)
