+++ .github/workflows/test-coverage.yml
@@ -1,12 +1,6 @@
 on:
   push:
-    branches:
-      - main
-      - master
-  pull_request:
-    branches:
-      - main
-      - master
+    branches: [main,dev]
 
 name: test-coverage
 
