### 2022-01-28 655c001
Exit gracefully from get_site_MODIS in case of timeout
