+++ .github/workflows/check-standard.yaml
@@ -95,7 +95,3 @@ jobs:
         shell: Rscript {0}
 
       - uses: r-lib/actions/check-r-package@v1
-      
-      - name: Test coverage
-        run: covr::codecov()
-        shell: Rscript {0}
