+++ .github/workflows/test-coverage.yml
@@ -9,7 +9,6 @@ jobs:
     runs-on: ubuntu-20.04
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
     steps:
       - uses: actions/checkout@v2
 
@@ -57,5 +56,5 @@ jobs:
         shell: Rscript {0}
 
       - name: Test coverage
-        run: covr::codecov(token = Sys.getenv("CODECOV_TOKEN"))
+        run: covr::codecov(quiet = FALSE)
         shell: Rscript {0}
