+++ .github/workflows/check-standard.yaml
@@ -96,7 +96,3 @@ jobs:
 
       - uses: r-lib/actions/check-r-package@v1
       
-      - name: Test coverage
-        run: covr::codecov(quiet = FALSE)
-        shell: Rscript {0}
-
