### 2022-01-31 0446517
function styling with lintr
