+++ R/get_site_MODIS.R
@@ -2,8 +2,8 @@
 #' @description This function acquires MODIS products
 #' from MODIS Land Products archive,
 #' crops to an eLTER site boundary, and
-#' plots a time series of the MODIS derived variable averaged over the eLTER site.
-#'
+#' plots a time series of the MODIS derived variable averaged over
+#' the eLTER site.
 #' @param deimsid  a `character`. The DEIMS ID of the site from
 #' DEIMS-SDR website. More information about DEIMS ID from:
 #' \href{https://deims.org/docs/deimsid.html}{page}.
@@ -18,13 +18,13 @@
 #' @param bands `character`, string array of bands in the requested product.
 #' Use the "show_bands" parameter first to get a list of available bands.
 #' Some examples: "LST_Day_1km", or multiple bands as c("NDVI", "EVI")
-#' Default is c("NDVI", "EVI")  
+#' Default is c("NDVI", "EVI").
 #' @param scale `boolean`. The MODIS raster values will be scaled (multiplied)
 #' by a factor. Some MODIS products are scaled up and saved as 16 bit integers.
 #' For example, Vegetation indices are scaled by 10000,
 #' and Land surface temperatures are scaled by 50.
 #' Scaling factors are derived from product metadata.
-#' For additional details refer to the  MODIS Product Tables: 
+#' For additional details refer to the  MODIS Product Tables:
 #' https://modis.gsfc.nasa.gov/data/dataprod/
 #' Default TRUE (scaling applied)
 #' @param from_date, `character`, the start date for acquiring MODIS
@@ -33,37 +33,36 @@
 #' @param to_date, `character`, the last date for acquiring MODIS
 #' formatted as YYYY.MM.DD.
 #' @param out_folder, `character`, Where to save the downloaded MODIS rasters
-#' Default is tempdir(). 
+#' Default is tempdir().
 #' (Note that if this parameter is left at default the downloaded MODIS rasters
 #' will be deleted on exit, along with other temporary R files.)
 #' @param save_ts_dir, `character`. Full path to directory to save a CSV file
 #' of the time series of site averaged MODIS bands values.
-#' Default is NA (saving is disabled). 
+#' Default is NA (saving is disabled).
 #' This directory is created if it does not exist.
-#' @param earthdata_user `character`. For authentication. 
+#' @param earthdata_user `character`. For authentication.
 #' Please create an account on: https://urs.earthdata.nasa.gov/users/new
 #' and enter authentication details here.
 #' @param earthdata_passwd `character` For authentication on Earthdata.
 #' @param show_products, a `boolean`, if TRUE, then show a list of all
 #' available MODIS products, and exit. Default is FALSE.
 #' @param show_bands, a `character`, the name of a MODIS product.
-#' The parameter could be one of: 
+#' The parameter could be one of:
 #'     "M*D13Q1", (Vegetation Indexes_16Days_250m),
 #'     "M*D21A2", (LST_3band_emissivity_8day_1km),
 #'     "MCD12Q1". (LandCover_Type_Yearly_500m)
 #' When this parameter is supplied, show a list of the bands
 #' for that product, and exit. Default is NA.
 #'
-#' @details The Nasa [EarthData](https://urs.earthdata.nasa.gov/home) site 
-#' offers MODIS products as 8 day, 16 day and  monthly aggregations. 
+#' @details The Nasa [EarthData](https://urs.earthdata.nasa.gov/home) site
+#' offers MODIS products as 8 day, 16 day and  monthly aggregations.
 #' This function accesses those aggregations
[…]
