+++ R/get_site_speciesOccurrences.R
@@ -1,4 +1,4 @@
-#' @title eLTER get_site_speciesOcc function
+#' @title eLTER get_site_speciesOccurrences function
 #' @description This function acquires data from
 #' GBIF \url{https://www.gbif.org} (via `rgbif`),
 #' iNaturalist \url{https://www.inaturalist.org/} and
@@ -9,7 +9,7 @@
 #' \href{https://deims.org/docs/deimsid.html}{page}.
 #' @param list_DS a `character`. Data source to get data from, any
 #' combination of gbif, inat and/or obis.
-#' @param show_map a `boolean`. If TRUE a Leaflet map with occurences
+#' @param show_map a `boolean`. If TRUE a Leaflet map with occurrences
 #' is shown. Default FALSE.
 #' @param limit a `numeric`. Number of records to return. This is passed
 #' across all sources. Default: 500 for each source. BEWARE: if you have a
@@ -17,8 +17,8 @@
 #' can take a while to collect. So, when you first query, set the limit to
 #' something smallish so that you can get a result quickly, then do more as
 #' needed.
-#' @return The output of the function is a tibble with the list of occurrences
-#' retived from species data sources selected into parameter list_DS.
+#' @return The output of the function is a `list` of `sf` one for each of the
+#' repositories specified in the list_DS parameter.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom leaflet layersControlOptions addLayersControl addLegend
 #' @importFrom leaflet addCircleMarkers addTiles addProviderTiles leaflet
@@ -27,32 +27,33 @@
 #' @importFrom tibble as_tibble
 #' @importFrom dplyr select mutate
 #' @importFrom spocc occ2df obis_search occ
-#' @importFrom sf st_as_text st_as_sfc st_bbox
+#' @importFrom sf st_as_text st_as_sfc st_bbox st_as_sf
 #' @export
 #' @examples
 #' \dontrun{
 #' # terrestrial site Saldur River Catchment
-#' occ_SRC <- get_site_speciesOcc(
+#' occ_SRC <- get_site_speciesOccurrences(
 #'   deimsid =
 #'   "https://deims.org/97ff6180-e5d1-45f2-a559-8a7872eb26b1",
 #'   list_DS = c("gbif", "inat"),
-#'   show_map = TRUE,
+#'   show_map = FALSE,
 #'   limit = 100
 #' )
 #' occ_SRC
 #'
 #' # marine site Gulf of Venice only obis
-#' occ_GoV <- get_site_speciesOcc(
+#' occ_GoV <- get_site_speciesOccurrences(
 #'   deimsid =
 #'   "https://deims.org/758087d7-231f-4f07-bd7e-6922e0c283fd",
-#'   list_DS = c("obis"),
+#'   list_DS = "obis",
 #'   show_map = TRUE,
 #'   limit = 100
 #' )
 #' occ_GoV
 #'
-#' # marine site Gulf of Venice only gbif, inat and obis
-#' occ_GoV_all <- get_site_speciesOcc(
+#' # marine site Gulf of Venice, all repositories are invoked
+#' # gbif, inat and obis
+#' occ_GoV_all <- get_site_speciesOccurrences(
 #'   deimsid =
 #'   "https://deims.org/758087d7-231f-4f07-bd7e-6922e0c283fd",
 #'   list_DS = c("gbif", "inat", "obis"),
@@ -62,8 +63,8 @@
 #' occ_GoV_all
 #' }
 #'
-### function get_site_speciesOcc
-get_site_speciesOcc <- function(
+### function get_site_speciesOccurrences
[…]
