### 2022-01-31 3292bf6
changes into package info after commits about functions
