+++ man/get_site_speciesOccurrences.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/get_site_speciesOcc.R
-\name{get_site_speciesOcc}
-\alias{get_site_speciesOcc}
-\title{eLTER get_site_speciesOcc function}
+% Please edit documentation in R/get_site_speciesOccurrences.R
+\name{get_site_speciesOccurrences}
+\alias{get_site_speciesOccurrences}
+\title{eLTER get_site_speciesOccurrences function}
 \usage{
-get_site_speciesOcc(deimsid, list_DS, show_map = FALSE, limit = 500)
+get_site_speciesOccurrences(deimsid, list_DS, show_map = FALSE, limit = 500)
 }
 \arguments{
 \item{deimsid}{a `character`. The DEIMS ID of the site from
@@ -14,7 +14,7 @@ DEIMS-SDR website. More information about DEIMS ID from:
 \item{list_DS}{a `character`. Data source to get data from, any
 combination of gbif, inat and/or obis.}
 
-\item{show_map}{a `boolean`. If TRUE a Leaflet map with occurences
+\item{show_map}{a `boolean`. If TRUE a Leaflet map with occurrences
 is shown. Default FALSE.}
 
 \item{limit}{a `numeric`. Number of records to return. This is passed
@@ -25,8 +25,8 @@ something smallish so that you can get a result quickly, then do more as
 needed.}
 }
 \value{
-The output of the function is a tibble with the list of occurrences
-retived from species data sources selected into parameter list_DS.
+The output of the function is a `list` of `sf` one for each of the
+repositories specified in the list_DS parameter.
 }
 \description{
 This function acquires data from
@@ -38,27 +38,28 @@ boundary, which is obtained from the DEIMS-SDR sites API.
 \examples{
 \dontrun{
 # terrestrial site Saldur River Catchment
-occ_SRC <- get_site_speciesOcc(
+occ_SRC <- get_site_speciesOccurrences(
   deimsid =
   "https://deims.org/97ff6180-e5d1-45f2-a559-8a7872eb26b1",
   list_DS = c("gbif", "inat"),
-  show_map = TRUE,
+  show_map = FALSE,
   limit = 100
 )
 occ_SRC
 
 # marine site Gulf of Venice only obis
-occ_GoV <- get_site_speciesOcc(
+occ_GoV <- get_site_speciesOccurrences(
   deimsid =
   "https://deims.org/758087d7-231f-4f07-bd7e-6922e0c283fd",
-  list_DS = c("obis"),
+  list_DS = "obis",
   show_map = TRUE,
   limit = 100
 )
 occ_GoV
 
-# marine site Gulf of Venice only gbif, inat and obis
-occ_GoV_all <- get_site_speciesOcc(
+# marine site Gulf of Venice, all repositories are invoked
+# gbif, inat and obis
+occ_GoV_all <- get_site_speciesOccurrences(
   deimsid =
   "https://deims.org/758087d7-231f-4f07-bd7e-6922e0c283fd",
   list_DS = c("gbif", "inat", "obis"),
