+++ _pkgdown.yaml
@@ -12,7 +12,7 @@ reference:
       - '`get_site_info`'
       - '`get_site_ODS`'
       - '`get_site_MODIS`'
-      - '`get_site_speciesOcc`'
+      - '`get_site_speciesOccurrences`'
   - subtitle: Networks
     desc: For retrive information about LTER networks stored in DEIMS-SDR website.
     contents:
