+++ NAMESPACE
@@ -15,7 +15,7 @@ export(get_network_sites)
 export(get_site_MODIS)
 export(get_site_ODS)
 export(get_site_info)
-export(get_site_speciesOcc)
+export(get_site_speciesOccurrences)
 export(get_sos_procedurelist)
 export(produce_network_points_map)
 export(produce_site_map)
