### 2022-02-01 0179ff2
added test_mode variable for codecov
