### 2022-02-01 9623aa5
Try removing source installation of terra
