+++ .github/workflows/test-coverage.yml
@@ -36,14 +36,6 @@ jobs:
           remotes::install_cran("covr")
         shell: Rscript {0}
 
-      - name: Reinstall {raster}
-        # temporary, see https://stackoverflow.com/questions/70674136/r-package-warning-multiple-methods-tables-found-for-direction
-        # remove when no longer needed
-        run: |
-          remotes::install_cran("terra", type = "source", repos = "https://stat.ethz.ch/CRAN/", force = TRUE)
-          remotes::install_cran("raster", type = "source", repos = "https://stat.ethz.ch/CRAN/", force = TRUE)
-        shell: Rscript {0}
-
       - name: Test coverage
         run: covr::codecov(quiet = FALSE)
         shell: Rscript {0}
