+++ .github/workflows/check-standard.yaml
@@ -88,13 +88,4 @@ jobs:
             libcairo2-dev \
             libgit2-dev
 
-      - name: Reinstall {raster}
-        # temporary, see https://stackoverflow.com/questions/70674136/r-package-warning-multiple-methods-tables-found-for-direction
-        # remove when no longer needed
-        if: runner.os == 'macOS'
-        run: |
-          remotes::install_cran("terra", type = "source", repos = "https://stat.ethz.ch/CRAN/", force = TRUE)
-          remotes::install_cran("raster", type = "source", repos = "https://stat.ethz.ch/CRAN/", force = TRUE)
-        shell: Rscript {0}
-
       - uses: r-lib/actions/check-r-package@v1
