+++ tests/testthat/test-get_site_speciesOccurrences.R
@@ -20,7 +20,7 @@ test_that("Expect error if internet connection is down", {
 
 skip_if_offline(host = "deims.org")
 
-test_that("Output of the function is 'list' as expected", {
+test_that("Output of the function is 'list' of 'sf' as expected", {
   result <- ReLTER::get_site_speciesOccurrences(
     deimsid = TESTURLSite,
     list_DS = c("gbif", "inat"),
@@ -28,8 +28,10 @@ test_that("Output of the function is 'list' as expected", {
     limit = 100
   )
   expect_type(result, "list")
+  expect_s3_class(result$gbif, "sf")
+  expect_s3_class(result$inat, "sf")
   expect_true(ncol(result$gbif) == 81)
-  expect_true(ncol(result$inat) == 152)
+  expect_true(ncol(result$inat) == 151)
   expect_true(nrow(result$gbif) == 100)
   expect_true(nrow(result$inat) == 100)
 })
