### 2022-02-01 b7e2b8c
Avoid warning
save_ts_dir should never exist in this case, but running tests for a time > 1 this could occur. this edit avoid showing warnings in this case.
