### 2022-02-01 6d28ed3
Run tests on all branches
