+++ .github/workflows/check-standard.yaml
@@ -1,8 +1,7 @@
 # Workflow derived from https://github.com/r-lib/actions/tree/master/examples
 # Need help debugging build failures? Start at https://github.com/r-lib/actions#where-to-find-help
 on:
-  push:
-    branches: [main,dev]
+  push
   pull_request:
     branches: [main]
 
