+++ .github/workflows/test-coverage.yml
@@ -1,6 +1,5 @@
 on:
-  push:
-    branches: [main,dev]
+  push
 
 name: test-coverage
 
