+++ .github/workflows/test-coverage.yml
@@ -1,5 +1,5 @@
 on:
-  push
+  push:
 
 name: test-coverage
 
