### 2022-02-02 95b8941
Fix Actions errors (#3)
* Try removing all skip_on_ci()

* Try fixing terra error, again

* Try forcing installing deps from binary

* typo

* Change pkgType only on Mac

* typo2

* typo3
