+++ tests/testthat/test-get_network_envcharacts.R
@@ -18,7 +18,6 @@ test_that("Expect error if internet connection is down", {
 })
 
 skip_if_offline(host = "deims.org")
-skip_on_ci()
 skip_if(skip_in_test_mode)
 
 test_that("Output of network environmental characteristics function constructs
