+++ .github/workflows/check-standard.yaml
@@ -63,6 +63,7 @@ jobs:
           
       - name: Install dependencies
         run: |
+          if (Sys.info()["sysname"]=="Darwin") {options("pkgType" = "binary")}
           remotes::install_deps(dependencies = TRUE)
           remotes::install_cran("rcmdcheck")
         shell: Rscript {0}
