### 2022-02-02 ffc42a6
Skip test if OpenDataScience Europe is offline
