+++ tests/testthat/test-get_site_ODS.R
@@ -1,6 +1,7 @@
 message("\n---- Test get_site_ODS() ----")
 
 library(testthat)
+skip_if_offline(host = "s3.eu-central-1.wasabisys.com")
 
 test_that("Expect error if internet connection is down", {
   Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
