+++ NAMESPACE
@@ -24,7 +24,10 @@ export(produce_site_parameters_waffle)
 export(produce_site_qrcode)
 export(taxon_id_pesi)
 export(taxon_id_worms)
-import(MODIStsp)
+importFrom(MODIStsp,MODIStsp)
+importFrom(MODIStsp,MODIStsp_get_prodlayers)
+importFrom(MODIStsp,MODIStsp_get_prodnames)
+importFrom(MODIStsp,load_prodopts)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(dplyr,as_tibble)
 importFrom(dplyr,bind_rows)
@@ -51,6 +54,9 @@ importFrom(ggplot2,scale_y_continuous)
 importFrom(ggplot2,theme)
 importFrom(ggplot2,theme_minimal)
 importFrom(grDevices,colorRampPalette)
+importFrom(grDevices,dev.off)
+importFrom(grDevices,png)
+importFrom(graphics,par)
 importFrom(grid,viewport)
 importFrom(httr,RETRY)
 importFrom(httr,content)
@@ -87,6 +93,7 @@ importFrom(sf,write_sf)
 importFrom(spocc,obis_search)
 importFrom(spocc,occ)
 importFrom(spocc,occ2df)
+importFrom(stats,complete.cases)
 importFrom(taxize,eubon_search)
 importFrom(terra,crop)
 importFrom(terra,crs)
@@ -109,6 +116,8 @@ importFrom(tmap,tm_scale_bar)
 importFrom(tmap,tm_shape)
 importFrom(utils,capture.output)
 importFrom(utils,read.csv)
+importFrom(utils,tail)
+importFrom(utils,write.csv)
 importFrom(waffle,waffle)
 importFrom(worrms,wm_records_names)
 importFrom(xml2,read_xml)
