+++ NAMESPACE
@@ -27,7 +27,6 @@ export(taxon_id_worms)
 importFrom(MODIStsp,MODIStsp)
 importFrom(MODIStsp,MODIStsp_get_prodlayers)
 importFrom(MODIStsp,MODIStsp_get_prodnames)
-importFrom(MODIStsp,load_prodopts)
 importFrom(RColorBrewer,brewer.pal)
 importFrom(dplyr,as_tibble)
 importFrom(dplyr,bind_rows)
