### 2022-02-02 4a0f039
Use USGS credentials defined in yaml + secrets
Saving credentials in a publicly accessible file is not suer. Moreover, using the secrets approach (multiple credentials) is a tentative to bypass errors.
