+++ tests/testthat/test-get_site_MODIS.R
@@ -66,8 +66,11 @@ test_that("Check for invalid dates", {
 
 # Test for successful download, function returns terra::SpatRaster
 test_that("After successfull download, Function returns SpatRaster", {
-  # Load username and password from sysdata.rda
-  creds <- ReLTER:::creds
+  # Load username and password from secrets
+  creds <- list(
+    "user" = Sys.getenv("USGS_USER"),
+    "password" = Sys.getenv("USGS_PASSWORD")
+  )
   ds <- get_site_MODIS(
           deimsid = "https://deims.org/45722713-80e3-4387-a47b-82c97a6ef62b",
           earthdata_user = creds$user,
