+++ .github/workflows/check-standard.yaml
@@ -17,14 +17,16 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: macOS-latest,   r: 'release', test_mode: 'FALSE'}
-          - {os: windows-latest, r: 'release', test_mode: 'FALSE'}
-          - {os: ubuntu-20.04,   r: 'devel',   test_mode: 'FALSE', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
-          - {os: ubuntu-20.04,   r: 'release', test_mode: 'FALSE', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
-          - {os: ubuntu-20.04,   r: 'oldrel',  test_mode: 'FALSE', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
+          - {os: macOS-latest,   r: 'release', usgs_user: 'modistsp_ci_macos_release',   test_mode: 'FALSE'}
+          - {os: windows-latest, r: 'release', usgs_user: 'modistsp_ci_windows_release', test_mode: 'FALSE'}
+          - {os: ubuntu-20.04,   r: 'devel',   usgs_user: 'modistsp_ci_ubuntu_release',  test_mode: 'FALSE', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
+          - {os: ubuntu-20.04,   r: 'release', usgs_user: 'modistsp_ci_ubuntu_devel',    test_mode: 'FALSE', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
+          - {os: ubuntu-20.04,   r: 'oldrel',  usgs_user: 'modistsp_ci_ubuntu_oldrel',   test_mode: 'FALSE', rspm: "https://packagemanager.rstudio.com/cran/__linux__/focal/latest"}
 
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      USGS_USER: ${{ matrix.config.usgs_user }}
+      USGS_PASSWORD: ${{ secrets.USGS_PASSWORD }}
       R_KEEP_PKG_SOURCE: yes
       LOCAL_DEIMS: ${{ matrix.config.test_mode }}
       
