### 2022-02-02 bb88eed
Skip test in case of MODIStsp's internal errors
