### 2022-02-23 #485 @allisonhorst
Here is my review of the `ReLTER` package! Happy to answer any questions and provide clarification.

## Package Review

None.

-   ☒ As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor *before* starting your review).

#### Documentation

-   ☐ **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
-   ☒ **Installation instructions:** for the development version of package and any non-standard dependencies in README
-   ☐ **Vignette(s):** demonstrating major functionality that runs successfully locally
-   ☒ **Function Documentation:** for all exported functions
-   ☒ **Examples:** (that run successfully locally) for all exported functions
-   ☒ **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).

#### Functionality

-   ☒ **Installation:** Installation succeeds as documented.
-   ☒ **Functionality:** Any functional claims of the software been confirmed.
-   ☒ **Performance:** Any performance claims of the software been confirmed.
-   ☐ **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
-   ☒ **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.

Estimated hours spent reviewing: 6

-   ☒ Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer (“rev” role) in the package DESCRIPTION file.

### Review Comments

**Overall comments:**

-   This seems like a valuable package to access and work with data for LTER users! It combines a *lot* of options for accessing information about networks, sites, and data that I imagine is very useful for LTER researchers working in R
- A clearer statement of need & a bit more background about LTER / eLTER on the homepage & README could provide a nicer entryway for beginner users, or those who happen up on it. It dives into the details a bit quickly, and a broader intro & motivation would be helpful. It would also be great to have a section on other options for how users can interact with the LTER data.
-   I really like the welcoming CONTRIBUTING documentation :heart:
-   I think the package would really benefit from a few vignettes, beyond what is in the README, with examples including explanatory text & annotation describing what’s going on, how to find the DEIMS.iD being used, etc. that walks adopters through how and why they might use functions in the package. Also, weirdly, when I install from GitHub `browseVignettes()` returns “No vignettes found” – but following local install the one vignette does show up. Could be weirdness on my end.
-   It took me a while to figure out where to find the DEIMS.iD for sites, networks, activities, data, etc. (and to understand that these are different - e.g. from what I tried a DEIMS.iD for a network will fail on a function built for a site). If it is possible on the homepage to clearly link (e.g. in the README) to sites that list all networks, sites, activities, etc. that could be really helpful for newcomers to the package.
-   For a number of functions more information in the documentation would be helpful for new users. I’ve noted suggestions for specific functions below.

# Checklist from `ropensci-org/pkgreviewr` package template

### Session info

-   System (from `sessionInfo()`):
    -   R version 4.1.2 (2021-11-01)
    -   Platform: x86\_64-apple-darwin17.0 (64-bit)
    -   Running under: macOS Big Sur 11.3.1

## Test installation

### Test local `ReLTER` install:

    install(pkg_dir, dependencies = T, build_vignettes = T)

No issues.

### Test install of `ReLTER` from GitHub with:

    devtools::install_github("oggioniale/ReLTER", dependencies = T, build_vignettes = T)

No issues.

## Check package integrity

### Run checks on `ReLTER` source:

`devtools::check()` returns `0 errors, 0 warnings, 0 notes`

### Run tests on `ReLTER` source:

`devtools::test()`: I’m having real issues with this, with of warnings and errors thrown, terminated early, etc. (see below of the summary). I’d love to try this again & learn if I’m doing something wrong on my end, this doesn’t seem to align with the testing checks done by the editor or authors?

    Maximum number of failures exceeded; quitting early.
    You can increase this number by setting `options(testthat.progress.max_fails)`

    ══ Results ═══════════════════════════════════════════════════════════════════════════════════════════════
    Duration: 1138.7 s

    [ FAIL 10 | WARN 8 | SKIP 0 | PASS 244 ]
    ══ Terminated early ═

### Check `ReLTER` for goodpractice:

`goodpractice::gp()` returns: “It is good practice to

✖ fix this R CMD check ERROR: Package suggested but not available: ‘ISOcodes’ The suggested packages are required for a complete check. Checking can be attempted without them by setting the environment variable *R\_CHECK\_FORCE\_SUGGESTS* to a false value. See section ‘The DESCRIPTION file’ in the ‘Writing R Extensions’ manual.”

## Check package metadata files

### Inspect

-   #### [README](https://github.com/oggioniale/ReLTER)

-   #### [DESCRIPTION](https://github.com/oggioniale/ReLTER/blob/master/DESCRIPTION)

-   #### [NAMESPACE](https://github.com/oggioniale/ReLTER/blob/master/NAMESPACE)

### Spell check

    devtools::spell_check(pkg_dir)

**Check spelling:**

-   “retrive” in README.md
-   “througth” in get\_network\_related\_resources.Rd:23 and get\_network\_research\_topics.Rd:23
-   “whitin” in NEWS.md

**README comments:**

-   I think the package would benefit from a clearer statement of need, some background information on eLTER, and expected audience (and possibly move this information up: “Functions currently implemented are derived from the discussion of the needs declared by eLTER users community” and add a bit more to motivate the package)?
-   There are a lot of abbreviations right at the beginning that are spelled out later on. It may be useful to have them written out the first time they show up, e.g. “Long Term Ecological Research (LTER)” then use the abbreviation thereafter
-   Check grammar throughout (I am happy to suggest edits directly if helpful for the authors)
-   I appreciate the use of emojis in the README & CONTRIBUTING docs - it makes the package documentation feel so welcoming!

**DESCRIPTION comments:**

-   Suggest updating Maintainer text: “The package maintainer Alessandro Oggioni, phD (2020)” to “Maintained by Alessandro Oggioni, PhD (2020)” or just “Alessandro Oggioni”
-   The description seems a bit long compared to other packages, possibly shorten to highlight what the package offers “bigger picture” (but honestly I’m not sure if it matters at all…)?

**NAMESPACE comments:**

None.

## Check documentation

-   Is the [documentation](https://ropensci.github.io/dev_guide/building.html#documentation) (installation instructions/vignettes/examples/demos) clear and sufficient?

### test `ReLTER` function help files:

`help(package = "ReLTER")` returns documentation as expected.

### test `ReLTER` vignettes:

`vignette(package = "ReLTER")` returns single vignette with information from the README. Annotated / explanatory vignettes would be really helpful.

## Test functionality:

-   Are there **user interface improvements** that could be made?

The [reference](https://oggioniale.github.io/ReLTER/reference/index.html) page descriptions of the function are the function name followed by “function,” but would be more helpful if they contained a brief description of what the function actually does instead (e.g. <https://docs.ropensci.org/GSODR/reference/index.html>)

-   Are there **performance improvements** that could be made?

I’ve noted specific recommendations for function use & documentation after the specific functions below.

## Feedback on individual package functions

Comment throughout on function documentation: in many functions’ documentation there is a line “More information about DEIMS network ID in these pages: page, and page the complete list of ILTER networks.” I think the link text needs to be updated (e.g. to “DEIMS.iD information and Complete list of ILTER networks” instead of “page and page”).

#### `get_site_info()`

-   404 error on [website page](https://oggioniale.github.io/ReLTER/reference/get_site_info.html) for this function
-   Example & test on a new site DEIMS.iD with varying categories seems to work fine

#### `get_site_ODS()`

-   I was like “no way this just works” and then it just worked beautifully
-   Unfortunately most of the sites I tried returned “No boundary for requested DEIMS site,” but I appreciated the clear message letting me know why it didn’t work

#### `get_network_envcharacts()`

-   Checked documentation and example: no issues
-   It’s really helpful that the list of networks is linked in the function documentation! I noticed that the DEIMS.iDs aren’t shown (as on other pages) below the page title. Could be worth noting that for the network sites, the DEIMS.iD is the URL for the network page? I kept looking for the DEIMS.iD right below the page title for everything.

#### `get_network_parameters()`

-   Checked documentation and example: no issues
-   Tested for 2 other networks with no problems

#### `get_network_research_topics()`

-   Checked documentation and example: no issues

#### `get_network_related_resources()`

-   Checked documentation and example: no issues

#### `get_network_sites()`

-   Checked documentation and example: no issues
-   Some of these examples feel really victorious, it’s great.

#### `get_ilter_envcharacts()`

-   Checked documentation and example: no issues
-   I am a bit confused about when this function would be useful based on the description - is it just taking a slice from a database of iLTER sites (e.g. the first *n* rows)? If yes, possible for a user to easily choose “all”?
-   Does it make sense for the default `sitesNum = 0` rather than the default of getting environmental characteristics for all sites?
-   I feel that the description of the `sitesNum` argument should be clarified and expanded on a bit - where are the rows getting pulled from? Does that order change or is it static?

#### `get_ilter_generalinfo()`

-   Checked documentation and two examples: no issues
-   Nice that the country\_name doesn’t require an exact match

#### `get_ilter_parameters()`

-   Checked documentation and example: no issues
-   See comment above re: why a user would want to just get information on the first n sites by position (rather than more meaningful queries)? Could just use a bit of explanation to clarify expected use.

#### `get_ilter_research_topics()`

-   Checked documentation and example: no issues
-   I like that the default here is `sitesNum = NULL` (and note the inconsistency with previous functions where `sitesNum = 0`) – unsure if that difference is intentional.

#### `get_activity_info()`

-   Checked documentation and example: no issues
-   May want to update the description (“activity (actually title)” phrasing is unclear)
-   Once again, I exclaimed joyfully upon successful map production :)

#### `get_dataset_info()`

-   Checked documentation and example: no issues

#### `produce_network_points_map()`

-   Checked documentation and example: no issues
-   The documentation says that the output is the tmap plot, but this also seems to create a nice sf object (`tbl_df`) with site names, coordinates, etc. - could be worth adding that the output is stored & usable for other stuff beyond the tmap?
-   One question I had: is this map customizable / exportable? It wasn’t clear that a graphical object was stored anywhere so that I could build on it, which could be nice.

#### `produce_site_map()`

-   Checked documentation and Lange Bramke example: no issues

#### `produce_site_qrcode()`

-   Checked documentation and example: no issues
-   Also, COOL

#### `produce_site_parameters_pie()`

-   Checked documentation and example: no issues
-   It isn’t obvious what the percentages refer to, more detail would be useful for users to responsibly understand what is being plotted and how it is useful
-   See above re: the output (a tibble in addition to the pie chart, which extends this function’s usefulness)

#### `produce_site_parameters_waffle()`

-   Checked documentation and example: no issues
-   Similar to above, more clarification in the documentation of what is being plotted here may be helpful. For example, does one square indicate that there is one agricultural parameter that has ever been recorded at that site?
-   See above re: output (a tibble here allows greater use / customization and flexibility)

#### `get_sos_procedurelist()`

-   Checked documentation and example: no issues
-   A bit more description might be helpful for users unfamiliar with SOS to clarify when & why someone might use this function
-   It looks like the name and URI are the same in the output. If possible to make the name more user-friendly / readable, that might be helpful here

#### `taxon_id_pesi()`

-   Checked documentation and example: Error returned in example after creating table when running `taxon_id_pesi(table = table, taxaColumn = 3)`. Returns: `Error in eubon_error(res) : Called from: eubon_error(res)`
-   The information re: CSV encoding potentially belongs in the function documentation (rather than in the examples section?)

#### `taxon_id_worms()`

-   Checked documentation and example: no issues
-   After entering a number to specify the record “most similar to the taxa name” (note: it currently says “mostsimilar”), it isn’t clear that anything happens (but the table does exist as a stored object). Possible to have a message returned after choosing that says “Object created” or something similar, just so the user knows it worked more obviously?

## Inspect code:

#### **Comments on package and code:**

-   Package function and argument naming conventions are intuitive and consistent. I note that the authors follow a verb-object naming scheme (instead of object-verb as recommended by [ROpenSci package guidelines](https://devguide.ropensci.org/building.html#function-and-argument-naming)).
-   Code is readable and organized throughout, and appears to follow the tidyverse style guide
-   Overall, I think the package mostly adheres to ROpenSci Packaging Guide. Minor note: Section 1.16 of the [ROpenSci Packaging Guide](https://devguide.ropensci.org/building.html) recommends “Avoid starting the description with the package name or”This package …”
-   I don’t note any code duplication (but also don’t have a lot of experience looking for it, so might be missing something)
-   Re: dependencies, installing this package seems like it may require a lot of updates for users (given that all dependencies except `terra` specify a minimum version). This may be necessary, else consider from the ROpenSci Packaging Guide (section 1.13): “If you know for a fact that your package will break below a certain dependency version, specify it explicitly. But if you don’t, then no need to specify a minimum dependency.” I’ll add a massive caveat here that I know basically nothing about handling dependencies, so you might just ignore this comment :)

## Review test suite:

### Test coverage

`covr::package_coverage(pkg_dir)` returns:

    ── Error (test-taxon_id_pesi.R:21:3): Output of taxon pesi function constructs 'tibble' as expected ──
    Error:
    Backtrace:
        ▆
     1. └─ReLTER::taxon_id_pesi(table = table, taxaColumn = 3) at test-taxon_id_pesi.R:21:2
     2.   └─taxize::eubon_search(query = table[[taxaColumn]][i], providers = "pesi")
     3.     └─taxize:::eubon_error(res)

    [ FAIL 1 | WARN 2 | SKIP 9 | PASS 234 ]
    Error: Test failures
    Execution halted

### Inspect [tests](https://github.com/oggioniale/ReLTER/blob/master/tests/testthat)

-   Testing (w/ `testthat`) appears thorough and reasonable, and tests exist for each function in the package
-   9 (of 21) test scripts include `skip_on_cran()` (not wholesale skipping)

### End review.

I look forward to answering any questions you have about this feedback, and am happy to re-test/review as you work on revisions! -Allison
