### 2022-02-24 #485 @oggioniale
@allisonhorst
Many many thanks for your review. We working on the answers and on fixing the issues that you are open.
In the next few days our answers and the commits for improve the quality of `ReLTER` package following your suggestions.
