### 2022-02-25 #485 @oggioniale
> Here is my review of the `ReLTER` package! Happy to answer any questions and provide clarification. Some of the testing issues may be on my end but I included anyway in case useful to see.

Dear @allisonhorst,
thanks a lot for your comments, moreover, for your revisions, suggestions and issues.
In line my answers.

>

----------------------------

> ### Review Comments
> **Overall comments:**
>
> * This seems like a valuable package to access and work with data for LTER users! It combines a _lot_ of options for accessing information about networks, sites, and data that I imagine is very useful for LTER researchers working in R

This is exactly the main aim of `ReLTER` package: interact with the informations and data provided by eLTER tools. In addition, since the eLTER community is global and provides long-term ecological research data, in perspective will be useful for ecology researchers working in R.

> * A clearer statement of need & a bit more background about LTER / eLTER on the homepage & README could provide a nicer entryway for beginner users, or those who happen up on it. It dives into the details a bit quickly, and a broader intro & motivation would be helpful. It would also be great to have a section on other options for how users can interact with the LTER data.

We will do. As part of the eLTER community we take it for seems obvious that community practices are known, it is good to have an outside view.
We should also describethe difference between LTER and eLTER.

> * I really like the welcoming CONTRIBUTING documentation ❤️

❤️

> * I think the package would really benefit from a few vignettes, beyond what is in the README, with examples including explanatory text & annotation describing what’s going on, how to find the DEIMS.iD being used, etc. that walks adopters through how and why they might use functions in the package. Also, weirdly, when I install from GitHub `browseVignettes()` returns “No vignettes found” – but following local install the one vignette does show up. Could be weirdness on my end.

In the few weeks, I started with the creation of "How to" pages describing some of the practical needs of researchers in the eLTER community (e.g. How to get the info of most than one LTER Sites?). Probably these are not true vignettes but can be a starting point.

> * It took me a while to figure out where to find the DEIMS.iD for sites, networks, activities, data, etc. (and to understand that these are different - e.g. from what I tried a DEIMS.iD for a network will fail on a function built for a site). If it is possible on the homepage to clearly link (e.g. in the README) to sites that list all networks, sites, activities, etc. that could be really helpful for newcomers to the package.

Another thing we took for seems obvious.

> * For a number of functions more information in the documentation would be helpful for new users. I’ve noted suggestions for specific functions below.

Thank you very much, we will consider all your suggestions to improve the use of the package.

-------------------------

> ## Inspect code:
> #### **Comments on package and code:**
> * Package function and argument naming conventions are intuitive and consistent. I note that the authors follow a verb-object naming scheme (instead of object-verb as recommended by [ROpenSci package guidelines](https://devguide.ropensci.org/building.html#function-and-argument-naming)).
> * Code is readable and organized throughout, and appears to follow the tidyverse style guide
> * Overall, I think the package mostly adheres to ROpenSci Packaging Guide. Minor note: Section 1.16 of the [ROpenSci Packaging Guide](https://devguide.ropensci.org/building.html) recommends “Avoid starting the description with the package name or”This package …”
> * I don’t note any code duplication (but also don’t have a lot of experience looking for it, so might be missing something)
> * Re: dependencies, installing this package seems like it may require a lot of updates for users (given that all dependencies except `terra` specify a minimum version). This may be necessary, else consider from the ROpenSci Packaging Guide (section 1.13): “If you know for a fact that your package will break below a certain dependency version, specify it explicitly. But if you don’t, then no need to specify a minimum dependency.” I’ll add a massive caveat here that I know basically nothing about handling dependencies, so you might just ignore this comment :)

-------------------------

> ### End review.
> I look forward to answering any questions you have about this feedback, and am happy to re-test/review as you work on revisions! -Allison

We hope not to give you too much work yet, we will strive to improve the package in the way you suggested.
Alessandro
