### 2022-02-25 6ac8e9a
descrizione di nuove funzioni in DESCRIPTION
