+++ vignettes/ReLTER.Rmd
@@ -1,50 +1,87 @@
 ---
-title: "ReLTER"
-output: rmarkdown::html_vignette
+title: 'ReLTER: An interface to the eLTER for the R statistical programming environment - v1.0.0'
+author: "Alessadro Oggioni ([alessandro.oggioni@cnr.it](mailto:alessandro.oggioni@cnr.it)),
+  Micha Silver ([silverm@post.bgu.ac.il](mailto:silverm@post.bgu.ac.il)),
+  Luigi Ranghetti ([ranghetti.l@irea.cnr.it](mailto:ranghetti.l@irea.cnr.it)),
+  Paolo Tagliolato ([tagliolato.a@irea.cnr.it](mailto:tagliolato.a@irea.cnr.it))"
+# bibliography: ReLTER.bib
+date: '`r Sys.Date()`'
+output:
+  rmarkdown::html_vignette:
+    fig_caption: yes
+    number_section: yes
+    toc: no
+    toc_depth: 2
+  rmarkdown::pdf_document:
+    fig_caption: yes
+    number_sections: yes
+    toc: yes
+    toc_depth: 2
+  rmarkdown::html_document:
+    fig_caption: yes
+    number_section: yes
+    toc: no
+    toc_depth: 1
+urlcolor: blue
+linkcolor: blue
 vignette: >
-  %\VignetteIndexEntry{ReLTER}
+  \usepackage[utf8]{inputenc}
+  %\VignetteIndexEntry{ReLTER: An interface to the eLTER for the R statistical programming environment - v1.0.0}
   %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
 ---
 
 ```{r, include = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  fig.path = "man/figures/README-",
+  fig.path = "vignettes/figures/vignettes-",
   fig.ext = "png",
   out.width = "100%"
 )
 ```
+<img src='../man/figures/logo.png' align="right" height="139" />
 
-## Description of <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span>
-<span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span> is an R package that: provides access to [DEIMS-SDR](https://deims.org/)), allows interact with software implemented by eLTER Research Infrastructure (RI) and improves the data/information shared by them.
-<span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span> is a R package devoted to access, interact and improve the information and the data shared by Long Term Ecological Research (LTER) network. This package is born within eLTER H2020 major project that will help advance the development of European Long-Term Ecosystem Research Infrastructures ([eLTER RI](https://elter-ri.eu)).
+<span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span> provides access to [DEIMS-SDR](https://deims.org/), and allows interaction with data and software implemented by [Long Term Ecological Research Research Infrastructure (eLTER-RI)](https://elter-ri.eu) thus improving data sharing among [European LTER projects](https://elter-ri.eu/projects).
 
-The <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span> package functions in particular allow to:
+<br>
+<br>
+<br>
+<br>
+<br>
 
-- retrive the information about entities (e.g. sites, datasets, and activities) shared by [DEIMS-SDR](https://deims.org/) (see e.g. [get_site_info function](https://oggioniale.github.io/ReLTER/reference/get_site_info.html));
+# Documentation of <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span>
 
[…]
