+++ _pkgdown.yaml
@@ -7,12 +7,12 @@ template:
 reference:
   - title: DEIMS-SDR interaction functions
   - subtitle: Sites
-    desc: For retrive information about LTER sites stored in DEIMS-SDR website and for acquires data from third-party services.
+    desc: For retrieve information about LTER sites stored in DEIMS-SDR website and for acquires data from third-party services.
     contents:
       - '`get_site_info`'
       - '`get_site_ODS`'
   - subtitle: Networks
-    desc: For retrive information about LTER networks stored in DEIMS-SDR website.
+    desc: For retrieve information about LTER networks stored in DEIMS-SDR website.
     contents:
       - '`get_network_envcharacts`'
       - '`get_network_parameters`'
@@ -20,14 +20,14 @@ reference:
       - '`get_network_related_resources`'
       - '`get_network_sites`'
   - subtitle: All ILTER
-    desc: For retrive information about all ILTER network stored in DEIMS-SDR website.
+    desc: For retrieve information about all ILTER network stored in DEIMS-SDR website.
     contents:
       - '`get_ilter_envcharacts`'
       - '`get_ilter_generalinfo`'
       - '`get_ilter_parameters`'
       - '`get_ilter_research_topics`'
   - subtitle: Other entities
-    desc: For retrive information about other entities (datasets and activites) stored in DEIMS-SDR website.
+    desc: For retrieve information about other entities (datasets and activites) stored in DEIMS-SDR website.
     contents:
       - '`get_activity_info`'
       - '`get_dataset_info`'
