### 2022-02-28 a664220
deleted tests for new functions
