### 2022-03-03 8e0ae65
Few edits following goodpractice::gp()
