+++ R/get_dataset_info.R
@@ -1,4 +1,6 @@
-#' @title eLTER get_dataset_info function
+#' @title Return a tibble object containing information, as a stored in
+#' \href{https://deims.org/}{DEIMS-SDR catalogue}, about a single eLTER dataset
+#' (e.g. \url{https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef}).
 #' @description This function allows to obtain the info of dataset provided in
 #' DEIMS-SDR.
 #' @param datasetid A character. It is the DEIMS iD of dataset make from
