+++ R/get_site_ODS.R
@@ -1,4 +1,6 @@
-#' @title eLTER get_site_ODS function
+#' @title Download and return a SpatRaster object containing requested
+#' by \href{https://maps.opendatascience.eu/}{ODS} dataset, cropped to a
+#' eLTER site boundaries.
 #' @description This function acquires various raster layers from
 #' ODS Europe:  https://maps.opendatascience.eu/
 #' and crops to an eLTER site boundary, which is obtained
@@ -129,7 +131,7 @@ get_site_ODS <- function(deimsid, dataset = "landcover") {
   boundary <- terra::vect(boundary)
   ds_site <- terra::mask(terra::crop(ds, boundary), boundary)
   # If this is NDVI, rescale back to (-1.0,1.0) range
-  if (length(grep(pattern="ndvi", x=dataset, fixed = TRUE)) > 0) {
+  if (length(grep(pattern = "ndvi", x = dataset, fixed = TRUE)) > 0) {
     ds_site <- (ds_site - 100) / 100.0
   }
   return(ds_site)
