+++ R/get_network_parameters.R
@@ -1,5 +1,8 @@
-#' @title eLTER get_network_parameters function
-#' @description This function obtains a list of the environmental parameters
+#' @title Return a tibble object containing Environmental Parameters collected
+#' by all of the eLTER sites belonging to an eLTER Network (e.g.
+#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
+#' Italy network}), as a stored into \href{https://deims.org}{DEIMS-SDR}.
+#' @description This function obtains a list of the Environmental Parameters
 #' collected on sites in the eLTER Network (e.g. Italy) through the DEIMS-SDR
 #' sites API.
 #' @param networkDEIMSID A `character`. The DEIMS ID of network
