+++ _pkgdown.yaml
@@ -1,76 +1,0 @@
-title: ReLTER
-url: https://oggioniale.github.io/ReLTER/
-template:
-  params:
-    bootswatch: flatly
-    ganalytics: UA-12345678-0
-reference:
-  - title: DEIMS-SDR interaction functions
-  - subtitle: Sites
-    desc: For retrieve information about LTER sites stored in DEIMS-SDR website and for acquires data from third-party services.
-    contents:
-      - '`get_site_info`'
-      - '`get_site_ODS`'
-  - subtitle: Networks
-    desc: For retrieve information about LTER networks stored in DEIMS-SDR website.
-    contents:
-      - '`get_network_envcharacts`'
-      - '`get_network_parameters`'
-      - '`get_network_research_topics`'
-      - '`get_network_related_resources`'
-      - '`get_network_sites`'
-  - subtitle: All ILTER
-    desc: For retrieve information about all ILTER network stored in DEIMS-SDR website.
-    contents:
-      - '`get_ilter_envcharacts`'
-      - '`get_ilter_generalinfo`'
-      - '`get_ilter_parameters`'
-      - '`get_ilter_research_topics`'
-  - subtitle: Other entities
-    desc: For retrieve information about other entities (datasets and activites) stored in DEIMS-SDR website.
-    contents:
-      - '`get_activity_info`'
-      - '`get_dataset_info`'
-  - title: Graphical functions
-    desc: Functions able of producing geographical images of sites and national networks or charts.
-    contents:
-      - '`produce_network_points_map`'
-      - '`produce_site_map`'
-      - '`produce_site_qrcode`'
-      - '`produce_site_parameters_pie`'
-      - '`produce_site_parameters_waffle`'
-  - title: Other functions
-    desc: Functions able to some other interactions.
-    contents:
-      - '`get_sos_procedurelist`'
-      - '`taxon_id_pesi`'
-      - '`taxon_id_worms`'
-navbar:
-  type: inverse
-  left: []
-  right:
-    - icon: fa-home fa-lg
-      href: index.html
-    - text: "Reference"
-      icon: fa-file-code
-      href: reference/index.html
-    #- text: "Articles"
-    #  icon: fa-book
-    #  menu:
-      #-  text: "Installation"
-      #   href: articles/installation.html
-      #-  text: "Run in a Docker container"
-      #   href: articles/docker.html
-      #-  text: "Using sen2r() from the GUI"
-      #   href: articles/sen2r_gui.html
-      #-  text: "Using sen2r() from the command line"
-      #   href: articles/sen2r_cmd.html
-      #-  text: "Output file structure"
-      #   href: articles/outstructure.html
-    - text: "Changelog"
-      icon: fa-newspaper
-      href: news/index.html
-    - icon: fa-user
-      href: https://www.cnr.it/people/alessandro.oggioni
-    - icon: fa-github fa-lg
-      href: https://github.com/oggioniale/ReLTER/
