+++ R/get_activity_info.R
@@ -5,7 +5,7 @@
 #' @description This function obtains the info of activity (actually
 #'  title) provided in DEIMS-SDR.
 #' @param activityid A character. The DEIMS activity ID from
-#' DEIMS-SDR website. More information about DEIMS iD in this
+#' DEIMS-SDR website. More information about DEIMS.iD in this
 #' \href{https://deims.org/docs/deimsid.html}{page}.
 #' @return The output of the function is a `tibble` with main features of
 #' the activities in a site, and a `leaflet` map plot.
