+++ man/produce_site_parameters_pie.Rd
@@ -11,8 +11,8 @@ Italy network}).}
 produce_site_parameters_pie(deimsid)
 }
 \arguments{
-\item{deimsid}{A `character`. It is the DEIMS ID of site/network from
-DEIMS-SDR website. More information about DEIMS ID from:
+\item{deimsid}{A `character`. It is the DEIMS.iD of site/network from
+DEIMS-SDR website. More information about DEIMS.iD from:
 \href{https://deims.org/docs/deimsid.html}{page}.}
 }
 \value{
