+++ man/produce_site_parameters_waffle.Rd
@@ -11,8 +11,8 @@ Italy network}).}
 produce_site_parameters_waffle(deimsid)
 }
 \arguments{
-\item{deimsid}{A `character`. The DEIMS ID of site/network from:
-DEIMS-SDR website. More information about DEIMS ID from:
+\item{deimsid}{A `character`. The DEIMS.iD of site/network from:
+DEIMS-SDR website. More information about DEIMS.iD from:
 \href{https://deims.org/docs/deimsid.html}{page}.}
 }
 \value{
