+++ man/produce_site_qrcode.Rd
@@ -8,8 +8,8 @@ site, activity).}
 produce_site_qrcode(deimsid, do_plot = FALSE)
 }
 \arguments{
-\item{deimsid}{a `character`. The DEIMS ID of entities from
-DEIMS-SDR website. More information about DEIMS iD in this
+\item{deimsid}{a `character`. The DEIMS.iD of entities from
+DEIMS-SDR website. More information about DEIMS.iD in this
 \href{https://deims.org/docs/deimsid.html}{page}.}
 
 \item{do_plot}{a `boolean`. Plot the computed QRCode. Default FALSE.}
