+++ man/get_network_envcharacts.Rd
@@ -11,8 +11,8 @@ Italy network}).}
 get_network_envcharacts(networkDEIMSID)
 }
 \arguments{
-\item{networkDEIMSID}{A `character`. DEIMS ID of network
-from DEIMS-SDR website. More information about DEIMS ID refer to these pages:
+\item{networkDEIMSID}{A `character`. DEIMS.iD of network
+from DEIMS-SDR website. More information about DEIMS.iD refer to these pages:
 \href{https://deims.org/docs/deimsid.html}{page}, and
 \href{https://deims.org/search?f[0]=result_type:network}{page}
 the complete list of ILTER networks.}
