+++ R/get_network_parameters.R
@@ -6,10 +6,9 @@
 #' collected on sites in the eLTER Network (e.g. Italy) through the DEIMS-SDR
 #' sites API.
 #' @param networkDEIMSID A `character`. The DEIMS.iD of network
-#' from DEIMS-SDR website. For more information about DEIMS.iD refer to these
-#' pages: \href{https://deims.org/docs/deimsid.html}{page}, and
-#' \href{https://deims.org/search?f[0]=result_type:network}{page}
-#' the complete list of ILTER networks.
+#' from DEIMS-SDR website. DEIMS.iD information 
+#' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
+#' networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
 #' @return The output of the function is a `tibble` containing the list
 #' of parameters and their URI (Uniform Resource Identifier) collected
 #' by the network's sites.
