+++ R/get_site_infrastructure.R
@@ -2,8 +2,8 @@
 #' @description This function obtains infrastructure information
 #' of an eLTER site through the DEIMS-SDR sites API.
 #' @param deimsid A `character`. The DEIMS.iD of the site from
-#' DEIMS-SDR website. More information about DEIMS.iD from:
-#' \href{https://deims.org/docs/deimsid.html}{page}.
+#' DEIMS-SDR website. DEIMS.iD information 
+#' \href{https://deims.org/docs/deimsid.html}{here}.
 #' @return The output of the function is a `tibble` with main features of the
 #' site and infrastructure information where available, such as:
 #' power supply, accessibility, maintenaince interval, etc.
