+++ R/get_site_research_topics.R
@@ -2,8 +2,8 @@
 #' @description This function obtains a list of research topics handled
 #' at an eLTER site through the DEIMS-SDR sites API.
 #' @param deimsid  A `character`. The DEIMS.iD of the site from
-#' DEIMS-SDR website. More information about DEIMS.iD from:
-#' \href{https://deims.org/docs/deimsid.html}{page}.
+#' DEIMS-SDR website. DEIMS.iD information 
+#' \href{https://deims.org/docs/deimsid.html}{here}.
 #' @return The output of the function is a `tibble` with main features of the
 #' site and a list of the research topics handled in this site.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
