+++ R/get_network_sites.R
@@ -7,8 +7,9 @@
 #' including title, date late updated, URI, and coordinates
 #' @param networkDEIMSID A `character`. The DEIMS.iD of the network from
 #' DEIMS-SDR website. DEIMS.iD information 
-#' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
+#' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of
 #' networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
+#' The DEIMS.iD of network is the URL for the network page.
 #' @return The output of the function is a point vector of `sf` class
 #' (package sf) of the network's sites.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
