+++ man/get_network_parameters.Rd
@@ -13,7 +13,8 @@ get_network_parameters(networkDEIMSID)
 \item{networkDEIMSID}{A `character`. The DEIMS.iD of network
 from DEIMS-SDR website. DEIMS.iD information 
 \href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
-networks \href{https://deims.org/search?f[0]=result_type:network}{here}.}
+networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
+The DEIMS.iD of network is the URL for the network page.}
 }
 \value{
 The output of the function is a `tibble` containing the list
