+++ man/get_dataset_info.Rd
@@ -9,9 +9,10 @@
 get_dataset_info(datasetid)
 }
 \arguments{
-\item{datasetid}{A character. It is the DEIMS.iD of dataset make from
+\item{datasetid}{a character. It is the DEIMS.iD of dataset make from
 DEIMS-SDR website. DEIMS.iD information 
-\href{https://deims.org/docs/deimsid.html}{here}.}
+\href{https://deims.org/docs/deimsid.html}{here}.
+The DEIMS.iD of dataset is the URL for the dataset page.}
 }
 \value{
 The output of the function is a `tibble` with main features
