### 2022-03-08 ddf08cd
only tmap output for this function
