### 2022-03-08 e4eed37
changed documentation and code of two functions
