+++ R/produce_site_parameters_waffle.R
@@ -1,14 +1,14 @@
 #' @title Return a waffle chart of Environmental Parameters, as a stored in
-#' \href{https://deims.org/}{DEIMS-SDR catalogue}, of a single eLTER site or
-#' all of the eLTER sites belonging to an eLTER Network (e.g.
-#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
-#' Italy network}).
+#' \href{https://deims.org/}{DEIMS-SDR catalogue}, of a single eLTER site.
 #' @description This function produces a waffle chart of the Environmental
 #' Parameters collected in a site or network grouped into clusters.
 #' @param deimsid A `character`. The DEIMS.iD of site/network from:
 #' DEIMS-SDR website. DEIMS.iD information 
 #' \href{https://deims.org/docs/deimsid.html}{here}.
-#' @return The output of the function is a waffle chart.
+#' @return The output of the function is a waffle chart and a `tibble`. Each
+#' of the squares represents a parameters measured into the selected eLTER
+#' site. The parameters with the same color belong to the same group (e.g.
+#' biological, atmospheric, etc.).
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom tibble as_tibble
 #' @importFrom dplyr group_by tally mutate filter
@@ -21,7 +21,7 @@
 #' waffle <- produce_site_parameters_waffle(
 #'   deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
 #' )
-#' print(waffle)
+#' waffle
 #' }
 #'
 ### function produce_site_parameters_waffle
@@ -34,9 +34,12 @@ produce_site_parameters_waffle <- function(deimsid) {
   paramsDeims <- tibble::as_tibble(paramsDeims$parameter[[1]])
   if (length(paramsDeims) != 0) {
     params <- tibble::as_tibble(paramsDeims)
-    params$parameterGroups <- parametersStructureEnvThes$category[
-      match(params$parameterLabel, parametersStructureEnvThes$parameter)
-    ]
+    params$parameterGroups <- paste0(
+      parametersStructureEnvThes$category[
+        match(params$parameterLabel, parametersStructureEnvThes$parameter)
+      ],
+      "s"
+    )
     groupsIsNa <- params %>% dplyr::filter(is.na(parameterGroups))
     # plot of parameters ----
     params <- params %>%
