+++ R/get_dataset_info.R
@@ -3,7 +3,7 @@
 #' (e.g. \url{https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef}).
 #' @description This function allows to obtain the info of dataset provided in
 #' DEIMS-SDR.
-#' @param datasetid a character. It is the DEIMS.iD of dataset make from
+#' @param datasetid A `character`. It is the DEIMS.iD of dataset make from
 #' DEIMS-SDR website. DEIMS.iD information 
 #' \href{https://deims.org/docs/deimsid.html}{here}.
 #' The DEIMS.iD of dataset is the URL for the dataset page.
