+++ man/get_activity_info.Rd
@@ -10,7 +10,7 @@ activity (e.g.
 get_activity_info(activityid)
 }
 \arguments{
-\item{activityid}{a character. It is the DEIMS.iD of activity make from
+\item{activityid}{A `character`. It is the DEIMS.iD of activity make from
 DEIMS-SDR website. DEIMS.iD information
 \href{https://deims.org/docs/deimsid.html}{here}.
 The DEIMS.iD of activity is the URL for the activity page.}

[…]
