+++ R/get_network_research_topics.R
@@ -1,10 +1,8 @@
-#' @title Return a tibble object containing Research Topics collected by all
+#' Obtain a list of Research Topics handled in an eLTER Network.
+#' @description This function obtains Research Topics collected by all
 #' of the eLTER sites belonging to an eLTER Network (e.g.
-#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
+#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
 #' Italy network}), as a stored into \href{https://deims.org}{DEIMS-SDR}.
-#' @description This function obtains a list of Research Topics
-#' handled in the eLTER Network (e.g. Italy) through
-#' the DEIMS-SDR sites API.
 #' @param networkDEIMSID A `character`. The DEIMS.iD of a network
 #' from DEIMS-SDR website. DEIMS.iD information 
 #' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
