+++ R/get_network_envcharacts.R
@@ -1,11 +1,10 @@
-#' @title Return a tibble object containing Environmental Characteristics,
-#' stored in \href{https://deims.org/}{DEIMS-SDR catalogue}, of all the
-#' eLTER sites belonging to an eLTER Network (e.g.
-#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
-#' Italy network}).
-#' @description This function obtains several Environmental Characteristics:
-#' title, URI, geocoordinates, country name, and elevation
-#' of sites in an eLTER Network (e.g. Italy) through the DEIMS-SDR sites API.
+#' Obtain a list of all Environmental Characteristics of sites in an
+#' eLTER Network.
+#' @description This function obtains all Environmental Characteristics:
+#' title, URI, geo-coordinates, country name, and elevation
+#' of eLTER Network sites (e.g.
+#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
+#' Italy network}), through the DEIMS-SDR API.
 #' @param networkDEIMSID A `character`. DEIMS.iD of network
 #' from DEIMS-SDR website. DEIMS.iD information 
 #' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
