+++ R/get_network_sites.R
@@ -1,10 +1,10 @@
-#' @title Return a spatial point vector object containing name and
-#' geo-information, stored in \href{https://deims.org/}{DEIMS-SDR catalogue},
-#' of all the eLTER sites belonging to an eLTER Network (e.g.
-#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
+#' Retrieve a list of sites in an eLTER Network.
+#' @description This function return a spatial point vector object including
+#' title, date late updated, URI, and coordinates, stored in
+#' \href{https://deims.org/}{DEIMS-SDR catalogue}, of all the eLTER sites
+#' belonging to an eLTER Network (e.g.
+#' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
 #' Italy network}).
-#' @description This function retrieves a list of sites in the provided network
-#' including title, date late updated, URI, and coordinates
 #' @param networkDEIMSID A `character`. The DEIMS.iD of the network from
 #' DEIMS-SDR website. DEIMS.iD information 
 #' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of
