### 2022-03-10 4fcb9a0
clarified the description of functions and of siteNum parameter
