+++ R/get_ilter_envcharacts.R
@@ -5,9 +5,23 @@
 #' \href{https://www.ilter.network/network/global-coverage}{ILTER sites
 #' (more than 1200 around the world)},
 #' through the DEIMS-SDR API.
+#' 
+#' This function aims to gather in a unique tibble all the Environmental
+#' Characteristics the whole of ILTER sites. If you want to obtain this
+#' results, it should be noted that the execution time for this function
+#' is very high.
+#' 
+#' If the objective is obtain information about Environmental Characteristics
+#' on a few sites, it is better to use other more specific functions (e.g.
+#' \href{get_network_envcharacts.html}{`get_network_envcharacts()`} or
+#' \href{get_site_info.html}{`get_site_info()`}) or using
+#' other methods
+#' (\href{../../articles/sites_information.html}{How to about sites informations}).
 #' @param sitesNum A `integer`. It is the number of the sites that are
-#' read to get the information. Use this parameters moreover for provide
-#' example of this function.
+#' read to get the information. Use this parameter only for get
+#' example of the output of this function. If the value of sitesNum is
+#' 0 (default) all the ILTER sites will be parsed and the waiting time will
+#' be long.
 #' @return The output of the function is a `tibble` containing all the
 #' Environmental Characteristics of ILTER's sites.
 #' @author Alessandro Oggioni, phD (2020) <oggioni.a@@irea.cnr.it>
