+++ man/get_ilter_envcharacts.Rd
@@ -8,8 +8,10 @@ get_ilter_envcharacts(sitesNum = 0)
 }
 \arguments{
 \item{sitesNum}{A `integer`. It is the number of the sites that are
-read to get the information. Use this parameters moreover for provide
-example of this function.}
+read to get the information. Use this parameter only for get
+example of the output of this function. If the value of sitesNum is
+0 (default) all the ILTER sites will be parsed and the waiting time will
+be long.}
 }
 \value{
 The output of the function is a `tibble` containing all the
@@ -22,6 +24,18 @@ of all
 \href{https://www.ilter.network/network/global-coverage}{ILTER sites
 (more than 1200 around the world)},
 through the DEIMS-SDR API.
+
+This function aims to gather in a unique tibble all the Environmental
+Characteristics the whole of ILTER sites. If you want to obtain this
+results, it should be noted that the execution time for this function
+is very high.
+
+If the objective is obtain information about Environmental Characteristics
+on a few sites, it is better to use other more specific functions (e.g.
+\href{get_network_envcharacts.html}{`get_network_envcharacts()`} or
+\href{get_site_info.html}{`get_site_info()`}) or using
+other methods
+(\href{../../articles/sites_information.html}{How to about sites informations}).
 }
 \examples{
 \dontrun{
