### 2022-03-11 #485 @oggioniale
> OK, here's my review. I haven't commented on some of the more detailed aspects of the review (such as strict adherence to the various reviewing/packaging guides) but please let me know if I've omitted anything crucial and I'll update it.
>
> I'll keep an eye out for any follow up questions that may arise but will be busy tomorrow afternoon.
>
...

Dear @WillOnGit,
thanks you for your review. Your comments, suggestions and improvements request, based on your programming experience, are a powerful impetus for improving the package.
In the next days our commits for answer to your issues.

Ale
