### 2022-03-22 3b8f428
deleted GBIF occurrences when they are form
iNaturalist
