### 2022-03-23 114b80c
Add functions for mapping outputs from get_site_speciesOccurrences into LTER reporting format
