+++ R/eLTER_reporting.R
@@ -1,0 +1,352 @@
+#' Harmonize outputs of get_site_speciesOccurrence and map them into eLTER reporting format
+#' @param x Tibble like one that can be obtained by
+#'   as_tibble(get_site_speciesOccurrence(deimsid, "gbif")$gbif)
+#' @param deimsid deims id of the site of interest
+#' @return list with the following named elements:
+#'   * deimsid: the same deimsid passed in input
+#'   * source: one of "gbif", "inat", "obis"
+#'   * data_mapping: tibble structured according to data_mapping of 
+#'      eLTER reporting format
+#'   * reference_TAXA: tibble structured according to reference_TAXA of 
+#'      eLTER reporting format
+#'   * reference_VARIABLES: tibble structured according to reference_VARIABLES 
+#'      of eLTER reporting format
+#' @author Paolo Tagliolato, phD (2020) \email{tagliolato.p@@irea.cnr.it}
+#' @importFrom tibble as_tibble
+#' @importFrom dplyr select mutate filter
+#' @export
+gbif2elter <- function(x, deimsid){#, reference_variables=reference_variables_gbif){
+  # "variables" (in this case they could be defined "custom fields", being just record metadata) 
+  # to be described in the reference.csv file
+  reference_variables<-reference_variables_gbif
+  # reference_variables<-as_tibble(data.frame(
+  #   FIELD_NAME=rep("VARIABLE",4),
+  #   VARIABLE_CODE=c("RECORD_ID","DATASET_KEY","INSTITUTION_CODE","LICENSE"),
+  #   VARIABLE_NAME=c("gbif record id",
+  #                   "gbif datasetKey", 
+  #                   "gbif institution code", 
+  #                   "licence applied to each gbif record"),
+  #   VARIABLE_DEFINITION=c("","","","")
+  # ))
+  
+  custom_fields=reference_variables$VARIABLE_CODE
+  
+  # compose output for eLTER reporting format 
+  # (intermediate step: info both for data and for reference)
+  lter_temp <- x %>% .addSiteCode(deimsid) %>% 
+    # add new (computed) columns
+    dplyr::mutate(VARIABLE="occurrence", 
+                  VALUE=TRUE, 
+                  TIME=paste0(eventDate,"T",eventTime),
+                  ref_CODE_URL=paste0("https://www.gbif.org/species/",taxonKey)) %>%
+    # select and alias existing columns
+    dplyr::select(
+      SITE_CODE,
+      ABS_POSITION=geometry,
+      TAXA = taxonKey, 
+      VARIABLE , 
+      VALUE, 
+      TIME,
+      ORG_NAME=prov,
+      ref_NAME=acceptedScientificName,
+      ref_CODE_URL,
+      # custom fields/variables
+      RECORD_ID=key, #tblGarda_gbif$institutionCode
+      DATASET_KEY=datasetKey, # UUID gbif
+      INSTITUTION_CODE=institutionCode,
+      LICENSE=license # this is a NEW field (a metadata of the record). Is it possibile to map it?
+    )
+  
+  # compose species list for reference_TAXA.csv
+  reference_TAXA<-lter_temp %>% 
+    mutate(FIELD_NAME="TAXA") %>%
+    select(FIELD_NAME,CODE=TAXA, NAME=ref_NAME, 
+           CODE_URL=ref_CODE_URL) %>% unique()
+  
+  data_mapping_LTER <- lter_temp %>% 
+    dplyr::select(any_of(c(
+      eLTER_reporting.mapping.coreFields,
+      eLTER_reporting.mapping.extendedFields,
+      custom_fields
[…]
