### 2022-03-23 69ba506
change eLTER_reporting function names to adhere code styling
