+++ R/eLTER_reporting.R
@@ -15,8 +15,9 @@
 #' @importFrom tibble as_tibble
 #' @importFrom dplyr select mutate filter
 #' @export
-map_occ_gbif2elter <- function(x, deimsid){#, reference_variables=reference_variables_gbif){
-  # "variables" (in this case they could be defined "custom fields", being just record metadata) 
+map_occ_gbif2elter <- function(x, deimsid){
+  # "variables" (in this case they could be defined "custom fields", 
+  # being just record metadata) 
   # to be described in the reference.csv file
   reference_variables<-reference_variables_gbif
   # reference_variables<-as_tibble(data.frame(
@@ -33,7 +34,7 @@ map_occ_gbif2elter <- function(x, deimsid){#, reference_variables=reference_vari
   
   # compose output for eLTER reporting format 
   # (intermediate step: info both for data and for reference)
-  lter_temp <- x %>% .addSiteCode(deimsid) %>% 
+  lter_temp <- x %>% .add_site_code(deimsid) %>% 
     # add new (computed) columns
     dplyr::mutate(VARIABLE="occurrence", 
                   VALUE=TRUE, 
@@ -77,7 +78,7 @@ map_occ_gbif2elter <- function(x, deimsid){#, reference_variables=reference_vari
               reference_VARIABLES=reference_variables_gbif))
 }
 
-#' @inherit gbif2elter
+#' @inherit map_occ_gbif2elter
 #' @export
 map_occ_inat2elter <- function(x, deimsid){
   reference_variables<-reference_variables_inat
@@ -92,7 +93,7 @@ map_occ_inat2elter <- function(x, deimsid){
   
   # compose output for eLTER reporting format 
   # (intermediate step: info both for data and for reference)
-  lter_temp <- x %>% .addSiteCode(deimsid) %>%
+  lter_temp <- x %>% .add_site_code(deimsid) %>%
     # add new (computed) columns
     dplyr::mutate(
       SITE_CODE = gardaid,
@@ -144,7 +145,7 @@ map_occ_inat2elter <- function(x, deimsid){
               reference_VARIABLES=reference_variables_inat))
 }
 
-#' @inherit gbif2elter
+#' @inherit map_occ_gbif2elter
 #' @export
 map_occ_obis2elter <- function(x, deimsid){
   # "variables" (in this case they could be defined "custom fields", being just record metadata) to be described in the reference.csv file
@@ -159,7 +160,7 @@ map_occ_obis2elter <- function(x, deimsid){
   
   # compose output for eLTER reporting format 
   # (intermediate step: info both for data and for reference)
-  lter_temp <- x %>% .addSiteCode(deimsid) %>%
+  lter_temp <- x %>% .add_site_code(deimsid) %>%
     # add new (computed) columns
     dplyr::mutate(
       SITE_CODE = GOFid,
@@ -221,9 +222,10 @@ map_occ_obis2elter <- function(x, deimsid){
 
 #' creates an archive with files following the eLTER reportingFormat
 #' @param eLterReportOut a list like the one created by gbif2elter
-#' @description creates a zip archive named biodiv_occurrence_site_<deimsid_code>_<source>.zip
-#'    where <deimsid_code> is the uuid in the last part of the deimsid, 
-#'    and <source> is one of "gbif", "inat", "obis"
+#' @description creates a zip archive named 
+#' biodiv_occurrence_site_<deimsid_code>_<source>.zip
+#' where <deimsid_code> is the uuid in the last part of the deimsid, 
+#' and <source> is one of "gbif", "inat", "obis"
 #' @author Paolo Tagliolato, phD (2020) \email{tagliolato.p@@irea.cnr.it}
 #' @importFrom utils zip write.csv2
 #' @export
[…]
