+++ _pkgdown.yaml
@@ -47,6 +47,13 @@ reference:
       - '`get_sos_procedurelist`'
       - '`taxon_id_pesi`'
       - '`taxon_id_worms`'
+  - title: eLTER reporting format
+    desc: Functions dealing with eLTER reporting format
+    contents:
+      - '`map_occ_gbif2elter`'
+      - '`map_occ_inat2elter`'
+      - '`map_occ_obis2elter`'
+      - '`save_occ_eLTER_reporting_Archive`'
 navbar:
   type: inverse
   left: []
