### 2022-03-23 30c7f06
integrate eLTER reporting functions
