+++ R/eLTER_reporting.R
@@ -1,99 +1,99 @@
-#' Harmonize outputs of get_site_speciesOccurrence and map them into eLTER reporting format
+#' Harmonize outputs of get_site_speciesOccurrence and map them into eLTER
+#' reporting format
 #' @param x Tibble like one that can be obtained by
-#'   as_tibble(get_site_speciesOccurrence(deimsid, "gbif")$gbif)
+#' as_tibble(get_site_speciesOccurrence(deimsid, "gbif")$gbif)
 #' @param deimsid deims id of the site of interest
 #' @return list with the following named elements:
-#'   * deimsid: the same deimsid passed in input
-#'   * source: one of "gbif", "inat", "obis"
-#'   * data_mapping: tibble structured according to data_mapping of 
-#'      eLTER reporting format
-#'   * reference_TAXA: tibble structured according to reference_TAXA of 
-#'      eLTER reporting format
-#'   * reference_VARIABLES: tibble structured according to reference_VARIABLES 
-#'      of eLTER reporting format
+#' * deimsid: the same deimsid passed in input
+#' * source: one of "gbif", "inat", "obis"
+#' * data_mapping: tibble structured according to data_mapping of
+#' eLTER reporting format
+#' * reference_TAXA: tibble structured according to reference_TAXA of
+#' eLTER reporting format
+#' * reference_VARIABLES: tibble structured according to reference_VARIABLES
+#' of eLTER reporting format
 #' @author Paolo Tagliolato, phD (2020) \email{tagliolato.p@@irea.cnr.it}
+#' @author Martina Zilioli \email{zilioli.m@@irea.cnr.it}
+#' @author Alessandro Oggioni, PhD (2020) \email{oggioni.a@@irea.cnr.it}
 #' @importFrom tibble as_tibble
 #' @importFrom dplyr select mutate filter
 #' @export
-map_occ_gbif2elter <- function(x, deimsid){
-  # "variables" (in this case they could be defined "custom fields", 
-  # being just record metadata) 
+map_occ_gbif2elter <- function(x, deimsid) {
+  # "variables" (in this case they could be defined "custom fields",
+  # being just record metadata)
   # to be described in the reference.csv file
-  reference_variables<-reference_variables_gbif
-  # reference_variables<-as_tibble(data.frame(
-  #   FIELD_NAME=rep("VARIABLE",4),
-  #   VARIABLE_CODE=c("RECORD_ID","DATASET_KEY","INSTITUTION_CODE","LICENSE"),
-  #   VARIABLE_NAME=c("gbif record id",
-  #                   "gbif datasetKey", 
-  #                   "gbif institution code", 
-  #                   "licence applied to each gbif record"),
-  #   VARIABLE_DEFINITION=c("","","","")
-  # ))
-  
-  custom_fields=reference_variables$VARIABLE_CODE
-  
-  # compose output for eLTER reporting format 
+  reference_variables <- reference_variables_gbif
+  custom_fields <- reference_variables$VARIABLE_CODE
+
+  # compose output for eLTER reporting format
   # (intermediate step: info both for data and for reference)
-  lter_temp <- x %>% .add_site_code(deimsid) %>% 
+  lter_temp <- x %>%
+    .add_site_code(deimsid) %>%
     # add new (computed) columns
-    dplyr::mutate(VARIABLE="occurrence", 
-                  VALUE=TRUE, 
-                  TIME=paste0(eventDate,"T",eventTime),
[…]
