+++ man/map_occ_inat2elter.Rd
@@ -1,0 +1,37 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/eLTER_reporting.R
+\name{map_occ_inat2elter}
+\alias{map_occ_inat2elter}
+\title{Harmonize outputs of get_site_speciesOccurrence and map them into eLTER
+reporting format}
+\usage{
+map_occ_inat2elter(x, deimsid)
+}
+\arguments{
+\item{x}{Tibble like one that can be obtained by
+as_tibble(get_site_speciesOccurrence(deimsid, "gbif")$gbif)}
+
+\item{deimsid}{deims id of the site of interest}
+}
+\value{
+list with the following named elements:
+* deimsid: the same deimsid passed in input
+* source: one of "gbif", "inat", "obis"
+* data_mapping: tibble structured according to data_mapping of
+eLTER reporting format
+* reference_TAXA: tibble structured according to reference_TAXA of
+eLTER reporting format
+* reference_VARIABLES: tibble structured according to reference_VARIABLES
+of eLTER reporting format
+}
+\description{
+Harmonize outputs of get_site_speciesOccurrence and map them into eLTER
+reporting format
+}
+\author{
+Paolo Tagliolato, phD (2020) \email{tagliolato.p@irea.cnr.it}
+
+Martina Zilioli \email{zilioli.m@irea.cnr.it}
+
+Alessandro Oggioni, PhD (2020) \email{oggioni.a@irea.cnr.it}
+}
