+++ vignettes/sites_information.Rmd
@@ -1,0 +1,152 @@
+---
+title: "How to get the info of many than one LTER Sites"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{How to get the info of many than one LTER Sites}
+  %\VignetteEngine{knitr::knitr}
+  %\usepackage[UTF-8]{inputenc}
+  bibliography: ./references/occurrences_into_site_refs.bib
+---
+
+Taking into account the guide on ["How to obtain site info?"](./site_information.html), here is offered two examples on how to aggregate information of many eLTER sites.
+
+## By get_network_* functions: {#firstExample}
+
+Through the 5 functions developed for get network information, access and download of some information of an entire eLTER network can be made.
+
+### Get general information of the network's sites:
+
+Knowing the network elTER id (DEIMS.iD of the Network, e.g. LTER-Italy <https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3>), the [`get_network_sites`](https://oggioniale.github.io/ReLTER/reference/get_network_sites.html) can download general info, such as name, DEIMS.iD and coordinates of the sites belonging to the network selected. A map of the sites can be returned.
+
+```{r getnetworksites, warning=FALSE, message=FALSE}
+library(dplyr)
+
+listItaSites <- ReLTER::get_network_sites(
+ networkDEIMSID =
+   "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
+)
+knitr::kable(
+  listItaSites[1:10, ],
+  caption = "The list of site for LTER-Italy network"
+)
+
+listItaSitesMap <- leaflet::leaflet(listItaSites) %>%
+  leaflet::addProviderTiles(provider = "CartoDB.PositronNoLabels",
+                            group = "Basemap",
+                            layerId = 123) %>%
+  leaflet::addTiles("http://{s}.basemaps.cartocdn.com/light_only_labels/{z}/{x}/{y}.png") %>%
+  leaflet::addCircleMarkers(
+    data = listItaSites,
+    radius = 3,
+    weight = 2,
+    opacity = 0.5,
+    fill = TRUE,
+    fillOpacity = 0.2
+  )
+listItaSitesMap
+```
+
+### Get other information of the network's sites:
+
+Another 4 functions are been implemented to obtain network's sites information. Each functions get specific information:
+
+* Environmental characteristics by [`get_network_envcharacts`](https://oggioniale.github.io/ReLTER/reference/get_network_envcharacts.html)
+* Parameters by [`get_network_parameters`](https://oggioniale.github.io/ReLTER/reference/get_network_parameters.html)
+* Related resources  by [`get_network_related_resources`](https://oggioniale.github.io/ReLTER/reference/get_network_related_resources.html)
+* Research topics by [`get_network_research_topics`](https://oggioniale.github.io/ReLTER/reference/get_network_research_topics.html)
+
+The following example perform the request to get all related resources (e.g. activities, datasets, etc.) of the network. The output is a table containing the title, id and time stamp of the last changes of the related resources shared by the network's sites.
+
+```{r relResNet, warning=FALSE, message=FALSE}
+listRelatedResources <- ReLTER::get_network_related_resources(
+  networkDEIMSID = "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
+)
+
+# Table of the network's related resources
+knitr::kable(
+  listRelatedResources,
[…]
