+++ vignettes/sites_contacts.Rmd
@@ -1,0 +1,89 @@
+---
+title: "How to get the list of contacts from list of DEIMS.iD within csv file"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{How to get the list of contacts from list of DEIMS.iD within csv file}
+  %\VignetteEngine{knitr::knitr}
+  %\usepackage[UTF-8]{inputenc}
+  bibliography: ./references/occurrences_into_site_refs.bib
+---
+
+Starting from the [DEIMS-SDR "Advanced Search - Sites" interface](https://deims.org/search/sites), an user can download a csv file of the sites. Some filters can be used for selecting sites based on countries, projects, Biome, etc. The list of sites selected can be downloaded in csv format file.
+
+This documentation, starting from this file csv, shows how to:
+
+## List of the contacts
+
+1. obtain the list of contacts for each site
+
+```{r listContacts, warning=FALSE, message=FALSE}
+library(dplyr)
+site_results <- read.csv("./data/site_contacts_data.csv", sep=";") %>%
+  dplyr::as_tibble()
+
+allSiteContact <- lapply(
+    as.list(
+      paste0("https://deims.org/", site_results$DEIMS.ID)
+    ),
+    ReLTER::get_site_info,
+    category = "Contacts"
+  )
+contacts <- tibble::tibble(
+  siteName = NA,
+  managerName = NA,
+  managerEmail = NA,
+  managerORCID = NA
+)
+for (i in seq_len(length(allSiteContact))) {
+  if (is.na(allSiteContact[[i]]$generalInfo.siteManager[[1]])) {
+    contacts <- contacts %>% 
+      tibble::add_row(
+        siteName = allSiteContact[[i]]$title,
+        managerName = NA,
+        managerEmail = NA,
+        managerORCID = NA
+      )
+  } else {
+    contacts <- contacts %>% 
+      tibble::add_row(
+        siteName = allSiteContact[[i]]$title,
+        managerName = allSiteContact[[i]]$generalInfo.siteManager[[1]]$name,
+        managerEmail = allSiteContact[[i]]$generalInfo.siteManager[[1]]$email,
+        managerORCID = allSiteContact[[i]]$generalInfo.siteManager[[1]]$orcid
+      )
+  }
+}
+# Contacts table
+knitr::kable(
+  contacts,
+  caption = "List of the contacts",
+  booktabs = TRUE, longtable = TRUE
+)
+```
+
+## Map of the sites
+
+2. how to map the sites selected
+
+```{r geoSites, warning=FALSE, message=FALSE}
+library(dplyr)
+site_results_geo <- sf::st_as_sf(
+      site_results,
+      wkt = "Coordinates"
+    )
+
+listItaSitesMap <- leaflet::leaflet(site_results_geo) %>%
+  leaflet::addProviderTiles(provider = "CartoDB.PositronNoLabels",
+                            group = "Basemap",
+                            layerId = 123) %>%
+  leaflet::addTiles("http://{s}.basemaps.cartocdn.com/light_only_labels/{z}/{x}/{y}.png") %>%
+  leaflet::addCircleMarkers(
+    data = site_results_geo,
+    radius = 3,
+    weight = 2,
+    opacity = 0.5,
+    fill = TRUE,
+    fillOpacity = 0.2
+  )
+listItaSitesMap
+```
