+++ vignettes/site_information.Rmd
@@ -5,7 +5,7 @@ vignette: >
   %\VignetteIndexEntry{How to get the info of a (one) LTER Site by knowing the DEIMS.ID}
   %\VignetteEngine{knitr::knitr}
   %\usepackage[UTF-8]{inputenc}
-  bibliography: ./references/occurrences_into_site_refs.bib
+author: "Alessandro Oggioni"
 ---
 
 For each LTER site the information are provided, by site manager, into DEIMS-SDR registry (see <https://deims.org>). The LTER site's information are grouped into main sections, the most important ones are:
