+++ vignettes/installation.Rmd
@@ -5,14 +5,13 @@ vignette: >
   %\VignetteIndexEntry{How to install `{ReLTER}` package}
   %\VignetteEngine{knitr::knitr}
   %\usepackage[UTF-8]{inputenc}
+author: "Alessandro Oggioni, Paolo Tagliolato"
 ---
 
 You can install the `{ReLTER}` package from [GitHub](https://github.com/oggioniale/ReLTER) with:
 
-```{r, installationInfo, eval = FALSE}
-install.packages("devtools")
-devtools::install_github("oggioniale/ReLTER")
-```
+    install.packages("devtools")
+    devtools::install_github("oggioniale/ReLTER")
 
 If you want to install different package branch (e.g. 'dev') can you use this command:
 
