### 2022-03-26 98085b0
added new param to these 2 functions
