+++ R/get_dataset_info.R
@@ -7,6 +7,8 @@
 #' DEIMS-SDR website. DEIMS.iD information 
 #' \href{https://deims.org/docs/deimsid.html}{here}.
 #' The DEIMS.iD of dataset is the URL for the dataset page.
+#' @param show_map A `boolean`. If TRUE a Leaflet map with occurrences
+#' is shown. Default FALSE.
 #' @return The output of the function is a `tibble` with main features
 #' of the site and the related resources collected by site.
 #' @author Alessandro Oggioni, phD (2020) \email{oggioni.a@@irea.cnr.it}
@@ -21,12 +23,13 @@
 #' @examples
 #' tDataset <- get_dataset_info(
 #'   datasetid =
-#'   "https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef"
+#'   "https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef",
+#'   show_map = TRUE
 #' )
 #' tDataset
 #'
 ### function get_dataset_info
-get_dataset_info <- function(datasetid) {
+get_dataset_info <- function(datasetid, show_map = FALSE) {
   q <- '{
        title: .title,
        abstract: .attributes.general.abstract,
@@ -104,8 +107,12 @@ get_dataset_info <- function(datasetid) {
             map <- map %>%
               leaflet::addPolygons()
           }
-          print(map)
-          geoDataset
+          if (show_map = TRUE) {
+            print(map)
+            geoDataset
+          } else {
+            geoDataset
+          }
         } else {
           map <- leaflet::leaflet() %>%
             leaflet::addTiles()
