### 2022-03-27 e947dcd
Adds new examples to the README
showing how to obtain DEIMS ID using country name and site name.
Minor grammar edits to the Descriptions in all functions
