+++ R/get_site_research_topics.R
@@ -1,8 +1,8 @@
 #' @title eLTER get_site_research_topics function
 #' @description This function obtains a list of research topics handled
 #' at an eLTER site through the DEIMS-SDR sites API.
-#' @param deimsid  A `character`. The DEIMS.iD of the site from
-#' DEIMS-SDR website. DEIMS.iD information 
+#' @param deimsid  A `character`. The DEIMS ID of the site from
+#' DEIMS-SDR website. DEIMS ID information 
 #' \href{https://deims.org/docs/deimsid.html}{here}.
 #' @return The output of the function is a `tibble` with main features of the
 #' site and a list of the research topics handled in this site.
@@ -47,7 +47,7 @@ get_site_research_topics <- function(deimsid) {
     }
   } else {
     message("\n----\nThe requested page could not be found.
-Please check again the DEIMS.iD\n----\n")
+Please check again the DEIMS ID\n----\n")
     researchTopics <- NULL
   }
   researchTopics
