+++ R/get_site_parameters.R
@@ -1,8 +1,8 @@
 #' @title eLTER get_site_parameters function
 #' @description This function obtains the list of parameters measured
 #' in the eLTER site through the DEIMS-SDR sites API.
-#' @param deimsid A `character`. It is the DEIMS.iD of the site from
-#' DEIMS-SDR website. DEIMS.iD information 
+#' @param deimsid A `character`. It is the DEIMS ID of the site from
+#' DEIMS-SDR website. DEIMS ID information 
 #' \href{https://deims.org/docs/deimsid.html}{here}.
 #' @return The output of the function is a `tibble` with main features of the
 #' site and the parameters collected.
@@ -47,7 +47,7 @@ get_site_parameters <- function(deimsid) {
     }
   } else {
     message("\n----\nThe requested page could not be found.
-Please check again the DEIMS.iD\n----\n")
+Please check again the DEIMS ID\n----\n")
     parameters <- NULL
   }
   parameters
