+++ R/produce_site_parameters_waffle.R
@@ -1,8 +1,8 @@
 #' Produce a waffle chart of the parameters collected in a site LTER.
 #' @description Return a waffle chart of Environmental Parameters, as a stored in
 #' \href{https://deims.org/}{DEIMS-SDR catalogue}, of a single eLTER site.
-#' @param deimsid A `character`. The DEIMS.iD of site/network from:
-#' DEIMS-SDR website. DEIMS.iD information 
+#' @param deimsid A `character`. The DEIMS ID of site/network from:
+#' DEIMS-SDR website. DEIMS ID information 
 #' \href{https://deims.org/docs/deimsid.html}{here}.
 #' @return The output of the function is a waffle chart and a `tibble`. Each
 #' of the squares represents a parameters measured into the selected eLTER
@@ -92,7 +92,7 @@ produce_site_parameters_waffle <- function(deimsid) {
     params
   } else {
     message("\n----\nThe requested page could not be found.
-Please check again the DEIMS.iD\n----\n")
+Please check again the DEIMS ID\n----\n")
     waffle <- NULL
     params <- NULL
   }

[…]
