### 2022-03-27 aa964d0
Fixes test errors in produce_network_points_map.R
