### 2022-03-28 f7afddb
Fixes get_ilter_research_topics
