+++ tests/testthat/test-get_ilter_research_topics.R
@@ -19,7 +19,7 @@ skip_if_offline(host = "deims.org")
 
 test_that("Output of ILTER research topics function constructs 'tibble' as
           expected", {
-  result <- ReLTER::get_ilter_research_topics(sitesNum = 4)
+  result <- ReLTER::get_ilter_research_topics(sitesNum = 20)
   expect_s3_class(result, "tbl_df")
   expect_true(ncol(result) == 2)
   expect_true(all(names(result) == c(
