### 2022-03-28 b2d7576
vignettes and image of functions done
