+++ R/get_activity_info.R
@@ -26,6 +26,9 @@
 #' )
 #' activities
 #'
+#' @section Here is an example graphic output:
+#' \figure{get_activity_info_fig.png}{Map of "Study of non-indigenous (alien) species in the Mar Piccolo of Taranto" activity}
+#'
 ### function get_activity_info
 get_activity_info <- function(activityid, show_map = FALSE) {
   q <- '{
