+++ vignettes/ReLTER.Rmd
@@ -1,97 +1,0 @@
----
-title: 'ReLTER: An interface to the eLTER for the R statistical programming environment - v1.0.0'
-author: "Alessadro Oggioni ([alessandro.oggioni@cnr.it](mailto:alessandro.oggioni@cnr.it)),
-  Micha Silver ([silverm@post.bgu.ac.il](mailto:silverm@post.bgu.ac.il)),
-  Luigi Ranghetti ([ranghetti.l@irea.cnr.it](mailto:ranghetti.l@irea.cnr.it)),
-  Paolo Tagliolato ([tagliolato.a@irea.cnr.it](mailto:tagliolato.a@irea.cnr.it))"
-# bibliography: ReLTER.bib
-date: '`r Sys.Date()`'
-output:
-  rmarkdown::html_vignette:
-    fig_caption: yes
-    number_section: yes
-    toc: no
-    toc_depth: 2
-  rmarkdown::pdf_document:
-    fig_caption: yes
-    number_sections: yes
-    toc: yes
-    toc_depth: 2
-  rmarkdown::html_document:
-    fig_caption: yes
-    number_section: yes
-    toc: no
-    toc_depth: 1
-urlcolor: blue
-linkcolor: blue
-vignette: >
-  \usepackage[utf8]{inputenc}
-  %\VignetteIndexEntry{ReLTER: An interface to the eLTER for the R statistical programming environment - v1.0.0}
-  %\VignetteEngine{knitr::rmarkdown}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "vignettes/figures/vignettes-",
-  fig.ext = "png",
-  out.width = "100%"
-)
-```
-<img src='../man/figures/logo.png' align="right" height="139" />
-
-<span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span> provides access to [DEIMS-SDR](https://deims.org/), and allows interaction with data and software implemented by [Long Term Ecological Research Research Infrastructure (eLTER-RI)](https://elter-ri.eu) thus improving data sharing among [European LTER projects](https://elter-ri.eu/projects).
-
-<br>
-<br>
-<br>
-<br>
-<br>
-
-# Documentation of <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span>
-
-Visit the <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span> [website](https://oggioniale.github.io/ReLTER/) for further documentation, examples, and installation of the package.
-
-# Installation of <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span>
-
-You can install the development version of <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span> from [GitHub](https://github.com/oggioniale/ReLTER) with:
-
-```{r, installationInfo, eval = FALSE}
-install.packages("devtools")
-devtools::install_github("oggioniale/ReLTER")
-```
-If you wish to help develop this package, please follow the [contributing guidelines](https://oggioniale.github.io/ReLTER/CONTRIBUTING.html).
-
-# Examples
-
-## Ex1 ...
-<img alt="xxx" width="250" style="padding: 0 15px; float: right;"/>
-
-# Problems and Issues of <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span>
-
-Please report any issues you may encounter in our [issues page on GitHub](https://github.com/oggioniale/ReLTER/issues)
-
-# Citation of <span style="color:#2878B3;font-size:90%;font-weight:normal;">ReLTER</span>
-
[…]

[…]
