+++ README.Rmd
@@ -17,8 +17,7 @@ knitr::opts_chunk$set(
   comment = "#>",
   fig.path = "man/figures/README-",
   fig.ext = 'png',
-  out.width = "100%"
-)
+  out.width = "100%")
 ```
 
 <!-- badges: start -->
@@ -61,7 +60,7 @@ network. This package was conceived within eLTER H2020 project and will
 help advance the development of European Long-Term Ecosystem Research
 Infrastructures ([eLTER RI](https://elter-ri.eu)).
 
-The `{ReLTER}` package functions in particular allow to:
+The `{ReLTER}` package functions in particular allows to:
 
 -   retrieve the information about entities (e.g. sites, datasets, and
     activities) shared by [DEIMS-SDR](https://deims.org/) (see e.g.
@@ -145,8 +144,7 @@ These examples demonstrate interaction with the DEIMS-SDR
 [API](https://deims.org/api).
 
 #### *get_ilter_generalinfo()*
-
-For the initial example, the *get_ilter_generalinfo* function is called
+In this initial example, the *get_ilter_generalinfo* function is called
 to find the DEIMS ID for a specific eLTER site. The function takes
 parameters "country_name" and "site_name", and returns the site URL (the
 DEIMS ID) in the DEIMS SDR infrastructure.
@@ -155,7 +153,7 @@ DEIMS ID) in the DEIMS SDR infrastructure.
 # Obtain DEIMS ID for the Eisenwurzen site in Austria
 eisenwurzen <- ReLTER::get_ilter_generalinfo(country_name = "Austri",
                                      site_name = "Eisen")
-# extract DEIMS.Id
+# extract DEIMS ID
 (eisenwurzen_deimsid <- eisenwurzen$uri)
 ```
 
@@ -163,10 +161,10 @@ In some countries, many sites have similar names. Here is a method to
 find the DEIMS ID of the specific site of interest.
 
 ```{r ilter_multi, message=FALSE, warning=FALSE}
-# Obtain DEIMS ID for the Eisenwurzen site in Austria
+# Obtain DEIMS ID for sites in Germany, with names like "TERENO"
 sites_germany <- ReLTER::get_ilter_generalinfo(country_name = "Germ",
                                      site_name = "TERENO")
-# List full names of all sites
+# List full names of these sites
 sites_germany$title
 # Get the DEIMS ID of the site of interest by partial string match on the title
 idx <- which(stringr::str_detect(sites_germany$title,
@@ -176,10 +174,7 @@ idx <- which(stringr::str_detect(sites_germany$title,
 
 #### *get_site_info(category = "Boundaries")*
 
-Next, the *get_site_info* function, using category "Boundaries", creates
-a Leaflet map overlayed with the boundaries of the site (e.g. Lake
-Maggiore) using the site geographical boundaries provided by
-[DEIMS-SDR](https://deims.org/). This function returns an `sf` object.
+Next, the *get_site_info* function, using category "Boundaries", returns an `sf` object. Furthermore, a map overlayed with the boundaries of the site is displayed, using the site geographical boundaries provided by [DEIMS-SDR](https://deims.org/).
 
 ```{r, message = FALSE, warning = FALSE, fig.show='hold', exampleSiteBoundaries}
 library(dplyr)
@@ -187,10 +182,14 @@ siteBoundaries <- ReLTER::get_site_info(
[…]
