+++ README.md
@@ -31,56 +31,54 @@ developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.re
 [![R build
 status](https://github.com/rossellhayes/ipa/workflows/R-CMD-check/badge.svg)](https://github.com/rossellhayes/ipa/actions)
 [![codecov](https://codecov.io/gh/oggioniale/ReLTER/branch/dev/graph/badge.svg)](https://codecov.io/gh/oggioniale/ReLTER)
+
 <!-- badges: end -->
 
-`{ReLTER}` is an R package that: provides access to
-[DEIMS-SDR](https://deims.org/), allows interact with software
-implemented by eLTER Research Infrastructure (RI) and improves the
-data/information shared by them.
-
-`{ReLTER}` is a R package devoted to access, interact and improve the
-information and the data shared by Long Term Ecological Research (LTER)
-network. This package is born within eLTER H2020 major project that will
+`{ReLTER}` is an R package that provides access to
+[DEIMS-SDR](https://deims.org/), allowing to interact with software
+implemented by eLTER Research Infrastructure (RI) and improving the
+data/information shared among the Long Term Ecological Research (LTER)
+network. This package was conceived within eLTER H2020 project and will
 help advance the development of European Long-Term Ecosystem Research
 Infrastructures ([eLTER RI](https://elter-ri.eu)).
 
-The `{ReLTER}` package functions in particular allow to:
+The `{ReLTER}` package functions in particular allows to:
 
 -   retrieve the information about entities (e.g. sites, datasets, and
-    activities) shared by [DEIMS-SDR](https://deims.org/) (see
-    e.g. [get_site_info
+    activities) shared by [DEIMS-SDR](https://deims.org/) (see e.g.
+    [get\_site\_info
     function](https://oggioniale.github.io/ReLTER/reference/get_site_info.html));
 
 -   elaborate the information of single site or merge info from national
     network sites or entire International LTER (ILTER) in order to
-    provide maps, figures, graphs etc (see
-    e.g. [get_network_sites](https://oggioniale.github.io/ReLTER/reference/get_network_sites.html),
-    [produce_site_map](https://oggioniale.github.io/ReLTER/reference/produce_site_map.html)
+    provide maps, figures, graphs etc (see e.g.
+    [get\_network\_sites](https://oggioniale.github.io/ReLTER/reference/get_network_sites.html),
+    [produce\_site\_map](https://oggioniale.github.io/ReLTER/reference/produce_site_map.html)
     or
-    [produce_site_parameters_pie](https://oggioniale.github.io/ReLTER/reference/produce_site_parameters_pie.html)
+    [produce\_site\_parameters\_pie](https://oggioniale.github.io/ReLTER/reference/produce_site_parameters_pie.html)
     functions);
 
 -   interact with the [ODSEurope](maps.opendatascience.eu) managed by
     members of the
     [Geo-harmonizer](https://opendatascience.eu/geoharmonizer-project/)
     project starting with the dataset shared by
-    [DEIMS-SDR](https://deims.org/) (see
-    e.g. [get_site_ODS](https://oggioniale.github.io/ReLTER/reference/get_site_ODS.html)
+    [DEIMS-SDR](https://deims.org/) (see e.g.
+    [get\_site\_ODS](https://oggioniale.github.io/ReLTER/reference/get_site_ODS.html)
     function);
 
--   improve the quality of the dataset (see
[…]
