+++ R/get_ilter_envcharacts.R
@@ -5,17 +5,18 @@
 #' \href{https://www.ilter.network/network/global-coverage}{ILTER sites
 #' (more than 1200 around the world)},
 #' through the DEIMS-SDR API.
-#' 
+#'
 #' This function gathers in a unique tibble all the Environmental
-#' Characteristics from all ILTER sites. Note that the execution time for this function
-#' is very high.
-#' 
+#' Characteristics from all ILTER sites. Note that the execution time
+#' for this function is very high.
+#'
 #' If the objective is obtain information about Environmental Characteristics
 #' on a few sites, it is better to use other more specific functions (e.g.
 #' \href{get_network_envcharacts.html}{`get_network_envcharacts()`} or
 #' \href{get_site_info.html}{`get_site_info()`}) or using
 #' other methods
-#' (\href{../../articles/sites_information.html}{How to about sites informations}).
+#' (\href{../../articles/sites_information.html}{How to about sites
+#' informations}).
 #' @param sitesNum A `integer`. The number of the sites that are
 #' read to get the information. Use this parameter only to get a
 #' sample of the output of this function. If the value of sitesNum is
