+++ R/get_sos_procedurelist.R
@@ -1,7 +1,8 @@
 #' List the procedures of a Sensor Observations Service (SOS).
-#' @description Return a list of procedures (e.g. method, algorithm, instrument,
-#' sensor, or system which may be used in making observations)
-#' store into a \href{http://opengeospatial.github.io/e-learning/sos/text/index.html}{SOS
+#' @description Return a list of procedures (e.g. method, algorithm,
+#' instrument, sensor, or system which may be used in making observations)
+#' store into a
+#' \href{http://opengeospatial.github.io/e-learning/sos/text/index.html}{SOS
 #' (Sensor Observations Service OGC)}.
 #' @param sosHost A `list`. An SOS endpoint (e.g.
 #' \url{http://getit.lteritalia.it/observations/sos/kvp?}).
