+++ R/get_network_research_topics.R
@@ -4,7 +4,7 @@
 #' \href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
 #' Italy network}), as a stored into \href{https://deims.org}{DEIMS-SDR}.
 #' @param networkDEIMSID A `character`. The DEIMS ID of a network
-#' from DEIMS-SDR website. DEIMS ID information 
+#' from DEIMS-SDR website. DEIMS ID information
 #' \href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
 #' networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
 #' The DEIMS ID of network is the URL for the network page.
