+++ R/get_ilter_parameters.R
@@ -1,21 +1,23 @@
 #' Obtain information about the parameters collected of all sites ILTER.
 #' @description Return a `tibble` object containing Parameters collected by all
 #' of the \href{https://www.ilter.network/network/global-coverage}{ILTER sites
-#' (more than 1200 around the world)}, available from \href{https://deims.org}{DEIMS-SDR}.
-#' 
+#' (more than 1200 around the world)}, available from
+#' \href{https://deims.org}{DEIMS-SDR}.
+#'
 #' This function gathers in a unique tibble all the Parameters
 #' from all ILTER sites. Note that the execution time for this function
 #' is very high.
-#' 
+#'
 #' If the objective is obtain information about Parameters
 #' on a few sites, it is better to use other more specific functions (e.g.
 #' \href{get_network_parameters.html}{`get_network_parameters()`} or
 #' \href{get_site_info.html}{`get_site_info()`}) or using
 #' other methods
-#' (\href{../../articles/sites_information.html}{How to about sites informations}).
+#' (\href{../../articles/sites_information.html}{How to about sites
+#' informations}).
 #' @param sitesNum A `integer`. The number of the sites that are
 #' read to get the information. Use this parameter only to sample the output
-#' of this function. If the value of sitesNum is #' 0 (default) 
+#' of this function. If the value of sitesNum is #' 0 (default)
 #' all the ILTER sites will be parsed and the waiting time will be long.
 #' @return The output of the function is a `tibble` containing the list
 #' of parameters and their URI (Uniform Resource Identifier) collected

[…]
