+++ .github/workflows/check-standard.yaml
@@ -44,6 +44,7 @@ jobs:
       - name: Query dependencies
         run: |
           install.packages("remotes")
+          install.packages("rnaturalearthdata")
           saveRDS(remotes::dev_package_deps(dependencies = TRUE), "depends.Rds", version = 2)
         shell: Rscript {0}
 
