### 2022-04-01 6e941e1
PhantomJS installation
