+++ .github/workflows/check-standard.yaml
@@ -45,6 +45,7 @@ jobs:
         run: |
           install.packages("remotes")
           install.packages("rnaturalearthdata")
+          install.packages("webshot")
           webshot::install_phantomjs()
           saveRDS(remotes::dev_package_deps(dependencies = TRUE), "depends.Rds", version = 2)
         shell: Rscript {0}
