### 2022-04-05 affdca6
Avoid error if GADM is offline
