+++ vignettes/sites_contacts.Rmd
@@ -5,7 +5,7 @@ vignette: >
   %\VignetteIndexEntry{How to get the list of contacts from list of DEIMS ID within csv file}
   %\VignetteEngine{knitr::knitr}
   %\usepackage[UTF-8]{inputenc}
-author: "Alessandro Oggioni"
+author: "Alessandro Oggioni, Paolo Tagliolato"
 ---
 
 Starting from the [DEIMS-SDR "Advanced Search - Sites" interface](https://deims.org/search/sites), an user can download a csv file of the sites. Some filters can be used for selecting sites based on countries, projects, Biome, etc. The list of sites selected can be downloaded in csv format file.
@@ -18,44 +18,25 @@ This documentation, starting from this file csv, shows how to:
 
 ```{r listContacts, warning=FALSE, message=FALSE}
 library(dplyr)
-site_results <- read.csv("./data/site_contacts_data.csv", sep=";") %>%
-  dplyr::as_tibble()
+site_results <- read.csv("./data/site_contacts_data.csv", sep = ";") %>%
+  dplyr::as_tibble() %>% 
+  dplyr::mutate(url = paste0("https://deims.org/", DEIMS.ID))
 
-allSiteContact <- lapply(
-    as.list(
-      paste0("https://deims.org/", site_results$DEIMS.ID)
-    ),
-    ReLTER::get_site_info,
-    category = "Contacts"
-  )
-contacts <- tibble::tibble(
-  siteName = NA,
-  managerName = NA,
-  managerEmail = NA,
-  managerORCID = NA
-)
-for (i in seq_len(length(allSiteContact))) {
-  if (is.na(allSiteContact[[i]]$generalInfo.siteManager[[1]])) {
-    contacts <- contacts %>% 
-      tibble::add_row(
-        siteName = allSiteContact[[i]]$title,
-        managerName = NA,
-        managerEmail = NA,
-        managerORCID = NA
-      )
-  } else {
-    contacts <- contacts %>% 
-      tibble::add_row(
-        siteName = allSiteContact[[i]]$title,
-        managerName = allSiteContact[[i]]$generalInfo.siteManager[[1]]$name,
-        managerEmail = allSiteContact[[i]]$generalInfo.siteManager[[1]]$email,
-        managerORCID = allSiteContact[[i]]$generalInfo.siteManager[[1]]$orcid
-      )
-  }
-}
+contacts <- site_results$url %>%
+  purrr::map_dfr(.f = function(x) {
+    x %>% ReLTER::get_site_info(category = "Contacts") %>%
+      dplyr::select(title, uri, generalInfo.siteManager, geoCoord) %>%
+      tidyr::unnest()
+  }) %>%
+  dplyr::mutate(
+    site = paste0("[", title, "](", uri, ")"),
+    contact = paste0("[", name, "](mailto:", email, ")")
+  ) %>%
+  unique()
+
 # Contacts table
 knitr::kable(
-  contacts,
+  (contacts %>% dplyr::select(site, contact)),
   caption = "List of the contacts",
   booktabs = TRUE, longtable = TRUE
 )
@@ -68,8 +49,8 @@ knitr::kable(
 ```{r geoSites, warning=FALSE, message=FALSE}
 library(dplyr)
 site_results_geo <- sf::st_as_sf(
-      site_results,
-      wkt = "Coordinates"
+      contacts,
+      wkt = "geoCoord"
     )
 
 listItaSitesMap <- leaflet::leaflet(site_results_geo) %>%
@@ -83,7 +64,13 @@ listItaSitesMap <- leaflet::leaflet(site_results_geo) %>%
     weight = 2,
     opacity = 0.5,
     fill = TRUE,
[…]
