### 2022-04-11 04923d4
authorship of some functions
