+++ man/get_sos_procedurelist.Rd
@@ -31,4 +31,6 @@ get_sos_procedurelist(
 }
 \author{
 Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
+
+Paolo Tagliolato, PhD (2021) \email{tagliolato.p@irea.cnr.it}
 }
