+++ R/get_activity_info.R
@@ -41,9 +41,6 @@ get_activity_info <- function(activityid, show_map = FALSE) {
         created: .created,
         changed: .changed,
         relatedSite: .attributes.general.relatedSite,
-        siteTitle: .attributes.general.relatedSite[].title,
-        DEIMSiD_prefix: .attributes.general.relatedSite[].id.prefix,
-        DEIMSiD_suffix: .attributes.general.relatedSite[].id.suffix,
         contacts: .attributes.contact,
         boundaries: .attributes.geographic.boundaries,
         observationParameters: .attributes.observations.parameters,
