+++ tests/testthat/test-get_activity_info.R
@@ -28,10 +28,10 @@ test_that("Output of activities information function constructs 'sf' and
   )
   expect_s3_class(result, "sf")
   expect_s3_class(result, "tbl_df")
-  expect_true(ncol(result) == 16)
+  expect_true(ncol(result) == 13)
   expect_true(all(names(result) == c(
     "title", "abstract", "keywords", "uri", "type", "created",            
-    "changed", "relatedSite", "siteTitle", "DEIMSiD_prefix", "DEIMSiD_suffix",
+    "changed", "relatedSite", 
     "contacts.corresponding", "contacts.metadataProvider", "boundaries",
     "observationParameters", "relatedResources"
   )))
