+++ R/get_dataset_info.R
@@ -42,9 +42,6 @@ get_dataset_info <- function(datasetid, show_map = FALSE) {
        type: .type,
        dateRange: .attributes.general.dateRange,
        relatedSite: .attributes.general.relatedSite,
-       siteTitle: .attributes.general.relatedSite[].title,
-       DEIMSiD_prefix: .attributes.general.relatedSite[].id.prefix,
-       DEIMSiD_suffix: .attributes.general.relatedSite[].id.suffix,
        contacts: .attributes.contact,
        observationParameters: .attributes.observations.parameters,
        observationSpecies: .attributes.observations.speciesGroups,
