+++ tests/testthat/test-get_activity_info.R
@@ -81,7 +81,7 @@ test_that("The activity ,iss the geo information", {
     show_map = FALSE
   )
   expect_s3_class(result, "tbl_df")
-  expect_true(ncol(result) == 16)
+  expect_true(ncol(result) == 13)
   expect_true(all(names(result) == c(
     "title", "abstract", "keywords", "uri", "type", "created",             
     "changed", "relatedSite", "siteTitle", "DEIMSiD_prefix",
