+++ tests/testthat/test-get_activity_info.R
@@ -42,7 +42,7 @@ test_that("Output of activities information function constructs 'sf' and
 test_that("Wrong input (but URL) constructs a NULL object", {
   Sys.setenv("LOCAL_DEIMS" = FALSE) # set online mode
   expect_error(
-    object = ReLTER::get_activity_info(
+    object <- ReLTER::get_activity_info(
       activityid = "https://deims.org/activity/ljhnhbkihubib",
       show_map = FALSE
     ),
@@ -74,7 +74,7 @@ test_that("Output of get activities information function constructs 'sf' with
   expect_true(any(result_valid))
 })
 
-test_that("The activity ,iss the geo information", {
+test_that("The activity don't have geo information", {
   result <- ReLTER::get_activity_info(
     activityid =
       "https://deims.org/activity/22983172-c53c-4ae9-9623-66f92cb222e3",
@@ -84,8 +84,8 @@ test_that("The activity ,iss the geo information", {
   expect_true(ncol(result) == 13)
   expect_true(all(names(result) == c(
     "title", "abstract", "keywords", "uri", "type", "created",             
-    "changed", "relatedSite", "siteTitle", "DEIMSiD_prefix",
-    "DEIMSiD_suffix", "contacts.corresponding", "contacts.metadataProvider",
+    "changed", "relatedSite", "contacts.corresponding",
+    "contacts.metadataProvider",
     "boundaries", "observationParameters", "relatedResources"
   )))
   expect_type(result$title, "character")
