+++ README.md
@@ -8,7 +8,7 @@ ReLTER
 
 [![](https://img.shields.io/badge/doi-10.5281/zenodo.5576813-yellow.svg)](https://doi.org/10.5281/zenodo.5576813)
 <!-- the version is determined via the DESCRIPTION file -->
-[![](https://img.shields.io/badge/devel%20version-1.0.0-blue.svg)](https://github.com/oggioniale/ReLTER)
+[![](https://img.shields.io/badge/devel%20version-1.1.0-blue.svg)](https://github.com/oggioniale/ReLTER)
 <!-- CRAN badges -->
 <!-- [![](https://www.r-pkg.org/badges/version/oggioniale/ReLTER?color=orange)](https://cran.r-project.org/package=oggioniale/ReLTER) -->
 <!-- [![CRAN checks](https://cranchecks.info/badges/summary/ReLTER)](https://cran.r-project.org/web/checks/check_results_ReLTER.html) -->
@@ -46,16 +46,16 @@ The `{ReLTER}` package functions in particular allows to:
 
 -   retrieve the information about entities (e.g. sites, datasets, and
     activities) shared by [DEIMS-SDR](https://deims.org/) (see e.g.
-    [get\_site\_info
+    [get_site_info
     function](https://oggioniale.github.io/ReLTER/reference/get_site_info.html));
 
 -   elaborate the information of single site or merge info from national
     network sites or entire International LTER (ILTER) in order to
     provide maps, figures, graphs etc (see e.g.
-    [get\_network\_sites](https://oggioniale.github.io/ReLTER/reference/get_network_sites.html),
-    [produce\_site\_map](https://oggioniale.github.io/ReLTER/reference/produce_site_map.html)
+    [get_network_sites](https://oggioniale.github.io/ReLTER/reference/get_network_sites.html),
+    [produce_site_map](https://oggioniale.github.io/ReLTER/reference/produce_site_map.html)
     or
-    [produce\_site\_parameters\_pie](https://oggioniale.github.io/ReLTER/reference/produce_site_parameters_pie.html)
+    [produce_site_parameters_pie](https://oggioniale.github.io/ReLTER/reference/produce_site_parameters_pie.html)
     functions);
 
 -   interact with the [ODSEurope](maps.opendatascience.eu) managed by
@@ -63,11 +63,11 @@ The `{ReLTER}` package functions in particular allows to:
     [Geo-harmonizer](https://opendatascience.eu/geoharmonizer-project/)
     project starting with the dataset shared by
     [DEIMS-SDR](https://deims.org/) (see e.g.
-    [get\_site\_ODS](https://oggioniale.github.io/ReLTER/reference/get_site_ODS.html)
+    [get_site_ODS](https://oggioniale.github.io/ReLTER/reference/get_site_ODS.html)
     function);
 
 -   improve the quality of the dataset (see e.g.
-    [get\_id\_worms](https://oggioniale.github.io/ReLTER/reference/get_id_worms.html)).
+    [get_id_worms](https://oggioniale.github.io/ReLTER/reference/get_id_worms.html)).
 
 Functions currently implemented are derived from the discussion of the
 needs declared by eLTER users community.
@@ -78,22 +78,22 @@ tools.
 
 <!-- about the icons https://github.com/ikatyang/emoji-cheat-sheet -->
 
-## :notebook\_with\_decorative\_cover: Citation
+## :notebook_with_decorative_cover: Citation
 
 To cite `{ReLTER}` please use: Alessandro Oggioni, Micha Silver, Luigi
-Ranghetti & Paolo Tagliolato. (2021). oggioniale/ReLTER: ReLTER v1.0.0
-(1.0.0). Zenodo. <https://doi.org/10.5281/zenodo.5576813>
[…]
