+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: ReLTER
 Type: Package
 Title: An Interface for the 'eLTER' Community
-Version: 1.0.0
+Version: 1.1.0
 Authors@R: c(
         person("Alessandro", "Oggioni", email = "alessandro.oggioni@cnr.it", 
           role = c("aut", "cre", "fnd"),
