### 2022-04-15 ropensci/ReLTER#50 @oggioniale
taxon_id_worms function message of the output
from @allisonhorst
After entering a number to specify the record “most similar to the taxa name” (note: it currently says “mostsimilar”), it isn’t clear that anything happens (but the table does exist as a stored object). Possible to have a message returned after choosing that says “Object created” or something similar, just so the user knows it worked more obviously?
