### 2022-04-15 #485 @oggioniale
Dear @@WillOnGit,
here our comments and answers in line.

> OK, here's my review. I haven't commented on some of the more detailed aspects of the review (such as strict adherence to the various reviewing/packaging guides) but please let me know if I've omitted anything crucial and I'll update it.
>
> I'll keep an eye out for any follow up questions that may arise but will be busy tomorrow afternoon.
>
> ## Package Review
> _Please check off boxes as applicable, and elaborate in comments below. Your review is not limited to these topics, as described in the reviewer guide_
>
> * **Briefly describe any working relationship you have (had) with the package authors.**
>
> I work on the eLTER project separately from the authors. I attended a workshop on the ReLTER package given by Micha Silver. I've also corresponded with Alessandro and Paolo about a separate R package.
>
> * [x]  As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work (if you are unsure whether you are in conflict, please speak to your editor _before_ starting your review).
>
> #### Documentation
> The package includes all the following forms of documentation:
>
> * [x]  **A statement of need:** clearly stating problems the software is designed to solve and its target audience in README
> * [x]  **Installation instructions:** for the development version of package and any non-standard dependencies in README
> * [ ]  **Vignette(s):** demonstrating major functionality that runs successfully locally

The Vignettes are now included.

> * [x]  **Function Documentation:** for all exported functions
> * [x]  **Examples:** (that run successfully locally) for all exported functions
> * [x]  **Community guidelines:** including contribution guidelines in the README or CONTRIBUTING, and DESCRIPTION with `URL`, `BugReports` and `Maintainer` (which may be autogenerated via `Authors@R`).
>
> #### Functionality
> * [ ]  **Installation:** Installation succeeds as documented.

We changed the Installation section in ReadMe file and we created a vignette for the installation process.

> * [x]  **Functionality:** Any functional claims of the software been confirmed.
> * [x]  **Performance:** Any performance claims of the software been confirmed.
> * [x]  **Automated tests:** Unit tests cover essential functions of the package and a reasonable range of inputs and conditions. All tests pass on the local machine.
> * [x]  **Packaging guidelines**: The package conforms to the rOpenSci packaging guidelines.
>
> Estimated hours spent reviewing: 8
>
> * [x]  Should the author(s) deem it appropriate, I agree to be acknowledged as a package reviewer ("rev" role) in the package DESCRIPTION file.

:+1:

It is a pleasure for us to include your name in the package description as a reviewer.

>
> ### Review Comments
> #### Overall comments
> On the whole, I really liked the package - once I could get the damn thing installed!
>
> The package would be good enough if it just standardised the busywork of fetching stuff from APIs such as DEIMS, but it provides lots of extra bonuses (like `produce_site_qrcode`, `produce_network_points_map` and `taxon_id_worms`) that go above and beyond the basics. The good thing is, if you don't like the output of these bonuses you can do your own thing with the data the package gives you.
>
> The documentation and other supporting material is very good, and I'm a stickler for that kind of thing. This helps elevate the package from being functional to enjoyable to use.
>
> I will definitely use this package myself from now on, which is really the ultimate endorsement!

We changed the Installation section in ReadMe file and we created a vignette for the installation process.

>
> ### Technical info
> ```
> > sessionInfo()
> R version 4.0.3 (2020-10-10)
> Platform: x86_64-w64-mingw32/x64 (64-bit)
> Running under: Windows 10 x64 (build 17763)
>
> Matrix products: default
>
> locale:
> [1] LC_COLLATE=English_United Kingdom.1252
> [2] LC_CTYPE=English_United Kingdom.1252
> [3] LC_MONETARY=English_United Kingdom.1252
> [4] LC_NUMERIC=C
> [5] LC_TIME=English_United Kingdom.1252
>
> attached base packages:
> [1] stats     graphics  grDevices utils     datasets  methods
> [7] base
>
> other attached packages:
> [1] ReLTER_1.0.0
>
> loaded via a namespace (and not attached):
> ...
> ```
>
> #### Installation
> I only tested the `devtools::install_github()` installation as that is all that is indicated in the documentation.
>
> Installation was extremely difficult and frustrating (~4 hours); an important caveat is that circumstances meant I only had two environments to use, Windows and Linux, both of which had limited functionality in different ways. In the end, I got the package working on a corporate Windows laptop after battling antivirus software and using several additional `install.packages` invocations. I would strongly recommend further testing on "normal", modern Windows and Linux environments to confirm that my experience was a product of the unusual setup I had to work with.
>
> Because of similar issues, I could not use the `ropensci-org/pkgreviewr` template.
>
> #### Installation comments
> The amount of dependencies is very large, which is understandable considering some of the advanced functionalities provided, especially (I assume) the graphical/GIS features. This meant that installation was slow, which may be offputting for users wanting to try the package out. Perhaps a minimal version/install option of the package would be nice for low resource users or those mainly dealing with DEIMS site metadata.
>
> This was especially noticeable in my case where the installation failed, and on each failure I had to wait a long time before seeing whether it was working again or not.
>
> #### Testing
> No problems here. All tests seem reasonable and all functions are covered. Everything passed locally, results are below.

:+1:

>
> ```
> > devtools::test()
> i Loading ReLTER
>
>
> ReLTER is specially drafted for the LTER community.
>
> ...
>
> == Results ===========================================================================
> Duration: 567.1 s
>
> [ FAIL 0 | WARN 2 | SKIP 0 | PASS 355 ]
> ```
>
> #### Code style
> The code looks well-written on the whole. The only suggestion I have is to introduce more "internal" functions (like `get_site_general` from `get_site_info`) to standardise procedures and cut out some boilerplate code.

We are thinking to work in this direction from next improvements of package.

>
> As an example, in `get_network_sites` there are a few lines at the beginning setting up a URL, defining retries, standardising DEIMS IDs, etc. Similar things appear in `produce_network_points_map`. Some sort of wrapper(s) for contacting the DEIMS API looks like it would be good to replace all these things with something like `getFromDeims("foo")`. As an example (a vain one, at that - run `git blame` on it and you'll see what I mean...), from a similar piece of work see the "normaliseDeimsID" function here: https://github.com/eLTER-RI/deimsPy/blob/main/deims.py.

It can be a good suggestion for next version of package. For sure the internal functions, that we already use, can be the best way for improve the quality and the performance of package.

>
> ### Documentation
> In terms of the [Divio documentation system](https://documentation.divio.com/) (my favourite way of documenting things), the reference material is the strongest, the explanation and tutorials are good (plenty of examples), and the how-to guides are lacking, most likely in the format of a few vignettes to expand on the README examples. Some other points are:
>
> * Even though there is no explicit statement of need, I feel the introduction to the README implicitly addresses most of the required points. I do think it would benefit from more concrete examples in opening paragraph though, or a link to that section.

@Micha changed the Introduction of README.

> * I see a vignette in the source code but none of `browseVignettes()`, `browseVignettes("ReLTER")`, `vignette(package="ReLTER")` or `vignette("ReLTER")` displayed it.

Now a Vignettes have been created.

> * As a native English speaker, there are parts of the documentation (particularly the README) which, while they aren't mistakes, are written unusually. I don't think it matters though because the intended meaning is always clear.

@Micha has review all documentation, including the README.

> * Related to the above, there are some outright typos/errors:
>
>   * `produce_site_map` documentation, examples section: "exclused", should be "excluded"

fixed

>   * `README`, first bullet point: "retrive", should be "retrieve"

done

>   * `README`, first paragraph: "improvements and develop of new tools", should be "improvement and development of new tools".

fixed

> * The function documentation is great. Each page was brief but not missing anything. All the examples I tried worked and they generally gave a clear picture of how to use the function. The only potential improvement I could think of is to hyperlink more to other function pages or vignettes.

We added some pictures to the function documentation and we created some Vignettes in order to explain better the use of the package.

> * The documentation index page is good, in that it links to all the relevant function documentation pages, but doesn't link to any vignettes or general pages about the package. The fact that each page links here satisfies the "principle of multiple points of entry" as far as I'm concerned.

...

> * ~I would suggest to either replicate the function documentation pages on the online help (https://oggioniale.github.io/ReLTER/, probably as a subpage like the authors page https://oggioniale.github.io/ReLTER/authors) or to emphasise that using help(package = "ReLTER") describes each function in detail.~ Oh my god, I just noticed from Allison's review that there's a reference section linked in the top-right. There's no way I would have found that by myself and I would suggest making it more prominent or adding additional links elsewhere in the README/vignettes.

...

>
> ### Functionality
> I've only included functions where I had specific comments here, so anything not mentioned you can assume I:
>
> * understood the point of the function
> * thought it was sufficiently well-documented
> * had no issues trying it out.

:+1:

> Some functions didn't do what I expected from the name (e.g. `produce_site_map`: I assumed you would just pass it a DEIMS ID and it would make a map of the site and its immediate surroundings) but on the whole were well-named and intuitive—once you're familiar with the terminology, although this package will largely be used by folks familiar with eLTER anyway I imagine.

We know that the function names could have been even more clear, but we think that the new documentation (e.g. figures in the references) and the vignettes allow a better explanation of the use of the package.

>
> * `get_activity_info` and `get_dataset_info`
>   I like these. An option to suppress the plot output might be useful though.

Added a new parameter `show_map` in order to suppress the plot of the map.

> * `get_network_sites`
>   The example works great and this function is very useful, but I couldn't get it to work with "LTER Europe"

The eLTER Europe Network has a site with MULTIPOINT WKT geometry. This multi-type geometry can't plot easily in a R LeafLet map. Now, with the improvements of this function, this issue is solved.

> ```
> > get_network_sites("https://deims.org/network/4742ffca-65ac-4aae-815f-83738500a1fc")
> Error in pointData.default(sanitize_sf(obj)) :
>   Don't know how to get location data from object of class sfc_GEOMETRY,sfc
> ```

See the comment above.

>  * `get_site_info`
>   The foundation of the package, works great.

:+1:

> * `get_site_ODS`
>   Simply magnificent.

:+1:

> * `produce_network_points_map`
>   I especially liked this function.

:+1:

> * `produce_site_map`
>   As mentioned, I assumed you would just pass it a DEIMS ID and it would make a map of the site and its immediate surroundings. I don't really understand why this function _requires_ a list of sites separately to the DEIMS ID. I also think the site of the DEIMS ID should be plotted in a different manner to those from the list, to make it more visible on the map.

Now this function returns a `tmap` object. This output can enables to use the map in a different manner by the user.

The parameters of the map consider not only the DEIMS ID of the site but also the Network DEIMS ID because in the viewport are reported the position of all the network sites. In the viewport only the marker of the site just indicated is coloured different.

> * `produce_site_parameters_pie` and `produce_site_parameters_waffle`
>   I like these and get they're shortcuts to save you plotting your own thing but I think a chart title would go a long way to making these more practical for use.

We added the title in the plots, both the pie and the waffle, and we added some text in the description of the function.

> * `produce_site_qrcode`
>   Very cool. Tested the resulting code on my phone and it worked.

:+1:

> * `taxon_id_pesi` and `taxon_id_worms`
>   These work fine and seem handy enough but I also don't really see what it has to do with eLTER. Should they be in a separate package?

These function are been made for some aspect related to the Biodiversity part of eLTER community.

Regards
ReLTER package authors
