### 2022-04-27 #485 @oggioniale
Dear @allisonhorst and @WillOnGit,
thanks a lot for your work.

@maurolepore,
can you provide further instructions about the transfer of the ReLTER repository to the rOpenSci repository?
