### 2022-04-27 #485 @ropensci-review-bot
Approved! Thanks @oggioniale for submitting and @WillOnGit, @allisonhorst for your reviews! :grin:
