### 2022-04-27 #485 @maurolepore
Great news, I now see you're already part of the [team](https://github.com/orgs/ropensci/teams/relter) and I expect that to allow you to transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.
