### 2022-04-29 #485 @oggioniale
Dear @maurolepore I would like to include you as editor of the package but I wanted to know if you agree.
