### 2022-05-02 #485 @maurolepore
Thank you! I think editors don't usually appear in DESCRIPTOIN but you are welcome to add the reviewers should you want to acknowledge their work (with their consent).
