### 2022-05-30 #485 @maurolepore
Dear @oggioniale,

Sorry for insisting. How are you going with the "To-dos" listed https://github.com/ropensci/software-review/issues/485#issuecomment-1111040141?
Are you able to check the boxes or that option is unavailable to you?
