### 2022-05-31 #485 @oggioniale
Dear @maurolepore,
sorry for my long silent.

I'm not able to check the boxes. Let me know how to add this option.
