### 2022-06-07 #485 @oggioniale
Dear @maurolepore,
I write you for a question about the better way for organza the work of rOpenSci/ReLTER repo with multiple (currently we are in 3) developers.
We must to fork the main repo, work each on their own private repo and make Pull Requests. This is the better way? Or we must work directly together in the same (main) repo?
Which is the suggestion by rOpenSci?

Thanks a lot for any suggestions!
