### 2022-06-10 #485 @maurolepore
Dear @oggioniale,

Great to know the package will benefit from multiple contributors, and great to hear you're thinking about how to it it as smoothly as possible.

I'm not aware of a recomendation by rOpenSci. (@maelle does rOpenSci recommend any specific Git remote setup?)

In my experience the "best" workflow varies from team to team. I suggest your team read about [common remote setups](https://happygitwithr.com/common-remote-setups.html) and [useful Git patterns for real life](https://happygitwithr.com/workflows-intro.html), discuss and test the options that seem more promising. After that I would pick the simplest workflow that gets the job done.

It may also help to write a document with the practices you agree to follow ([example](https://2degreesinvesting.github.io/practices/git.html)).

I'm curious to know what your teams finally agrees on. Please let me know.
